amp (MongooseIM v4.0.1)

Models and business logic for XEP-0079: Advanced Message Processing This work was sponsored by Grindr LLC

Types

amp_action()

Specs

t:amp.amp_action/0

amp_any_rule()

Specs

t:amp.amp_any_rule/0

amp_condition()

Specs

t:amp.amp_condition/0

amp_error()

Specs

t:amp.amp_error/0

amp_invalid_rule()

Specs

t:amp.amp_invalid_rule/0

amp_match_result()

Specs

t:amp.amp_match_result/0

amp_rule()

Specs

t:amp.amp_rule/0

amp_rules()

Specs

t:amp.amp_rules/0

amp_value()

Specs

t:amp.amp_value/0

Functions

are_valid_deliver_params/2

are_valid_expire_at_params/2

are_valid_match_resource_params/2

binaries_to_rule/3

Specs

amp.binaries_to_rule/3

error_amp_attrs/3

error_code/1

error_marker_name/1

extract_requested_rules/1

Specs

amp.extract_requested_rules/1

from_bin_/1

Specs

amp.from_bin_/1

WARNING! This is a partial function. Only values that have been verified as legal in binaries_to_rule/3 should be passed here. DO NOT EXPORT!!! conditons

is_amp_el/1

Specs

amp.is_amp_el/1

is_amp_request/1

Specs

amp.is_amp_request/1

We want to keep client->server AMPed messages, but filter out server->client AMPed responses. We can distinguish them by the fact that s2c messages MUST have a status attr on the <amp> element.

is_valid_action/1

is_valid_rule/1

Specs

amp.is_valid_rule/1

make_error_el/2

Specs

amp.make_error_el/2

make_error_response/4

Specs

amp.make_error_response/4

make_response/3

Specs

amp.make_response/3

mk_amp_invalid_rule/3

mk_amp_rule/3

parse_rule/1

Specs

amp.parse_rule/1

parse_rules/1

Specs

amp.parse_rules/1

rule_container_name/1

rule_to_xmlel/1

Specs

amp.rule_to_xmlel/1

strip_amp_el/1

Specs

amp.strip_amp_el/1

to_bin_/1

Specs

amp.to_bin_/1