cyrsasl_external behaviour (MongooseIM v4.0.1)

SASL EXTERNAL implementation (XEP178)

SASL EXTERNAL mechanism requires client's SSL certificate. the purpose of this module is to parse the certificate and get authorization identity (if any provided by the client). this module doesn't make authorization, it only prepares all the data and provides it to auth. backend.
the next extra fields are added to the mongoose_credentials record by this module: * pem_cert - certificate in PEM format * der_cert - certificate in DER format * common_name - CN field (bitstring) of the client's cert (if available) * xmpp_addresses - list of provided "xmppAddr" fields (bare jids) provided in the client's certificate (empty list if not available) * auth_id - authorization identity (bare jid, if provided by the client)

Types

sasl_external_state()

Specs

t:cyrsasl_external.sasl_external_state/0

Functions

auth_id_verification/1

authorize/1

common_name_verification/1

custom_verification/2

do_mech_step/1

get_common_name/1

get_credentials/2

get_verification_list/1

get_xmpp_addresses/1

maybe_add_auth_id/2

maybe_extract_certs/2

mech_new/3

Specs

cyrsasl_external.mech_new/3

mech_step/2

Specs

cyrsasl_external.mech_step/2

Callback implementation for c::cyrsasl.mech_step/2.

mechanism/0

Specs

cyrsasl_external.mechanism/0

Callback implementation for c::cyrsasl.mechanism/0.

standard_verification/1

verification_loop/2

verify_creds/2

verify_server/2

Callbacks

verify_creds(creds)

Specs

c:cyrsasl_external.verify_creds/1