ejabberd_admin (MongooseIM v4.0.1)

Functions

backup_mnesia/1

Specs

ejabberd_admin.backup_mnesia/1

commands/0

Specs

ejabberd_admin.commands/0

delete_expired_messages/0

Specs

ejabberd_admin.delete_expired_messages/0

delete_old_messages/1

Specs

ejabberd_admin.delete_old_messages/1

do_import/2

Specs

ejabberd_admin.do_import/2

do_join_cluster/1

do_leave_cluster/0

do_register/1

Specs

ejabberd_admin.do_register/1

dump_mnesia/1

Specs

ejabberd_admin.dump_mnesia/1

dump_tab/2

Specs

ejabberd_admin.dump_tab/2

dump_table/2

Specs

ejabberd_admin.dump_table/2

dump_tables/2

Specs

ejabberd_admin.dump_tables/2

dump_to_textfile/1

Specs

ejabberd_admin.dump_to_textfile/1

dump_to_textfile/2

Specs

ejabberd_admin.dump_to_textfile/2

dump_to_textfile/3

Specs

ejabberd_admin.dump_to_textfile/3

generate_user/0

get_local_tables/0

Specs

ejabberd_admin.get_local_tables/0

get_loglevel/0

get_results_from_registrator/1

Specs

ejabberd_admin.get_results_from_registrator/1

import_users/1

Specs

ejabberd_admin.import_users/1

install_fallback_mnesia/1

Specs

ejabberd_admin.install_fallback_mnesia/1

join_cluster/1

Specs

ejabberd_admin.join_cluster/1

keep_modules_tables/0

Specs

ejabberd_admin.keep_modules_tables/0

Returns the list of modules tables in use, according to the list of actually loaded modules

keep_tables/0

Specs

ejabberd_admin.keep_tables/0

This function return a list of tables that should be kept from a previous version backup. Obsolete tables or tables created by module who are no longer used are not restored and are ignored.

leave_cluster/0

Specs

ejabberd_admin.leave_cluster/0

load_mnesia/1

Specs

ejabberd_admin.load_mnesia/1

mnesia_change_nodename/4

Specs

ejabberd_admin.mnesia_change_nodename/4

module_tables/1

Specs

ejabberd_admin.module_tables/1

Mapping between modules and their tables

register/2

Specs

ejabberd_admin.register/2

register/3

Specs

ejabberd_admin.register/3

registered_users/1

Specs

ejabberd_admin.registered_users/1

registrator_proc/1

Specs

ejabberd_admin.registrator_proc/1

registrator_proc/2

Specs

ejabberd_admin.registrator_proc/2

remove_dead_node/1

remove_from_cluster/1

Specs

ejabberd_admin.remove_from_cluster/1

remove_rpc_alive_node/1

restore/1

Mnesia database restore This function is called from ejabberd_ctl, ejabberd_web_admin and mod_configure/adhoc

restore_mnesia/1

Specs

ejabberd_admin.restore_mnesia/1

send_job_to_registrator/2

send_service_message_all_mucs/2

Specs

ejabberd_admin.send_service_message_all_mucs/2

set_master/1

Specs

ejabberd_admin.set_master/1

spawn_link_workers/0

Specs

ejabberd_admin.spawn_link_workers/0

start/0

status/0

Specs

ejabberd_admin.status/0

stop/0

unregister/2

Specs

ejabberd_admin.unregister/2