ejabberd_app (MongooseIM v4.0.1)

Functions

broadcast_c2s_shutdown/0

Specs

ejabberd_app.broadcast_c2s_shutdown/0

connect_nodes/0

Specs

ejabberd_app.connect_nodes/0

db_init/0

delete_pid_file/0

Specs

ejabberd_app.delete_pid_file/0

prep_stop/1

Prepare the application for termination. This function is called when an application is about to be stopped, before shutting down the processes of the application.

start/2

Callback implementation for :application.start/2.

start_modules/0

Specs

ejabberd_app.start_modules/0

Start all the modules in all the hosts

start_services/0

Specs

ejabberd_app.start_services/0

stop/1

Callback implementation for :application.stop/1.

stop_modules/0

Specs

ejabberd_app.stop_modules/0

stop_services/0

Specs

ejabberd_app.stop_services/0

update_status_file/1

write_pid_file/0

Specs

ejabberd_app.write_pid_file/0

write_pid_file/2

Specs

ejabberd_app.write_pid_file/2