ejabberd_auth (MongooseIM v4.0.1)

Types

authmodule()

Specs

t:ejabberd_auth.authmodule/0

passterm()

Specs

t:ejabberd_auth.passterm/0

Functions

auth_methods/1

Specs

ejabberd_auth.auth_methods/1

auth_modules/0

Specs

ejabberd_auth.auth_modules/0

Return the lists of all the auth modules actually used in the configuration

auth_modules/1

Specs

ejabberd_auth.auth_modules/1

authorize/1

Specs

ejabberd_auth.authorize/1

authorize_loop/2

Specs

ejabberd_auth.authorize_loop/2

authorize_with_check_password/2

Specs

ejabberd_auth.authorize_with_check_password/2

check_digest/4

Specs

ejabberd_auth.check_digest/4

check_password/2

Specs

ejabberd_auth.check_password/2

Check if the user and password can login in server.

check_password/4

Specs

ejabberd_auth.check_password/4

Check if the user and password can login in server.

check_password_loop/2

Specs

ejabberd_auth.check_password_loop/2

check_password_with_authmodule/2

Specs

ejabberd_auth.check_password_with_authmodule/2

Check if the user and password can login in server. The user can login if at least an authentication method accepts the user and the password.

check_password_with_authmodule/4

Specs

ejabberd_auth.check_password_with_authmodule/4

dirty_get_registered_users/0

Specs

ejabberd_auth.dirty_get_registered_users/0

Registered users list do not include anonymous users logged

do_authorize_loop/2

do_check_password_loop/2

do_get_vh_registered_users/1

do_get_vh_registered_users/2

do_get_vh_registered_users_number/1

do_get_vh_registered_users_number/2

do_try_register_if_does_not_exist/3

Specs

ejabberd_auth.do_try_register_if_does_not_exist/3

do_try_register_if_does_not_exist_timed/3

do_try_register_in_backend/4

does_user_exist/1

Specs

ejabberd_auth.does_user_exist/1

Returns true if the user exists in the DB or if an anonymous user is logged under the given name

does_user_exist_in_given_modules/4

does_user_exist_timed/2

ensure_metrics/1

entropy/1

Specs

ejabberd_auth.entropy/1

Calculate informational entropy.

get_auth_method_as_a_list/1

get_opt/2

get_opt/3

Specs

ejabberd_auth.get_opt/3

get_passterm_with_authmodule/1

Specs

ejabberd_auth.get_passterm_with_authmodule/1

Get the password(like thing) of the user and the auth module.

get_password/1

Specs

ejabberd_auth.get_password/1

Get the password of the user.

get_password_s/1

Specs

ejabberd_auth.get_password_s/1

get_type_information/2

Specs

ejabberd_auth.get_type_information/2

get_vh_registered_users/1

Specs

ejabberd_auth.get_vh_registered_users/1

Registered users list do not include anonymous users logged

get_vh_registered_users/2

Specs

ejabberd_auth.get_vh_registered_users/2

get_vh_registered_users_number/1

Specs

ejabberd_auth.get_vh_registered_users_number/1

get_vh_registered_users_number/2

Specs

ejabberd_auth.get_vh_registered_users_number/2

is_user_exists_in_other_modules/2

Specs

ejabberd_auth.is_user_exists_in_other_modules/2

remove_user/1

Specs

ejabberd_auth.remove_user/1

ejabberd_auth.remove_user/1

Remove user. Note: it may return ok even if there was some problem removing the user.

set_opts/2

Specs

ejabberd_auth.set_opts/2

set_password/2

Specs

ejabberd_auth.set_password/2

start/0

Specs

ejabberd_auth.start/0

start/1

Specs

ejabberd_auth.start/1

stop/1

Specs

ejabberd_auth.stop/1

supports_sasl_module/2

Specs

ejabberd_auth.supports_sasl_module/2

timed_call/4

Specs

ejabberd_auth.timed_call/4

try_register/2

Specs

ejabberd_auth.try_register/2