ejabberd_auth_anonymous (MongooseIM v4.0.1)

Functions

allow_multiple_connections/1

Specs

ejabberd_auth_anonymous.allow_multiple_connections/1

Return true if multiple connections have been allowed in the config file defaults to false

anonymous_protocol/1

Specs

ejabberd_auth_anonymous.anonymous_protocol/1

Return the anonymous protocol to use: sasl_anon|login_anon|both defaults to login_anon

anonymous_user_exist/2

Specs

ejabberd_auth_anonymous.anonymous_user_exist/2

Check if user exist in the anonymous database

authorize/1

Specs

ejabberd_auth_anonymous.authorize/1

Callback implementation for c::ejabberd_gen_auth.authorize/1.

check_password/3

Specs

ejabberd_auth_anonymous.check_password/3

When anonymous login is enabled, check the password for permanent users before allowing access

check_password/5

dirty_get_registered_users/0

Specs

ejabberd_auth_anonymous.dirty_get_registered_users/0

Callback implementation for c::ejabberd_gen_auth.dirty_get_registered_users/0.

does_user_exist/2

Specs

ejabberd_auth_anonymous.does_user_exist/2

Returns true if the user exists in the DB or if an anonymous user is logged under the given name

get_password/2

Specs

ejabberd_auth_anonymous.get_password/2

Return password of permanent user or false for anonymous users

get_password/3

Specs

ejabberd_auth_anonymous.get_password/3

get_password_s/2

gen_auth unimplemented callbacks

get_vh_registered_users/1

Specs

ejabberd_auth_anonymous.get_vh_registered_users/1

Callback implementation for c::ejabberd_gen_auth.get_vh_registered_users/1.

get_vh_registered_users/2

Specs

ejabberd_auth_anonymous.get_vh_registered_users/2

Callback implementation for c::ejabberd_gen_auth.get_vh_registered_users/2.

get_vh_registered_users_number/1

Callback implementation for c::ejabberd_gen_auth.get_vh_registered_users_number/1.

get_vh_registered_users_number/2

Callback implementation for c::ejabberd_gen_auth.get_vh_registered_users_number/2.

is_login_anonymous_enabled/1

Specs

ejabberd_auth_anonymous.is_login_anonymous_enabled/1

Return true if anonymous login is enabled on the server anonymous login can be used with a standard authentication method (i.e. with clients that do not support anonymous login)

is_sasl_anonymous_enabled/1

Specs

ejabberd_auth_anonymous.is_sasl_anonymous_enabled/1

Return true if SASL ANONYMOUS mechanism is enabled one the server

login/2

Specs

ejabberd_auth_anonymous.login/2

purge_hook/3

Specs

ejabberd_auth_anonymous.purge_hook/3

Launch the hook to purge user data only for anonymous users.

register_connection/4

Specs

ejabberd_auth_anonymous.register_connection/4

Register connection

remove_connection/3

Specs

ejabberd_auth_anonymous.remove_connection/3

Remove connection from Mnesia tables

remove_user/2

Specs

ejabberd_auth_anonymous.remove_user/2

Callback implementation for c::ejabberd_gen_auth.remove_user/2.

session_cleanup/5

Specs

ejabberd_auth_anonymous.session_cleanup/5

set_password/3

Specs

ejabberd_auth_anonymous.set_password/3

When anonymous login is enabled, check that the user is permanent before changing its password

start/1

Create the anonymous table if at least one virtual host has anonymous features enabled. Register to login / logout events

stop/1

Callback implementation for c::ejabberd_gen_auth.stop/1.

supports_sasl_module/2

Specs

ejabberd_auth_anonymous.supports_sasl_module/2

Callback implementation for c::ejabberd_gen_auth.supports_sasl_module/2.

try_register/3

Specs

ejabberd_auth_anonymous.try_register/3

When anonymous login is enabled, check if permanent users are allowed on the server:

unregister_connection/5

Specs

ejabberd_auth_anonymous.unregister_connection/5

Remove an anonymous user from the anonymous users table