ejabberd_auth_external (MongooseIM v4.0.1)

Functions

authorize/1

Specs

ejabberd_auth_external.authorize/1

Callback implementation for c::ejabberd_gen_auth.authorize/1.

check_cache_last_options/1

Specs

ejabberd_auth_external.check_cache_last_options/1

check_caches/3

check_password/3

Specs

ejabberd_auth_external.check_password/3

check_password/5

Specs

ejabberd_auth_external.check_password/5

check_password_cache/4

Specs

ejabberd_auth_external.check_password_cache/4

check_password_extauth/3

Specs

ejabberd_auth_external.check_password_extauth/3

check_password_external_cache/3

Check the password using extauth; if success then cache it

check_password_internal/3

Specs

ejabberd_auth_external.check_password_internal/3

dirty_get_registered_users/0

Specs

ejabberd_auth_external.dirty_get_registered_users/0

Callback implementation for c::ejabberd_gen_auth.dirty_get_registered_users/0.

does_user_exist/2

Specs

ejabberd_auth_external.does_user_exist/2

Callback implementation for c::ejabberd_gen_auth.does_user_exist/2.

get_cache_option/1

Specs

ejabberd_auth_external.get_cache_option/1

get_last_access/2

Specs

ejabberd_auth_external.get_last_access/2

Code copied from mod_configure.erl Code copied from web/ejabberd_web_admin.erl

get_last_info/2

Specs

ejabberd_auth_external.get_last_info/2

get_mod_last_configured/1

Specs

ejabberd_auth_external.get_mod_last_configured/1

get_password/2

Specs

ejabberd_auth_external.get_password/2

The password can only be returned if cache is enabled, cached info exists and is fresh enough.

get_password_cache/3

Specs

ejabberd_auth_external.get_password_cache/3

get_password_internal/2

get_password_s/2

Specs

ejabberd_auth_external.get_password_s/2

Callback implementation for c::ejabberd_gen_auth.get_password_s/2.

get_vh_registered_users/1

Specs

ejabberd_auth_external.get_vh_registered_users/1

Callback implementation for c::ejabberd_gen_auth.get_vh_registered_users/1.

get_vh_registered_users/2

Specs

ejabberd_auth_external.get_vh_registered_users/2

Callback implementation for c::ejabberd_gen_auth.get_vh_registered_users/2.

get_vh_registered_users_number/1

Specs

ejabberd_auth_external.get_vh_registered_users_number/1

Callback implementation for c::ejabberd_gen_auth.get_vh_registered_users_number/1.

get_vh_registered_users_number/2

Specs

ejabberd_auth_external.get_vh_registered_users_number/2

Callback implementation for c::ejabberd_gen_auth.get_vh_registered_users_number/2.

is_configured/2

is_fresh_enough/2

Specs

ejabberd_auth_external.is_fresh_enough/2

maybe_set_password_internal/4

remove_user/2

Specs

ejabberd_auth_external.remove_user/2

Callback implementation for c::ejabberd_gen_auth.remove_user/2.

set_password/3

Specs

ejabberd_auth_external.set_password/3

Callback implementation for c::ejabberd_gen_auth.set_password/3.

set_password_internal/3

Specs

ejabberd_auth_external.set_password_internal/3

start/1

Specs

ejabberd_auth_external.start/1

Callback implementation for c::ejabberd_gen_auth.start/1.

stop/1

Callback implementation for c::ejabberd_gen_auth.stop/1.

supports_sasl_module/2

Specs

ejabberd_auth_external.supports_sasl_module/2

Callback implementation for c::ejabberd_gen_auth.supports_sasl_module/2.

try_register/3

Specs

ejabberd_auth_external.try_register/3

Callback implementation for c::ejabberd_gen_auth.try_register/3.

try_register_extauth/3

Specs

ejabberd_auth_external.try_register_extauth/3

try_register_external_cache/3

Specs

ejabberd_auth_external.try_register_external_cache/3

Try to register using extauth; if success then cache it