ejabberd_c2s (MongooseIM v4.0.1)

Types

authenticated_state()

Specs

t:ejabberd_c2s.authenticated_state/0

aux_key()

Specs

t:ejabberd_c2s.aux_key/0

aux_value()

Specs

t:ejabberd_c2s.aux_value/0

blocking_type()

Specs

t:ejabberd_c2s.blocking_type/0

broadcast()

Specs

t:ejabberd_c2s.broadcast/0

broadcast_result()

Specs

t:ejabberd_c2s.broadcast_result/0

broadcast_type()

Specs

t:ejabberd_c2s.broadcast_type/0

conntype()

Specs

t:ejabberd_c2s.conntype/0

debug_presences()

Specs

t:ejabberd_c2s.debug_presences/0

fsm_return()

Specs

t:ejabberd_c2s.fsm_return/0

jid_set()

Specs

t:ejabberd_c2s.jid_set/0

pack_tree()

Specs

t:ejabberd_c2s.pack_tree/0

packet()

Specs

t:ejabberd_c2s.packet/0

routing_result()

Specs

t:ejabberd_c2s.routing_result/0

routing_result_atom()

Specs

t:ejabberd_c2s.routing_result_atom/0

state()

Specs

t:ejabberd_c2s.state/0

statename()

Specs

t:ejabberd_c2s.statename/0

Functions

add_timestamp/3

am_i_available_to/3

am_i_subscribed_to_presence/3

blocking_presence_to_contacts/3

Specs

ejabberd_c2s.blocking_presence_to_contacts/3

blocking_push_to_resources/3

Specs

ejabberd_c2s.blocking_push_to_resources/3

bounce_csi_buffer/1

bounce_messages/1

Specs

ejabberd_c2s.bounce_messages/1

buffer_out_stanza/3

Specs

ejabberd_c2s.buffer_out_stanza/3

c2s_stream_error/2

Specs

ejabberd_c2s.c2s_stream_error/2

calc_to_drop/2

can_use_tls/3

can_use_zlib_compression/2

change_shaper/2

Specs

ejabberd_c2s.change_shaper/2

check_compression_auth/3

check_compression_method/3

check_from/2

Specs

ejabberd_c2s.check_from/2

Check from attributes.

check_incoming_accum_for_conflicts/2

Specs

ejabberd_c2s.check_incoming_accum_for_conflicts/2

check_privacy_and_route/2

Specs

ejabberd_c2s.check_privacy_and_route/2

check_privacy_and_route/3

Specs

ejabberd_c2s.check_privacy_and_route/3

check_privacy_and_route_or_ignore/6

Specs

ejabberd_c2s.check_privacy_and_route_or_ignore/6

check_privacy_and_route_probe/5

Specs

ejabberd_c2s.check_privacy_and_route_probe/5

code_change/4

Callback implementation for c::p1_fsm_old.code_change/4.

compress_setup_failed/0

compress_unsupported_method/0

compressed/0

compression_zlib/0

default_language/0

default_replaced_wait_timeout/0

defer_resource_constraint_check/1

del_aux_field/2

Specs

ejabberd_c2s.del_aux_field/2

determine_features/2

do_handle_ack/5

do_handover_session/2

do_open_session/3

Specs

ejabberd_c2s.do_open_session/3

do_open_session_common/3

do_resume_session/4

Specs

ejabberd_c2s.do_resume_session/4

do_send_element/2

drop_last/2

Specs

ejabberd_c2s.drop_last/2

Drop last N elements from List. It's not an error if N > length(List). The actual number of dropped elements and an empty list is returned.

element_to_origin_accum/2

Specs

ejabberd_c2s.element_to_origin_accum/2

This function is executed when c2s receives a stanza from TCP connection.

enable_stream_resumption/1

filter_mechanism/2

finish_state/3

fix_message_from_user/2

flush_csi_buffer/1

Specs

ejabberd_c2s.flush_csi_buffer/1

flush_messages/1

Specs

ejabberd_c2s.flush_messages/1

flush_messages/2

Specs

ejabberd_c2s.flush_messages/2

flush_or_buffer_packets/1

flush_stream_mgmt_buffer/1

fsm_limit_opts/1

fsm_next_state/2

fsm_next_state: Generate the next_state FSM tuple with different timeout, depending on the future state

fsm_next_state_gc/2

fsm_next_state_gc: Garbage collect the process heap to make use of the newly packed StateData structure.

fsm_next_state_pack/2

fsm_next_state_pack: Pack the StateData structure to improve sharing.

fsm_reply/3

fsm_reply: Generate the reply FSM tuple with different timeout, depending on the future state

generate_random_resource/0

Specs

ejabberd_c2s.generate_random_resource/0

get_ack_freq/0

Specs

ejabberd_c2s.get_ack_freq/0

get_aux_field/2

Specs

ejabberd_c2s.get_aux_field/2

get_buffer_max/0

Specs

ejabberd_c2s.get_buffer_max/0

get_conn_type/1

Specs

ejabberd_c2s.get_conn_type/1

get_info/1

get_msg/1

get_peer_cert/1

Specs

ejabberd_c2s.get_peer_cert/1

get_presence/1

Return Username, Resource and presence information

get_priority_from_presence/1

Specs

ejabberd_c2s.get_priority_from_presence/1

get_replaced_wait_timeout/1

get_resume_timeout/0

Specs

ejabberd_c2s.get_resume_timeout/0

get_showtag/1

get_statustag/1

get_subscribed/1

Return list of all available resources of contacts,

get_subscription/2

Specs

ejabberd_c2s.get_subscription/2

get_xml_lang/1

handle_broadcast_result/4

Specs

ejabberd_c2s.handle_broadcast_result/4

handle_event/3

Callback implementation for c::p1_fsm_old.handle_event/3.

handle_incoming_message/3

handle_info/3

Callback implementation for c::p1_fsm_old.handle_info/3.

handle_routed/5

handle_routed_available_presence/4

Specs

ejabberd_c2s.handle_routed_available_presence/4

handle_routed_broadcast/3

Specs

ejabberd_c2s.handle_routed_broadcast/3

handle_routed_iq/4

Specs

ejabberd_c2s.handle_routed_iq/4

handle_routed_iq/5

Specs

ejabberd_c2s.handle_routed_iq/5

handle_routed_presence/4

Specs

ejabberd_c2s.handle_routed_presence/4

handle_sasl_step/2

handle_sasl_success/2

handle_stream_start/2

handle_sync_event/4

Specs

ejabberd_c2s.handle_sync_event/4

Callback implementation for c::p1_fsm_old.handle_sync_event/4.

handover_session/2

hibernate/0

Specs

ejabberd_c2s.hibernate/0

hook_enabled_features/1

increment_sm_counter/2

increment_sm_incoming/1

init/1

Callback implementation for c::p1_fsm_old.init/1.

invisible_to/3

is_buffer_full/2

is_channel_binding_supported/1

is_fast_tls_configured/2

is_ip_blacklisted/1

Specs

ejabberd_c2s.is_ip_blacklisted/1

Used by c2s blacklist plugins

is_subscribed_to_my_presence/2

is_subscribed_to_my_presence/3

lowcase_and_bare/1

make_available_to/3

make_c2s_info/1

maybe_activate_session/2

maybe_add_cert/2

maybe_add_timestamp/3

maybe_compress_feature/2

maybe_csi_inactive_optimisation/3

Specs

ejabberd_c2s.maybe_csi_inactive_optimisation/3

maybe_do_compress/3

maybe_enable_stream_mgmt/3

maybe_enter_resume_session/1

maybe_enter_resume_session/2

maybe_hibernate/1

Specs

ejabberd_c2s.maybe_hibernate/1

maybe_inactivate_session/2

maybe_increment_sm_incoming/2

maybe_notify_unacknowledged_msg/2

maybe_open_session/2

Specs

ejabberd_c2s.maybe_open_session/2

maybe_resume_session/3

maybe_roster_versioning_feature/1

maybe_sasl_mechanisms/1

maybe_send_ack_request/2

maybe_send_element_from_server_jid_safe/2

Specs

ejabberd_c2s.maybe_send_element_from_server_jid_safe/2

maybe_send_element_from_server_jid_safe/3

Specs

ejabberd_c2s.maybe_send_element_from_server_jid_safe/3

maybe_send_sm_ack/5

maybe_send_trailer_safe/1

Specs

ejabberd_c2s.maybe_send_trailer_safe/1

maybe_text_tag/1

maybe_unexpected_sm_request/3

maybe_update_presence/3

mechanism/1

merge_state/2

new_id/0

Specs

ejabberd_c2s.new_id/0

notify_unacknowledged_messages/1

notify_unacknowledged_msg/2

notify_unacknowledged_msg_if_in_resume_state/2

open_session_allowed_hook/2

pack/1

Specs

ejabberd_c2s.pack/1

Try to reduce the heap footprint of the four presence sets by ensuring that we re-use strings and Jids wherever possible.

pack_jid_set/2

Specs

ejabberd_c2s.pack_jid_set/2

pack_jids/3

Specs

ejabberd_c2s.pack_jids/3

pack_string/2

Specs

ejabberd_c2s.pack_string/2

peerip/2

Specs

ejabberd_c2s.peerip/2

preprocess_and_ship/5

Specs

ejabberd_c2s.preprocess_and_ship/5

presence_broadcast/3

Specs

ejabberd_c2s.presence_broadcast/3

presence_broadcast_first/4

Specs

ejabberd_c2s.presence_broadcast_first/4

presence_broadcast_to_trusted/6

Specs

ejabberd_c2s.presence_broadcast_to_trusted/6

presence_track/2

Specs

ejabberd_c2s.presence_track/2

User sends a directed presence packet

presence_update/3

Specs

ejabberd_c2s.presence_update/3

User updates his presence (non-directed presence packet)

presence_update_to_available/4

Specs

ejabberd_c2s.presence_update_to_available/4

presence_update_to_available/7

Specs

ejabberd_c2s.presence_update_to_available/7

the first one is run when presence changes from unavailable to anything else

print_state/1

Specs

ejabberd_c2s.print_state/1

privacy_check_packet/4

Specs

ejabberd_c2s.privacy_check_packet/4

privacy_check_packet/5

Specs

ejabberd_c2s.privacy_check_packet/5

privacy_check_packet/6

Specs

ejabberd_c2s.privacy_check_packet/6

privacy_list_push_iq/1

process_incoming_stanza/5

process_incoming_stanza_with_conflict_check/5

process_outgoing_stanza/2

Specs

ejabberd_c2s.process_outgoing_stanza/2

Process packets sent by user (coming from user on c2s XMPP connection) eventually it should return {mongoose_acc:t(), fsm_return()} so that the accumulator comes back whence it originated

process_outgoing_stanza/4

process_presence_probe/4

Specs

ejabberd_c2s.process_presence_probe/4

process_presence_subscription_and_route/3

Specs

ejabberd_c2s.process_presence_subscription_and_route/3

process_privacy_iq/3

Specs

ejabberd_c2s.process_privacy_iq/3

process_privacy_iq/4

Specs

ejabberd_c2s.process_privacy_iq/4

process_unauthenticated_stanza/2

Specs

ejabberd_c2s.process_unauthenticated_stanza/2

re_route_packets/1

remove_session_info/3

reroute_unacked_messages/2

Specs

ejabberd_c2s.reroute_unacked_messages/2

resend_csi_buffer/1

resend_offline_message/6

resend_offline_messages/2

Specs

ejabberd_c2s.resend_offline_messages/2

resend_subscription_requests/2

Specs

ejabberd_c2s.resend_subscription_requests/2

response_iq_deny/4

response_negative/5

result_to_amp_event/1

resume_session/2

resume_session/3

roster_change/4

Specs

ejabberd_c2s.roster_change/4

run_remote_hook/3

run_remote_hook_after/4

sasl_challenge_stanza/1

Specs

ejabberd_c2s.sasl_challenge_stanza/1

sasl_failure_stanza/1

Specs

ejabberd_c2s.sasl_failure_stanza/1

sasl_success_stanza/1

Specs

ejabberd_c2s.sasl_success_stanza/1

send_and_maybe_buffer_stanza/3

Specs

ejabberd_c2s.send_and_maybe_buffer_stanza/3

send_and_maybe_buffer_stanza_no_ack/3

Specs

ejabberd_c2s.send_and_maybe_buffer_stanza_no_ack/3

send_back_error/4

send_element/3

Specs

ejabberd_c2s.send_element/3

This is the termination point - from here stanza is sent to the user

send_element_from_server_jid/2

Specs

ejabberd_c2s.send_element_from_server_jid/2

send_element_from_server_jid/3

Specs

ejabberd_c2s.send_element_from_server_jid/3

send_filtered/5

send_header/4

Specs

ejabberd_c2s.send_header/4

send_text/2

Specs

ejabberd_c2s.send_text/2

send_trailer/1

send_unavail_if_newly_blocked/4

send_unavail_if_newly_blocked/6

session_established/2

Specs

ejabberd_c2s.session_established/2

session_established/3

set_aux_field/3

Specs

ejabberd_c2s.set_aux_field/3

ship_to_local_user/3

Specs

ejabberd_c2s.ship_to_local_user/3

should_close_session/1

should_retransmit_last_presence/3

sm_handled_count_too_high_stanza/2

Specs

ejabberd_c2s.sm_handled_count_too_high_stanza/2

socket_type/0

specifically_visible_to/2

Is generally invisible, but visible to a particular resource?

start/2

Specs

ejabberd_c2s.start/2

start_link/2

starttls_stanza/1

stop/1

Stops the session gracefully, entering resume state if applicable

store_session_info/3

stream_features/1

stream_mgmt_ack/1

stream_mgmt_enabled/0

stream_mgmt_enabled/1

stream_mgmt_failed/1

stream_mgmt_failed/2

stream_mgmt_handle_ack/3

stream_mgmt_parse_h/1

stream_mgmt_request/0

stream_mgmt_resumed/2

stream_start_by_protocol_version/2

stream_start_error/2

stream_start_features_after_auth/1

stream_start_features_before_auth/1

stream_start_negotiate_features/1

strip_c2s_fields/1

Specs

ejabberd_c2s.strip_c2s_fields/1

terminate/4

Specs

ejabberd_c2s.terminate/4

Callback implementation for c::p1_fsm_old.terminate/4.

terminate_session/2

terminates the session immediately and unconditionally, sending the user a stream conflict error specifying the reason

terminate_when_tls_required_but_not_enabled/4

timestamp_xml/2

tls_proceed/0

update_priority/4

Specs

ejabberd_c2s.update_priority/4

update_stanza/4

Specs

ejabberd_c2s.update_stanza/4

user_allowed/2

verify_opts/1

wait_for_feature_after_auth/2

Specs

ejabberd_c2s.wait_for_feature_after_auth/2

wait_for_feature_before_auth/2

Specs

ejabberd_c2s.wait_for_feature_before_auth/2

wait_for_sasl_response/2

Specs

ejabberd_c2s.wait_for_sasl_response/2

wait_for_session_or_sm/2

Specs

ejabberd_c2s.wait_for_session_or_sm/2

wait_for_stream/2

Specs

ejabberd_c2s.wait_for_stream/2