ejabberd_config (MongooseIM v4.0.1)

Types

compare_result()

Specs

t:ejabberd_config.compare_result/0

host()

Specs

t:ejabberd_config.host/0

key()

Specs

t:ejabberd_config.key/0

state()

Specs

t:ejabberd_config.state/0

value()

Specs

t:ejabberd_config.value/0

Functions

add_global_option/2

Specs

ejabberd_config.add_global_option/2

add_local_option/2

Specs

ejabberd_config.add_local_option/2

add_virtual_host/1

Specs

ejabberd_config.add_virtual_host/1

all_cluster_nodes/0

apply_reload_changes/4

apply_reloading_change/1

assert_config_reloaded/1

assert_local_config_reloaded/0

assert_required_files_exist/1

clean_acls_opts/0

clean_global_opts/0

clean_local_opts/0

compute_config_file_version/0

compute_loaded_config_version/0

config_info/0

config_state/0

Returns configs on disc and in memory for this node. This function prepares all state data to pass into pure code part (i.e. mongoose_config_parser and mongoose_config_reload).

config_states/0

config_states/1

Returns config states from all nodes in cluster State from the local node comes as head of a list

del_local_option/1

Specs

ejabberd_config.del_local_option/1

do_reload_cluster/1

do_set_opts/1

Specs

ejabberd_config.do_set_opts/1

dump_reload_state/2

dump_reload_state_filename/0

get_categorized_options/0

Returns current all options in memory, grouped by category.

get_config_path/0

Specs

ejabberd_config.get_config_path/0

Get the filename of the ejabberd configuration file. The filename can be specified with: erl -config "/path/to/mongooseim.toml". It can also be specified with the environment variable EJABBERD_CONFIG_PATH. If not specified, the default value 'mongooseim.toml' is assumed.

get_global_config/0

Specs

ejabberd_config.get_global_config/0

get_global_option/1

Specs

ejabberd_config.get_global_option/1

get_host_local_config/0

Specs

ejabberd_config.get_host_local_config/0

get_local_config/0

Specs

ejabberd_config.get_local_config/0

get_local_option/1

Specs

ejabberd_config.get_local_option/1

get_local_option/2

Specs

ejabberd_config.get_local_option/2

get_local_option_or_default/2

Specs

ejabberd_config.get_local_option_or_default/2

get_vh_by_auth_method/1

Return the list of hosts handled by a given module

handle_config_add/1

handle_config_change/1

handle_config_del/1

handle_local_config_add/1

handle_local_config_change/1

handle_local_config_del/1

handle_local_hosts_config_add/1

handle_local_hosts_config_change/1

handle_local_hosts_config_del/1

handle_table_does_not_exist_error/1

is_mongooseim_node/1

Specs

ejabberd_config.is_mongooseim_node/1

load_file/1

Specs

ejabberd_config.load_file/1

Load the ejabberd configuration file. It also includes additional configuration files and replaces macros. This function will crash if finds some error in the configuration file.

maybe_clean_acls_opts/1

maybe_clean_global_opts/1

maybe_clean_local_opts/1

methods_to_auth_modules/1

missing_files/1

other_cluster_nodes/0

Specs

ejabberd_config.other_cluster_nodes/0

print_flat_config/0

reload_cluster/0

Specs

ejabberd_config.reload_cluster/0

reload_cluster_dryrun/0

Specs

ejabberd_config.reload_cluster_dryrun/0

reload_config/1

Specs

ejabberd_config.reload_config/1

reload_listeners/1

Specs

ejabberd_config.reload_listeners/1

reload_local/0

Specs

ejabberd_config.reload_local/0

reload_local_config/1

Specs

ejabberd_config.reload_local_config/1

reload_local_hosts_config/1

Specs

ejabberd_config.reload_local_hosts_config/1

reload_nodes/3

remove_virtual_host/1

Specs

ejabberd_config.remove_virtual_host/1

set_opts/1

Specs

ejabberd_config.set_opts/1

start/0

Specs

ejabberd_config.start/0

state_to_missing_and_required_files/1

try_reload_cluster/2