ejabberd_ctl (MongooseIM v4.0.1)

Management of mongooseimctl commands and frontend to ejabberd commands.

An mongooseimctl command is an abstract function identified by a name, with a defined number of calling arguments, that can be defined in any Erlang module and executed using mongooseimctl administration script.

Note: strings cannot have blankspaces
Does not support commands that have arguments with ctypes: list, tuple

Types

cmd()

Specs

t:ejabberd_ctl.cmd/0

format()

Specs

t:ejabberd_ctl.format/0

format_type()

Specs

t:ejabberd_ctl.format_type/0

Functions

call_command/3

Specs

ejabberd_ctl.call_command/3

filter_commands/2

filter_commands_regexp/2

format_command_lines/5

Specs

ejabberd_ctl.format_command_lines/5

format_result/2

Specs

ejabberd_ctl.format_result/2

format_status/1

format_usage_ctype/2

format_usage_tuple/2

get_accesscommands/0

Specs

ejabberd_ctl.get_accesscommands/0

get_dist_proto/0

get_list_commands/0

Specs

ejabberd_ctl.get_list_commands/0

get_list_ctls/0

Specs

ejabberd_ctl.get_list_ctls/0

get_mongoose_status/0

get_shell_info/0

Specs

ejabberd_ctl.get_shell_info/0

Get some info about the shell: how many columns of width and guess if it supports text formatting codes.

get_uptime/0

init/0

Specs

ejabberd_ctl.init/0

join/2

Specs

ejabberd_ctl.join/2

Join words in a segment, but stop adding to a segment if adding this word would pass L

join/5

Specs

ejabberd_ctl.join/5

make_status/1

Specs

ejabberd_ctl.make_status/1

mix_desc_segments/2

Specs

ejabberd_ctl.mix_desc_segments/2

prepare_description/3

Specs

ejabberd_ctl.prepare_description/3

Split this command description in several lines of proper length

prepare_long_line/3

Specs

ejabberd_ctl.prepare_long_line/3

print_usage/0

print_usage/3

Specs

ejabberd_ctl.print_usage/3

print_usage_command/3

print_usage_command/4

print_usage_commands/3

print_usage_commands/4

Specs

ejabberd_ctl.print_usage_commands/4

print_usage_commands2/3

print_usage_help/2

print_usage_tags/2

print_usage_tags/3

process/1

Specs

ejabberd_ctl.process/1

The commands status, stop and restart are defined here to ensure they are usable even if ejabberd is completely stopped.

process2/2

Specs

ejabberd_ctl.process2/2

register_commands/3

Specs

ejabberd_ctl.register_commands/3

split_desc_segments/2

start/0

Specs

ejabberd_ctl.start/0

try_call_command/3

Specs

ejabberd_ctl.try_call_command/3

try_run_ctp/3

Specs

ejabberd_ctl.try_run_ctp/3

tuple_command_help/1

Specs

ejabberd_ctl.tuple_command_help/1

unregister_commands/3

Specs

ejabberd_ctl.unregister_commands/3