ejabberd_hooks (MongooseIM v4.0.1)

Types

hook()

Specs

t:ejabberd_hooks.hook/0

Functions

add/1

Specs

ejabberd_hooks.add/1

add/4

Specs

ejabberd_hooks.add/4

Add a fun to the given hook. The integer sequence is used to sort the calls: low numbers are executed before high numbers.

add/5

Specs

ejabberd_hooks.add/5

Add a module and function to the given hook. The integer sequence is used to sort the calls: low numbers are executed before high numbers.

add_or_del/2

Specs

ejabberd_hooks.add_or_del/2

code_change/3

Callback implementation for :gen_server.code_change/3.

delete/1

Specs

ejabberd_hooks.delete/1

delete/4

Specs

ejabberd_hooks.delete/4

Delete a module and function from this hook. It is important to indicate exactly the same information than when the call was added.

delete/5

Specs

ejabberd_hooks.delete/5

error_running_hook/3

handle_call/3

Callback implementation for :gen_server.handle_call/3.

handle_cast/2

Callback implementation for :gen_server.handle_cast/2.

handle_info/2

Callback implementation for :gen_server.handle_info/2.

hook_apply_function/4

init/1

Callback implementation for :gen_server.init/1.

run/2

Specs

ejabberd_hooks.run/2

Run the calls of this hook in order, don't care about function results. If a call returns stop, no more calls are performed.

run/3

Specs

ejabberd_hooks.run/3

run_fold/3

Specs

ejabberd_hooks.run_fold/3

Run the calls of this hook in order. The arguments passed to the function are: [Val | Args]. The result of a call is used as Val for the next call. If a call returns 'stop', no more calls are performed and 'stopped' is returned. If a call returns {stop, NewVal}, no more calls are performed and NewVal is returned.

run_fold/4

Specs

ejabberd_hooks.run_fold/4

run_fold1/4

start_link/0

terminate/2

Callback implementation for :gen_server.terminate/2.