ejabberd_listener (MongooseIM v4.0.1)

Types

addr()

Specs

t:ejabberd_listener.addr/0

listener_option()

Specs

t:ejabberd_listener.listener_option/0

port_ip_proto()

Specs

t:ejabberd_listener.port_ip_proto/0

portnum()

Specs

t:ejabberd_listener.portnum/0

proto()

Specs

t:ejabberd_listener.proto/0

Functions

add_listener/3

Specs

ejabberd_listener.add_listener/3

Add a listener and store in config if success

add_proto/2

Specs

ejabberd_listener.add_proto/2

certfile_readable/1

Specs

ejabberd_listener.certfile_readable/1

check_listener_options/1

Specs

ejabberd_listener.check_listener_options/1

delete_listener/2

Specs

ejabberd_listener.delete_listener/2

delete_listener/3

Specs

ejabberd_listener.delete_listener/3

format_error/1

Specs

ejabberd_listener.format_error/1

get_ip_tuple/2

Specs

ejabberd_listener.get_ip_tuple/2

get_proto/1

Specs

ejabberd_listener.get_proto/1

includes_deprecated_ssl_option/1

Specs

ejabberd_listener.includes_deprecated_ssl_option/1

Parse the options of the socket, and return if the deprecated option 'ssl' is included

init/1

normalize_proto/1

Specs

ejabberd_listener.normalize_proto/1

opts_to_listener_args/2

Specs

ejabberd_listener.opts_to_listener_args/2

parse_listener_portip/2

Specs

ejabberd_listener.parse_listener_portip/2

Parse any kind of ejabberd listener specification. The parsed options are returned in several formats. OptsClean does not include inet/inet6 or ip options. Opts can include the options inet6 and {ip, Tuple}, but they are only used when no IP address was specified in the PortIP. The IP version (either IPv4 or IPv6) is inferred from the IP address type, so the option inet/inet6 is only used when no IP is specified at all.

prepare_opts/3

Specs

ejabberd_listener.prepare_opts/3

report_duplicated_portips/1

Specs

ejabberd_listener.report_duplicated_portips/1

socket_error/6

start_link/0

Specs

ejabberd_listener.start_link/0

start_listener/3

Specs

ejabberd_listener.start_listener/3

start_listener_sup/3

Specs

ejabberd_listener.start_listener_sup/3

start_listeners/0

Specs

ejabberd_listener.start_listeners/0

start_module_sup/2

Specs

ejabberd_listener.start_module_sup/2

stop_listener/2

Specs

ejabberd_listener.stop_listener/2

stop_listeners/0

Specs

ejabberd_listener.stop_listeners/0

strip_ip_option/1

Specs

ejabberd_listener.strip_ip_option/1