ejabberd_receiver (MongooseIM v4.0.1)

Types

state()

Specs

t:ejabberd_receiver.state/0

Functions

activate_socket/1

Specs

ejabberd_receiver.activate_socket/1

become_controller/2

change_shaper/2

Specs

ejabberd_receiver.change_shaper/2

close/1

Specs

ejabberd_receiver.close/1

code_change/3

Callback implementation for :gen_server.code_change/3.

compress/2

deactivate_socket/1

Specs

ejabberd_receiver.deactivate_socket/1

element_wrapper/1

Specs

ejabberd_receiver.element_wrapper/1

Element coming from XML parser are wrapped inside xmlstreamelement When we receive directly xmlel tuple (from a socket module speaking directly Erlang XML), we wrap it inside the same xmlstreamelement coming from the XML parser.

free_parser/1

gen_fsm/0

gen_server_call_or_noproc/2

get_socket/1

handle_call/3

Callback implementation for :gen_server.handle_call/3.

handle_cast/2

Callback implementation for :gen_server.handle_cast/2.

handle_info/2

Callback implementation for :gen_server.handle_info/2.

hibernate/0

Specs

ejabberd_receiver.hibernate/0

init/1

Specs

ejabberd_receiver.init/1

Callback implementation for :gen_server.init/1.

maybe_hibernate/1

Specs

ejabberd_receiver.maybe_hibernate/1

maybe_pause/2

maybe_run_keep_alive_hook/2

process_data/2

Specs

ejabberd_receiver.process_data/2

Data processing for connectors directly generating xmlel in Erlang data structure. WARNING: Shaper does not work with Erlang data structure.

reset_parser/1

resolve_peername/2

shutdown_socket_and_wait_for_peer_to_close/2

start/4

start_link/4

Specs

ejabberd_receiver.start_link/4

starttls/2

terminate/2

Callback implementation for :gen_server.terminate/2.

update_stanza_size/3

wrap_if_xmlel/1