ejabberd_router (MongooseIM v4.0.1)

Types

domain()

Specs

t:ejabberd_router.domain/0

external_component()

Specs

t:ejabberd_router.external_component/0

return_hidden()

Specs

t:ejabberd_router.return_hidden/0

Functions

all_routes/1

Specs

ejabberd_router.all_routes/1

check_component/2

Specs

ejabberd_router.check_component/2

Check if the component/route is already registered somewhere; ok means it is not, so we are ok to proceed, anything else means the domain/node pair is already serviced. true and false are there because that's how orelse works.

check_component_global/2

check_component_local/2

check_component_route/1

code_change/3

Callback implementation for :gen_server.code_change/3.

compile_routing_module/0

default_routing_modules/0

dirty_get_all_domains/0

Specs

ejabberd_router.dirty_get_all_domains/0

dirty_get_all_domains/1

Specs

ejabberd_router.dirty_get_all_domains/1

dirty_get_all_routes/0

Specs

ejabberd_router.dirty_get_all_routes/0

dirty_get_all_routes/1

Specs

ejabberd_router.dirty_get_all_routes/1

do_register_component/4

do_unregister_component/2

get_global_component/2

handle_call/3

Callback implementation for :gen_server.handle_call/3.

handle_cast/2

Callback implementation for :gen_server.handle_cast/2.

handle_info/2

Callback implementation for :gen_server.handle_info/2.

init/1

Callback implementation for :gen_server.init/1.

lookup_component/1

Specs

ejabberd_router.lookup_component/1

Returns a list of components registered for this domain by any node, the choice is yours.

lookup_component/2

Specs

ejabberd_router.lookup_component/2

Returns a list of components registered for this domain at the given node. (must be only one, or nothing)

make_routing_module_source/1

register_component/2

Specs

ejabberd_router.register_component/2

components are registered in two places: external_components table as local components and external_components_global as globals. Registration should be done by register_component/1 or register_components/1, which registers them for current node; the arity 2 funcs are here for testing.

register_component/3

Specs

ejabberd_router.register_component/3

register_component/4

Specs

ejabberd_router.register_component/4

register_components/2

Specs

ejabberd_router.register_components/2

register_components/3

Specs

ejabberd_router.register_components/3

register_components/4

Specs

ejabberd_router.register_components/4

register_route/2

Specs

ejabberd_router.register_route/2

register_route_to_ldomain/3

Specs

ejabberd_router.register_route_to_ldomain/3

register_routes/2

Specs

ejabberd_router.register_routes/2

route/3

Specs

ejabberd_router.route/3

The main routing function. It puts the message through a chain of filtering/routing modules, as defined in config 'routing_modules' setting (default is hardcoded in make_routing_module_source function of this module). Each of those modules should use xmpp_router behaviour and implement two functions: filter/3 - should return either 'drop' atom or its args route/3, which should either: - deliver the message locally by calling mongoose_local_delivery:do_local_route/3 and return 'done' - deliver the message it its own way and return 'done' - return its args - return a tuple of {From, To, Packet} which might be modified For both functions, returning a 'drop' or 'done' atom terminates the procedure, while returning a tuple means 'proceed' and the tuple is passed to the next module in sequence.

route/4

Specs

ejabberd_router.route/4

route/5

Specs

ejabberd_router.route/5

route_error/4

Specs

ejabberd_router.route_error/4

route_error_reply/4

Specs

ejabberd_router.route_error_reply/4

routes_cleanup_on_nodedown/2

Specs

ejabberd_router.routes_cleanup_on_nodedown/2

routing_modules_list/0

start_link/0

Specs

ejabberd_router.start_link/0

terminate/2

Callback implementation for :gen_server.terminate/2.

unregister_component/1

Specs

ejabberd_router.unregister_component/1

unregister_component/2

Specs

ejabberd_router.unregister_component/2

unregister_components/1

Specs

ejabberd_router.unregister_components/1

unregister_components/2

Specs

ejabberd_router.unregister_components/2

unregister_route/1

unregister_routes/1

update_tables/0