ejabberd_s2s_out (MongooseIM v4.0.1)

Types

element_queue()

Specs

t:ejabberd_s2s_out.element_queue/0

fsm_return()

Specs

t:ejabberd_s2s_out.fsm_return/0

-define(DBGFSM, true).

state()

Specs

t:ejabberd_s2s_out.state/0

statename()

Specs

t:ejabberd_s2s_out.statename/0

FSM handler return value

Functions

bounce_element/3

Specs

ejabberd_s2s_out.bounce_element/3

Bounce a single message (xmlel)

bounce_messages/1

Specs

ejabberd_s2s_out.bounce_messages/1

bounce_queue/2

Specs

ejabberd_s2s_out.bounce_queue/2

calc_addr_index/1

cancel_timer/1

Specs

ejabberd_s2s_out.cancel_timer/1

code_change/4

do_get_predefined_addresses/1

Specs

ejabberd_s2s_out.do_get_predefined_addresses/1

fsm_limit_opts/0

Specs

ejabberd_s2s_out.fsm_limit_opts/0

get_acc_with_new_sext/3

get_acc_with_new_tls/3

get_addr_list/1

Specs

ejabberd_s2s_out.get_addr_list/1

get_addr_port/1

Specs

ejabberd_s2s_out.get_addr_port/1

get_addrs/2

Specs

ejabberd_s2s_out.get_addrs/2

get_max_retry_delay/0

Get the maximum allowed delay for retry to reconnect (in milliseconds). The default value is 5 minutes. The option {s2s_max_retry_delay, Seconds} can be used (in seconds).

get_predefined_addresses/1

Specs

ejabberd_s2s_out.get_predefined_addresses/1

Get IPs predefined for a given s2s domain in the configuration

get_timeout_interval/1

Specs

ejabberd_s2s_out.get_timeout_interval/1

Calculate timeout depending on which state we are in: Can return integer > 0 | infinity

get_tls_opts_with_certfile/1

get_tls_opts_with_ciphers/1

handle_event/3

handle_info/3

handle_parsed_features/1

handle_sync_event/4

init/1

Specs

ejabberd_s2s_out.init/1

is_verify_res/1

Specs

ejabberd_s2s_out.is_verify_res/1

log_s2s_out/4

Human readable S2S logging: Log only new outgoing connections as INFO Do not log dialback

new_id/0

Specs

ejabberd_s2s_out.new_id/0

open_socket/2

Specs

ejabberd_s2s_out.open_socket/2

open_socket1/2

Specs

ejabberd_s2s_out.open_socket1/2

open_socket2/3

Specs

ejabberd_s2s_out.open_socket2/3

outgoing_s2s_families/0

Specs

ejabberd_s2s_out.outgoing_s2s_families/0

outgoing_s2s_port/0

Specs

ejabberd_s2s_out.outgoing_s2s_port/0

outgoing_s2s_timeout/0

Specs

ejabberd_s2s_out.outgoing_s2s_timeout/0

print_state/1

relay_to_bridge/2

Specs

ejabberd_s2s_out.relay_to_bridge/2

remove_addr_index/1

reopen_socket/2

Specs

ejabberd_s2s_out.reopen_socket/2

send_db_request/1

Specs

ejabberd_s2s_out.send_db_request/1

send_element/2

Specs

ejabberd_s2s_out.send_element/2

send_element/3

Specs

ejabberd_s2s_out.send_element/3

send_event/3

send_queue/2

Specs

ejabberd_s2s_out.send_queue/2

send_text/2

Specs

ejabberd_s2s_out.send_text/2

srv_lookup/1

Specs

ejabberd_s2s_out.srv_lookup/1

srv_lookup/3

Specs

ejabberd_s2s_out.srv_lookup/3

XXX - this behaviour is suboptimal in the case that the domain has a "_xmpp-server._tcp." but not a "_jabber._tcp." record and we don't get a DNS reply for the "_xmpp-server._tcp." lookup. In this case we'll give up when we get the "_jabber._tcp." nxdomain reply.

start/3

Specs

ejabberd_s2s_out.start/3

start_connection/1

start_link/3

Specs

ejabberd_s2s_out.start_link/3

stop_connection/1

stream_established/2

Specs

ejabberd_s2s_out.stream_established/2

terminate/3

terminate_if_waiting_delay/2

Terminate s2s_out connections that are in state wait_before_retry

test_get_addr_port/1

wait_before_reconnect/1

Specs

ejabberd_s2s_out.wait_before_reconnect/1

This function is intended to be called at the end of a state function that want to wait for a reconnect delay before stopping.

wait_before_retry/2

Specs

ejabberd_s2s_out.wait_before_retry/2

This state is use to avoid reconnecting to often to bad sockets

wait_for_auth_result/2

Specs

ejabberd_s2s_out.wait_for_auth_result/2

wait_for_features/2

Specs

ejabberd_s2s_out.wait_for_features/2

wait_for_starttls_proceed/2

Specs

ejabberd_s2s_out.wait_for_starttls_proceed/2

wait_for_stream/2

Specs

ejabberd_s2s_out.wait_for_stream/2

wait_for_validation/2

Specs

ejabberd_s2s_out.wait_for_validation/2