ejabberd_sm (MongooseIM v4.0.1)

Types

backend()

Specs

t:ejabberd_sm.backend/0

close_reason()

Specs

t:ejabberd_sm.close_reason/0

info()

Specs

t:ejabberd_sm.info/0

info_item()

Specs

t:ejabberd_sm.info_item/0

info_key()

Specs

t:ejabberd_sm.info_key/0

priority()

Specs

t:ejabberd_sm.priority/0

ses_tuple()

Specs

t:ejabberd_sm.ses_tuple/0

session()

Specs

t:ejabberd_sm.session/0

sid()

Specs

t:ejabberd_sm.sid/0

state()

Specs

t:ejabberd_sm.state/0

Functions

bounce_offline_message/4

Specs

ejabberd_sm.bounce_offline_message/4

check_existing_resources/3

Specs

ejabberd_sm.check_existing_resources/3

check_for_sessions_to_replace/1

Specs

ejabberd_sm.check_for_sessions_to_replace/1

On new session, check if some existing connections need to be replace

check_in_subscription/6

Specs

ejabberd_sm.check_in_subscription/6

check_max_sessions/3

Specs

ejabberd_sm.check_max_sessions/3

clean_session_list/1

Specs

ejabberd_sm.clean_session_list/1

clean_session_list/2

Specs

ejabberd_sm.clean_session_list/2

close_session/4

Specs

ejabberd_sm.close_session/4

close_session/6

close_session_unset_presence/5

Specs

ejabberd_sm.close_session_unset_presence/5

close_session_unset_presence/7

code_change/3

Callback implementation for :gen_server.code_change/3.

commands/0

Specs

ejabberd_sm.commands/0

disconnect_removed_user/3

Specs

ejabberd_sm.disconnect_removed_user/3

do_filter/3

do_route/4

Specs

ejabberd_sm.do_route/4

do_route_no_resource/6

Specs

ejabberd_sm.do_route_no_resource/6

do_route_no_resource_presence/5

Specs

ejabberd_sm.do_route_no_resource_presence/5

do_route_no_resource_presence_prv/6

Specs

ejabberd_sm.do_route_no_resource_presence_prv/6

do_route_offline/6

Specs

ejabberd_sm.do_route_offline/6

force_update_presence/1

Specs

ejabberd_sm.force_update_presence/1

get_cached_unique_count/0

Specs

ejabberd_sm.get_cached_unique_count/0

get_full_session_list/0

Specs

ejabberd_sm.get_full_session_list/0

get_info/2

Specs

ejabberd_sm.get_info/2

get_max_user_sessions/2

Specs

ejabberd_sm.get_max_user_sessions/2

Get the user_max_session setting This option defines the max number of time a given users are allowed to log in. Defaults to infinity

get_node_sessions_number/0

Specs

ejabberd_sm.get_node_sessions_number/0

get_raw_sessions/1

Specs

ejabberd_sm.get_raw_sessions/1

get_raw_sessions/2

get_session/1

Specs

ejabberd_sm.get_session/1

get_session/3

get_session_ip/1

Specs

ejabberd_sm.get_session_ip/1

get_session_ip/3

get_session_pid/1

Specs

ejabberd_sm.get_session_pid/1

get_session_pid/3

get_total_sessions_number/0

Specs

ejabberd_sm.get_total_sessions_number/0

get_unique_sessions_number/0

Specs

ejabberd_sm.get_unique_sessions_number/0

get_user_present_pids/2

Specs

ejabberd_sm.get_user_present_pids/2

get_user_present_resources/1

Specs

ejabberd_sm.get_user_present_resources/1

get_user_present_resources/2

get_user_resources/1

Specs

ejabberd_sm.get_user_resources/1

get_user_resources/2

get_vh_session_list/1

Specs

ejabberd_sm.get_vh_session_list/1

get_vh_session_number/1

Specs

ejabberd_sm.get_vh_session_number/1

handle_call/3

Callback implementation for :gen_server.handle_call/3.

handle_cast/2

Callback implementation for :gen_server.handle_cast/2.

handle_info/2

Specs

ejabberd_sm.handle_info/2

Callback implementation for :gen_server.handle_info/2.

hooks/1

init/1

Specs

ejabberd_sm.init/1

Callback implementation for :gen_server.init/1.

is_offline/1

Specs

ejabberd_sm.is_offline/1

is_privacy_allow/4

Specs

ejabberd_sm.is_privacy_allow/4

The default list applies to the user as a whole, and is processed if there is no active list set for the target session/resource to which a stanza is addressed, or if there are no current sessions for the user.

is_privacy_allow/5

Specs

ejabberd_sm.is_privacy_allow/5

Check if privacy rules allow this delivery Function copied from ejabberd_c2s.erl

make_new_sid/0

Specs

ejabberd_sm.make_new_sid/0

node_cleanup/2

open_session/3

Specs

ejabberd_sm.open_session/3

open_session/4

Specs

ejabberd_sm.open_session/4

open_session/5

open_session/6

process_iq/4

Specs

ejabberd_sm.process_iq/4

process_iq/5

register_iq_handler/4

register_iq_handler/5

remove_info/2

Specs

ejabberd_sm.remove_info/2

remove_info/4

route/3

Specs

ejabberd_sm.route/3

route/4

route_message/4

Specs

ejabberd_sm.route_message/4

route_message_by_type/5

set_presence/6

Specs

ejabberd_sm.set_presence/6

set_presence/8

set_session/4

Specs

ejabberd_sm.set_session/4

sm_backend/0

Specs

ejabberd_sm.sm_backend/0

sm_backend/1

Specs

ejabberd_sm.sm_backend/1

start/0

Specs

ejabberd_sm.start/0

start_link/0

Specs

ejabberd_sm.start_link/0

store_info/2

Specs

ejabberd_sm.store_info/2

store_info/4

terminate/2

Specs

ejabberd_sm.terminate/2

Callback implementation for :gen_server.terminate/2.

unregister_iq_handler/2

unset_presence/5

Specs

ejabberd_sm.unset_presence/5

unset_presence/7

update_session/4

Specs

ejabberd_sm.update_session/4

user_resources/2

Specs

ejabberd_sm.user_resources/2