eldap_utils (MongooseIM v4.0.1)

Types

check_fun()

Specs

t:eldap_utils.check_fun/0

Functions

case_insensitive_match/2

Specs

eldap_utils.case_insensitive_match/2

check_and_convert_restricted_string/2

Specs

eldap_utils.check_and_convert_restricted_string/2

check_filter/1

Specs

eldap_utils.check_filter/1

collect_parts/1

Specs

eldap_utils.collect_parts/1

collect_parts/2

Specs

eldap_utils.collect_parts/2

collect_parts_bit/3

Specs

eldap_utils.collect_parts_bit/3

decode_octet_string/3

Specs

eldap_utils.decode_octet_string/3

decode_restricted_string/3

Specs

eldap_utils.decode_restricted_string/3

find_ldap_attrs/2

Specs

eldap_utils.find_ldap_attrs/2

Not tail-recursive, but it is not very terribly. It stops finding on the first not empty value.

generate_subfilter/1

Specs

eldap_utils.generate_subfilter/1

Generate an 'or' LDAP query on one or several attributes If there is only one attribute

generate_substring_list/1

Specs

eldap_utils.generate_substring_list/1

generate_substring_list/2

get_base/1

get_deref_aliases/1

get_dn_filter_with_attrs/1

get_ldap_attr/2

Specs

eldap_utils.get_ldap_attr/2

get_mod_opt/2

Specs

eldap_utils.get_mod_opt/2

get_mod_opt/3

Specs

eldap_utils.get_mod_opt/3

get_mod_opt/4

Specs

eldap_utils.get_mod_opt/4

get_search_filter/1

get_state/2

Specs

eldap_utils.get_state/2

get_uids/2

get_user_filter/2

get_user_part/2

Specs

eldap_utils.get_user_part/2

make_filter/2

Specs

eldap_utils.make_filter/2

make_filter/3

Specs

eldap_utils.make_filter/3

match_filter_name/2

match_tags/2

maybe_b2list/1

maybe_list2b/1

prepare_opt_val/4

Specs

eldap_utils.prepare_opt_val/4

process_uids/1

singleton_value/1

Specs

eldap_utils.singleton_value/1

subfilter/1

Specs

eldap_utils.subfilter/1

Subfilter for a single attribute

traverse_filter_fun/1

uids_domain_subst/2

Specs

eldap_utils.uids_domain_subst/2

From the list of uids attribute: we look from alias domain (%d) and make the substitution with the actual host domain. This helps when you need to configure many virtual domains.