gen_mod behaviour (MongooseIM v4.0.1)

Types

dep_arguments()

Specs

t:gen_mod.dep_arguments/0

deps_list()

Specs

t:gen_mod.deps_list/0

module_deps_list()

Specs

t:gen_mod.module_deps_list/0

service_deps_list()

Specs

t:gen_mod.service_deps_list/0

Functions

clear_opts/2

Specs

gen_mod.clear_opts/2

del_module_mnesia/2

Specs

gen_mod.del_module_mnesia/2

expand_hosts/1

Specs

gen_mod.expand_hosts/1

get_deps/3

Specs

gen_mod.get_deps/3

get_module_opt/4

get_module_opt_by_subhost/4

Specs

gen_mod.get_module_opt_by_subhost/4

get_module_opt_subhost/3

Specs

gen_mod.get_module_opt_subhost/3

get_module_opts/2

get_module_proc/2

Specs

gen_mod.get_module_proc/2

get_opt/2

get_opt/3

get_opt/4

get_opt_subhost/3

Specs

gen_mod.get_opt_subhost/3

get_opt_subhost/4

Specs

gen_mod.get_opt_subhost/4

get_required_services/3

Specs

gen_mod.get_required_services/3

is_app_running/1

Specs

gen_mod.is_app_running/1

is_common_test_running/0

is_loaded/2

Specs

gen_mod.is_loaded/2

is_mim_or_ct_running/0

loaded_modules/1

Specs

gen_mod.loaded_modules/1

loaded_modules_with_opts/1

Specs

gen_mod.loaded_modules_with_opts/1

make_subhost/2

Specs

gen_mod.make_subhost/2

make_subhosts/2

Specs

gen_mod.make_subhosts/2

opts_for_module/2

Specs

gen_mod.opts_for_module/2

reload_module/3

Specs

gen_mod.reload_module/3

set_module_opt/4

Specs

gen_mod.set_module_opt/4

Non-atomic! You have been warned.

set_module_opt_by_subhost/4

Specs

gen_mod.set_module_opt_by_subhost/4

set_module_opts/3

Specs

gen_mod.set_module_opts/3

Replaces all module options

set_module_opts_mnesia/3

Specs

gen_mod.set_module_opts_mnesia/3

set_opt/3

Specs

gen_mod.set_opt/3

start/0

Specs

gen_mod.start/0

start_backend_module/2

Specs

gen_mod.start_backend_module/2

start_backend_module/3

start_module/3

Specs

gen_mod.start_module/3

start_module_for_host/3

stop_module/2

Specs

gen_mod.stop_module/2

Stop the module in a host, and forget its configuration.

stop_module_for_host/2

stop_module_keep_config/2

Specs

gen_mod.stop_module_keep_config/2

Stop the module in a host, but keep its configuration. As the module configuration is kept in the Mnesia local_config table, when ejabberd is restarted the module will be started again. This function is useful when ejabberd is being stopped and it stops all modules.

wait_for_process/1

Specs

gen_mod.wait_for_process/1

wait_for_stop/2

Specs

gen_mod.wait_for_stop/2

wait_for_stop1/1

Specs

gen_mod.wait_for_stop1/1

Callbacks

start(host, opts)

Specs

c:gen_mod.start/2

stop(host)

Specs

c:gen_mod.stop/1