gen_mod_deps (MongooseIM v4.0.1)

Types

gen_mod_list()

Specs

t:gen_mod_deps.gen_mod_list/0

gen_mod_map()

Specs

t:gen_mod_deps.gen_mod_map/0

gen_mod_params()

Specs

t:gen_mod_deps.gen_mod_params/0

hardness()

Specs

t:gen_mod_deps.hardness/0

module_deps_list()

Specs

t:gen_mod_deps.module_deps_list/0

Functions

add_deps/2

Specs

gen_mod_deps.add_deps/2

Adds deps into module list. Side-effect free.

find_edge/3

Specs

gen_mod_deps.find_edge/3

find_soft_edge/2

Specs

gen_mod_deps.find_soft_edge/2

get_deps/3

Specs

gen_mod_deps.get_deps/3

merge_args/3

Specs

gen_mod_deps.merge_args/3

Merges proplists prioritizing the new list, and warns on overrides.

process_dep/4

Specs

gen_mod_deps.process_dep/4

process_module_dep/4

Specs

gen_mod_deps.process_module_dep/4

replace_modules/3

Specs

gen_mod_deps.replace_modules/3

Replaces OldModules (along with dependencies) with NewModules (along with dependencies). The dependencies are resolved only for given lists of modules, so for certain arguments a still-needed dependency might be removed. Thus, the function is meant to replace all modules on the host.

resolve_deps/2

Specs

gen_mod_deps.resolve_deps/2

Determines all modules to start, along with their parameters.

NOTE: A dependency will not be discarded during resolving, e.g. if the resolver processes dependencies in order:

deps(mod_a, []) -> [{mod_b, []}, {mod_c, []}] deps(mod_parent, []) -> [{mod_a, [param]}]

then the dependency for mod_a will be reevaluated with new parameters and might return:

deps(mod_a, [param]) -> [{mod_c, []}]
In this case, mod_b will still be started.

resolve_deps/4

Specs

gen_mod_deps.resolve_deps/4

sort_deps/2

Specs

gen_mod_deps.sort_deps/2

start_modules/2

Specs

gen_mod_deps.start_modules/2