mod_bosh_socket (MongooseIM v4.0.1)

Types

cached_response()

Specs

t:mod_bosh_socket.cached_response/0

fsm_return()

Specs

t:mod_bosh_socket.fsm_return/0

rid()

Specs

t:mod_bosh_socket.rid/0

state()

Specs

t:mod_bosh_socket.state/0

statename()

Specs

t:mod_bosh_socket.statename/0

Functions

add_handler/2

Specs

mod_bosh_socket.add_handler/2

bosh_body/1

Specs

mod_bosh_socket.bosh_body/1

Bosh body for an ordinary stream element(s).

bosh_stream_end_body/0

Specs

mod_bosh_socket.bosh_stream_end_body/0

bosh_stream_start_body/2

Specs

mod_bosh_socket.bosh_stream_start_body/2

Bosh body for a session creation response.

bosh_unwrap/3

Specs

mod_bosh_socket.bosh_unwrap/3

bosh_wrap/3

Specs

mod_bosh_socket.bosh_wrap/3

cache_response/2

Specs

mod_bosh_socket.cache_response/2

cache_up_to/2

Specs

mod_bosh_socket.cache_up_to/2

cancel_inactivity_timer/1

Specs

mod_bosh_socket.cancel_inactivity_timer/1

change_shaper/2

Specs

mod_bosh_socket.change_shaper/2

close/1

Specs

mod_bosh_socket.close/1

closing/2

closing/3

code_change/4

compress/1

Specs

mod_bosh_socket.compress/1

Should be negotiated on HTTP level.

compress/3

Specs

mod_bosh_socket.compress/3

determine_next_state/3

determine_report_action/4

Specs

mod_bosh_socket.determine_report_action/4

forward_to_c2s/2

Specs

mod_bosh_socket.forward_to_c2s/2

get_attr/3

Specs

mod_bosh_socket.get_attr/3

get_cached_responses/1

Specs

mod_bosh_socket.get_cached_responses/1

get_client_acks/1

Specs

mod_bosh_socket.get_client_acks/1

get_client_acks/4

Specs

mod_bosh_socket.get_client_acks/4

get_handlers/1

get_maxpause/0

get_peer_certificate/1

Specs

mod_bosh_socket.get_peer_certificate/1

get_pending/1

get_sockmod/1

Specs

mod_bosh_socket.get_sockmod/1

handle_request/2

Specs

mod_bosh_socket.handle_request/2

handle_stream_event/4

ignore_undefined/1

inactivity/1

Specs

mod_bosh_socket.inactivity/1

is_acceptable_rid/2

Specs

mod_bosh_socket.is_acceptable_rid/2

is_expected_rid/2

Specs

mod_bosh_socket.is_expected_rid/2

is_stream_event/1

Specs

mod_bosh_socket.is_stream_event/1

is_stream_prefix/1

is_valid_ack/2

Specs

mod_bosh_socket.is_valid_ack/2

last_processed/2

Specs

mod_bosh_socket.last_processed/2

ls/2

maxpause/1

Specs

mod_bosh_socket.maxpause/1

maybe_ack/2

Specs

mod_bosh_socket.maybe_ack/2

maybe_add/2

Specs

mod_bosh_socket.maybe_add/2

maybe_add_default_ns/1

maybe_add_default_ns_to_children/1

maybe_add_handler/3

Specs

mod_bosh_socket.maybe_add_handler/3

maybe_diff/2

Specs

mod_bosh_socket.maybe_diff/2

maybe_is_retransmission/3

Specs

mod_bosh_socket.maybe_is_retransmission/3

maybe_report/1

Specs

mod_bosh_socket.maybe_report/1

maybe_send_report/1

Specs

mod_bosh_socket.maybe_send_report/1

maybe_stream_prefix/2

maybe_trim_cache/2

Specs

mod_bosh_socket.maybe_trim_cache/2

monitor/1

Specs

mod_bosh_socket.monitor/1

Callback implementation for c::mongoose_transport.monitor/1.

normal/2

Specs

mod_bosh_socket.normal/2

normal/3

pause/2

Specs

mod_bosh_socket.pause/2

TODO: no handler for this call is present! No check for violating maxpause is made when calling this!

peername/1

Specs

mod_bosh_socket.peername/1

Callback implementation for c::mongoose_transport.peername/1.

pick_handler/1

Specs

mod_bosh_socket.pick_handler/1

process_acked_stream_event/3

Specs

mod_bosh_socket.process_acked_stream_event/3

process_deferred_events/2

Specs

mod_bosh_socket.process_deferred_events/2

process_pause_event/2

Specs

mod_bosh_socket.process_pause_event/2

process_stream_event/4

Specs

mod_bosh_socket.process_stream_event/4

resend_cached/2

Specs

mod_bosh_socket.resend_cached/2

return_surplus_handlers/2

Specs

mod_bosh_socket.return_surplus_handlers/2

Keep in mind the hardcoding for hold == 1.

rid/2

schedule_report/2

Specs

mod_bosh_socket.schedule_report/2

send/2

Specs

mod_bosh_socket.send/2

send_or_store/2

Specs

mod_bosh_socket.send_or_store/2

Send data to the client if a request handler is available, that matches next RID. Otherwise, store for sending later.

send_to_handler/2

Specs

mod_bosh_socket.send_to_handler/2

send_to_handler() assumes that Handlers is not empty! Be sure that's the case if calling it.

send_to_handler/3

Specs

mod_bosh_socket.send_to_handler/3

send_wrapped_to_handler/3

Specs

mod_bosh_socket.send_wrapped_to_handler/3

This is the most specific variant of send_to_handler() and the *only one* actually performing a send to the cowboy_loop_handler serving a HTTP request.

send_xml/2

Specs

mod_bosh_socket.send_xml/2

Callback implementation for c::mongoose_transport.send_xml/2.

server_ack/2

Specs

mod_bosh_socket.server_ack/2

set_client_acks/2

Specs

mod_bosh_socket.set_client_acks/2

setup_inactivity_timer/1

Specs

mod_bosh_socket.setup_inactivity_timer/1

start/3

Specs

mod_bosh_socket.start/3

start_link/3

Specs

mod_bosh_socket.start_link/3

start_supervisor/0

starttls/2

Specs

mod_bosh_socket.starttls/2

Should be negotiated on HTTP level.

starttls/3

Specs

mod_bosh_socket.starttls/3

Callback implementation for c::mongoose_transport.starttls/3.

store/2

Specs

mod_bosh_socket.store/2

Store data for sending later.

stream_start/2

Specs

mod_bosh_socket.stream_start/2

terminate/3