mod_cowboy (MongooseIM v4.0.1)

HTTP routing layer for MongooseIM's Cowboy listener

Allows to set more than one handler for an endpoint: - one handler for http - one handler for each WS protocol

It's a proxy between cowboy and actual handlers.

So, this module handles calls from cowboy, and forwards the calls to the handler modules.
I.e. it both implements and uses the cowboy_handler and cowboy_websocket behaviours.

Types

option()

Specs

t:mod_cowboy.option/0

state()

Specs

t:mod_cowboy.state/0

Functions

handle_http_init/2

handle_websocket_init/1

handle_ws_init/2

http_handler/1

init/2

Specs

mod_cowboy.init/2

init_http_handler/3

init_state/4

init_state/5

init_unsafe/2

init_ws_handler/4

protocol/1

terminate/3

Specs

mod_cowboy.terminate/3

update_handler_state/2

websocket_handle/2

websocket_handle_unsafe/2

websocket_info/2

websocket_info_unsafe/2

websocket_init/1

ws_handler/2

ws_protocol/1