mod_keystore behaviour (MongooseIM v4.0.1)

Types

key()

Specs

t:mod_keystore.key/0

key_id()

Specs

t:mod_keystore.key_id/0

key_list()

Specs

t:mod_keystore.key_list/0

key_name()

Specs

t:mod_keystore.key_name/0

A key ID is used to uniquely identify a key for storage backends. It's used to maintain separate instances of a key with the same name for different virtual hosts.

key_type()

Specs

t:mod_keystore.key_type/0

raw_key()

Specs

t:mod_keystore.raw_key/0

Functions

clear_keystore_ets/1

config_metrics/1

Callback implementation for c::mongoose_module_metrics.config_metrics/1.

create_keystore_ets/0

does_table_exist/1

ets_get_key/1

ets_store_key/2

get_key/2

Specs

mod_keystore.get_key/2

get_key_size/1

hook_handlers/0

init_key/3

Specs

mod_keystore.init_key/3

init_keys/2

maybe_add_heir/3

start/2

Specs

mod_keystore.start/2

Callback implementation for c::gen_mod.start/2.

stop/1

Specs

mod_keystore.stop/1

Callback implementation for c::gen_mod.stop/1.

validate_key_ids/1

validate_opts/1

Callbacks

get_key(iD)

Specs

c:mod_keystore.get_key/1

init(domain, opts)

Specs

c:mod_keystore.init/2

init_ram_key(proposedKey)

Specs

c:mod_keystore.init_ram_key/1