mod_last_riak (MongooseIM v4.0.1)

Riak backend for last activity XEP

The backend uses the existing riak connection pool, which is "globally" defined in the ejebberd.cfg file. Therefore, we don't need to start anything in the init function.

The module follows the approach taken by the other riak backends - it creates the following bucket {<<"last">>, <<"example.com">>} for each xmpp domain. The <<"last">> bucket type has the following props: {"props":{"last_write_wins":true}}

Basically, there are 3 operations the backend need to provide set/get and count the number of users which has been active since given date. However, the count query can be only called using mongooseimctl.

Username is used as a key and the value is just a user status, which is just a binary. Moreover, we have a secondary integer index on the timestamp, which is used in the count active users query.

   Data Layout:
   KV: {Username:binary, Status:binary}
   2i: [Timestamp:integer]

Set/Get are rather simple operations - they map to riak's put/get functions. The count query uses mapred task to calculate the number of users satisfying the criteria. As the input timestamp index range is used. The approach is described here: http://docs.basho.com/riak/latest/dev/using/2i/#Querying in the *Count Bucket Objects via $bucket Index* section. The alternative approach could be the "riakc_pb_socket:get_index_range" function
The pros and cons of the both approaches have been discussed here: https://github.com/esl/MongooseIM/pull/567

Functions

bucket_type/1

count_active_users/2

Specs

mod_last_riak.count_active_users/2

Callback implementation for c::mod_last.count_active_users/2.

get_last/2

Specs

mod_last_riak.get_last/2

Callback implementation for c::mod_last.get_last/2.

infinity/0

Specs

mod_last_riak.infinity/0

init/2

Specs

mod_last_riak.init/2

Callback implementation for c::mod_last.init/2.

remove_user/2

Specs

mod_last_riak.remove_user/2

Callback implementation for c::mod_last.remove_user/2.

set_last_info/4

Specs

mod_last_riak.set_last_info/4

Callback implementation for c::mod_last.set_last_info/4.