mod_mam_cassandra_arch behaviour (MongooseIM v4.0.1)

RDBMS backend for Message Archive Management.

Types

filter()

Specs

t:mod_mam_cassandra_arch.filter/0

message_id()

Specs

t:mod_mam_cassandra_arch.message_id/0

server_hostname()

Specs

t:mod_mam_cassandra_arch.server_hostname/0

Functions

after_id/2

archive_message/3

Callback implementation for c::ejabberd_gen_mam_archive.archive_message/3.

archive_message2/1

archive_size/4

Callback implementation for c::ejabberd_gen_mam_archive.archive_size/4.

before_id/2

calc_before/5

Specs

mod_mam_cassandra_arch.calc_before/5

Count of elements in RSet before the passed element.
The element with the passed UID can be already deleted.

calc_count/4

Specs

mod_mam_cassandra_arch.calc_count/4

Get the total result set size. "SELECT COUNT(*) as "count" FROM mam_message WHERE "

calc_count_cql/1

calc_count_queries/0

calc_index/5

Specs

mod_mam_cassandra_arch.calc_index/5

Calculate a zero-based index of the row with UID in the result test.
If the element does not exists, the ID of the next element will be returned instead.

calc_offset/7

Specs

mod_mam_cassandra_arch.calc_offset/7

calc_offset_to_start_id/4

Specs

mod_mam_cassandra_arch.calc_offset_to_start_id/4

Convert offset to index of the first entry Returns undefined if not there are not enough rows

compile_params_module/1

db_message_format/0

Specs

mod_mam_cassandra_arch.db_message_format/0

ensure_params_loaded/1

eval_filter_params/1

expand_simple_param/1

extract_messages/6

Specs

mod_mam_cassandra_arch.extract_messages/6

extract_messages_cql/1

extract_messages_queries/0

extract_messages_r_cql/1

extract_messages_r_queries/0

fetch_user_messages/3

fetch_user_messages_cql/0

filter_to_cql/0

from_id/2

get_mam_pm_gdpr_data/2

Specs

mod_mam_cassandra_arch.get_mam_pm_gdpr_data/2

Callback implementation for c::ejabberd_gen_mam_archive.get_mam_pm_gdpr_data/2.

insert_offset_hint_query_cql/0

Insert offset -> message id 'hint'

insert_query_cql/0

list_message_ids_cql/1

list_message_ids_queries/0

lookup_messages/3

Specs

mod_mam_cassandra_arch.lookup_messages/3

Callback implementation for c::ejabberd_gen_mam_archive.lookup_messages/3.

lookup_messages2/10

lookup_messages_after_id/6

lookup_messages_before_id/6

lookup_messages_by_offset/6

lookup_messages_first_page/6

lookup_messages_last_page/6

lookup_messages_simple/6

maybe_full_jid/1

Specs

mod_mam_cassandra_arch.maybe_full_jid/1

maybe_save_offset_hint/6

Specs

mod_mam_cassandra_arch.maybe_save_offset_hint/6

Saves offset hint for future use in order to speed up queries with similar offset Hint is save only if previous offset hint was 50+ entires from current query This function returns given StartId as passthrough for convenience

message_to_params/1

offset_to_start_id/4

Specs

mod_mam_cassandra_arch.offset_to_start_id/4

Convert offset to index of the first entry Returns undefined if not there are not enough rows Uses previously calculated offsets to speed up queries

packet_to_stored_binary/1

params_helper/1

pool_name/0

Specs

mod_mam_cassandra_arch.pool_name/0

prepare_filter/5

prepare_filter_cql/2

prepare_filter_params/4

prepared_queries/0

Callback implementation for c::mongoose_cassandra.prepared_queries/0.

prev_offset_query_cql/0

Get closest offset -> message id 'hint' for specified offset

remove_archive/2

remove_archive/4

remove_archive_offsets_query_cql/0

remove_archive_query_cql/0

row_to_message_id/1

row_to_uniform_format/1

rows_to_gdpr_mam_message/1

rows_to_uniform_format/1

rsm_to_strategy/1

select_filter/1

select_filter/2

Specs

mod_mam_cassandra_arch.select_filter/2

select_for_removal_query_cql/0

simple_params/0

start/2

Callback implementation for c::gen_mod.start/2.

start_pm/2

stop/1

Callback implementation for c::gen_mod.stop/1.

stop_pm/1

stored_binary_to_packet/1

to_id/2

write_messages/2

Callbacks

decode(binary)

Specs

c:mod_mam_cassandra_arch.decode/1

encode(binary)

Specs

c:mod_mam_cassandra_arch.encode/1