mod_mam_muc behaviour (MongooseIM v4.0.1)

XEP-0313: Message Archive Management

The module uses several backend modules:

Preference manager (mod_mam_muc_rdbms_prefs);

Writer (mod_mam_muc_rdbms_arch or mod_mam_muc_rdbms_async_pool_writer);

Archive manager (mod_mam_muc_rdbms_arch);

User's ID generator (mod_mam_muc_user).

Preferencies can be also stored in Mnesia (mod_mam_mnesia_prefs). This module handles MUC archives.

This module should be started for each host. Message archivation is not shaped here (use standard support for this). MAM's IQs are shaped inside shaper_srv.

Message identifiers (or UIDs in the spec) are generated based on:

date (using timestamp());

node number (using ejabberd_node_id).

Types

packet()

Specs

t:mod_mam_muc.packet/0

row()

Specs

t:mod_mam_muc.row/0

row_batch()

Specs

t:mod_mam_muc.row_batch/0

Functions

archive_id/2

Specs

mod_mam_muc.archive_id/2

archive_id_int/2

Specs

mod_mam_muc.archive_id_int/2

archive_message/2

Specs

mod_mam_muc.archive_message/2

archive_room_packet/6

Specs

mod_mam_muc.archive_room_packet/6

Archive without validation.

archive_size/2

Specs

mod_mam_muc.archive_size/2

archive_size/3

Specs

mod_mam_muc.archive_size/3

can_access_room/2

Specs

mod_mam_muc.can_access_room/2

delete_archive/2

Specs

mod_mam_muc.delete_archive/2

filter_room_packet/2

Specs

mod_mam_muc.filter_room_packet/2

Handle public MUC-message.

forget_room/2

Specs

mod_mam_muc.forget_room/2

This hook is called from mod_muc:forget_room(Host, Name).

forget_room/3

Specs

mod_mam_muc.forget_room/3

This hook is called from mod_muc:forget_room(Host, Name).

forward_messages/6

get_behaviour/5

Specs

mod_mam_muc.get_behaviour/5

get_personal_data/2

Specs

mod_mam_muc.get_personal_data/2

get_prefs/4

Specs

mod_mam_muc.get_prefs/4

Load settings from the database.

handle_error_iq/5

Specs

mod_mam_muc.handle_error_iq/5

handle_get_message_form/3

Specs

mod_mam_muc.handle_get_message_form/3

handle_get_prefs/2

Specs

mod_mam_muc.handle_get_prefs/2

handle_get_prefs_result/3

handle_lookup_result/4

Specs

mod_mam_muc.handle_lookup_result/4

handle_mam_iq/4

Specs

mod_mam_muc.handle_mam_iq/4

handle_set_message_form/3

Specs

mod_mam_muc.handle_set_message_form/3

handle_set_prefs/2

Specs

mod_mam_muc.handle_set_prefs/2

handle_set_prefs_result/5

is_action_allowed/3

Specs

mod_mam_muc.is_action_allowed/3

is_action_allowed_by_default/3

Specs

mod_mam_muc.is_action_allowed_by_default/3

is_archivable_message/3

Specs

mod_mam_muc.is_archivable_message/3

is_room_action_allowed_by_default/3

Specs

mod_mam_muc.is_room_action_allowed_by_default/3

is_room_owner/2

Specs

mod_mam_muc.is_room_owner/2

is_user_identity_hidden/2

Specs

mod_mam_muc.is_user_identity_hidden/2

Return true if user element should be removed from results

lookup_messages/2

Specs

mod_mam_muc.lookup_messages/2

lookup_messages_without_policy_violation_check/2

maybe_delete_x_user_element/3

message_row_to_ext_id/1

Specs

mod_mam_muc.message_row_to_ext_id/1

message_row_to_xml/6

Specs

mod_mam_muc.message_row_to_xml/6

remove_archive/3

Specs

mod_mam_muc.remove_archive/3

replace_from_to_attributes/2

report_issue/4

report_issue/5

return_action_not_allowed_error_iq/1

Specs

mod_mam_muc.return_action_not_allowed_error_iq/1

return_error_iq/2

return_max_delay_reached_error_iq/1

Specs

mod_mam_muc.return_max_delay_reached_error_iq/1

return_message_form_iq/2

room_process_mam_iq/4

Specs

mod_mam_muc.room_process_mam_iq/4

To is an account or server entity hosting the archive. Servers that archive messages on behalf of local users SHOULD expose archives to the user on their bare JID (i.e. From.luser), while a MUC service might allow MAM queries to be sent to the room's bare JID (i.e To.luser).

send_messages_and_iq_result/4

set_prefs/6

Specs

mod_mam_muc.set_prefs/6

start/2

Specs

mod_mam_muc.start/2

stop/1

Specs

mod_mam_muc.stop/1

Callbacks

is_complete_message(module, dir, packet)

Specs

c:mod_mam_muc.is_complete_message/3