mod_mam_rdbms_async_pool_writer (MongooseIM v4.0.1)

Collect messages and flush them into the database.

Some backends allow ArcID can be not set (riak?). This means, that there are no "jid => integer" mapping.

But this module requires ArcID to be an integer. ArcID is set by mod_mam_rdbms_user. If ArcID is undefined, it means that there can be issues in mod_mam_rdbms_user.
We have is_integer(ArcID) check on each hook handler, so cases when ArcID is undefined would fail at the module entrance.

Functions

archive_message/3

Specs

mod_mam_rdbms_async_pool_writer.archive_message/3

Callback implementation for c::ejabberd_gen_mam_archive.archive_message/3.

archive_size/4

Specs

mod_mam_rdbms_async_pool_writer.archive_size/4

Callback implementation for c::ejabberd_gen_mam_archive.archive_size/4.

cancel_and_flush_timer/1

code_change/3

do_run_flush/2

handle_cast/2

handle_info/2

init/1

is_overloaded/1

lookup_messages/3

Specs

mod_mam_rdbms_async_pool_writer.lookup_messages/3

Callback implementation for c::ejabberd_gen_mam_archive.lookup_messages/3.

queue_length/1

queue_lengths/1

run_flush/1

select_worker/2

start/2

Callback implementation for c::gen_mod.start/2.

start_muc/2

start_pm/2

start_worker/4

start_workers/2

stop/1

Callback implementation for c::gen_mod.stop/1.

stop_muc/1

stop_pm/1

stop_worker/1

stop_workers/1

terminate/2

worker_count/1

worker_name/2

worker_names/1

worker_number/2

worker_prefix/0

worker_queue_length/1