mod_mam_utils (MongooseIM v4.0.1)

General functions for MAM.

Types

archive_behaviour()

Specs

t:mod_mam_utils.archive_behaviour/0

archive_behaviour_bin()

Specs

t:mod_mam_utils.archive_behaviour_bin/0

<<"roster">> | <<"always">> | <<"never">>.

iso8601_datetime_binary()

Specs

t:mod_mam_utils.iso8601_datetime_binary/0

Microseconds from 01.01.1970

ne_binary()

Specs

t:mod_mam_utils.ne_binary/0

unix_timestamp()

Specs

t:mod_mam_utils.unix_timestamp/0

Functions

action_to_global_shaper_name/1

Specs

mod_mam_utils.action_to_global_shaper_name/1

action_to_shaper_name/1

Specs

mod_mam_utils.action_to_shaper_name/1

append_arcid_elem/4

Specs

mod_mam_utils.append_arcid_elem/4

append_x_user_element/4

apply_end_border/2

Specs

mod_mam_utils.apply_end_border/2

apply_start_border/2

Specs

mod_mam_utils.apply_start_border/2

bare_jid/1

Specs

mod_mam_utils.bare_jid/1

binary_jid_to_lower/1

Normalize JID to be used when comparing JIDs in DB

borders/4

Specs

mod_mam_utils.borders/4

borders_decode/1

Specs

mod_mam_utils.borders_decode/1

calculate_msg_id_borders/3

Specs

mod_mam_utils.calculate_msg_id_borders/3

calculate_msg_id_borders/4

Specs

mod_mam_utils.calculate_msg_id_borders/4

check_for_item_not_found/3

Specs

mod_mam_utils.check_for_item_not_found/3

Check for XEP-313 item-not-found error condition, that is if a message ID passed in a before/after query is actually present in the archive. See https://xmpp.org/extensions/xep-0313.html#query-paging for details.
In a backend it's reasonable to query for PageSize + 1 messages, so that once the interval endpoint with requested ID is discarded we actually return (up to) PageSize messages.

check_result_for_policy_violation/2

Specs

mod_mam_utils.check_result_for_policy_violation/2

decode_compact_uuid/1

Specs

mod_mam_utils.decode_compact_uuid/1

Extract date and node id from a message id.

decode_optimizations/1

Specs

mod_mam_utils.decode_optimizations/1

delay/2

Specs

mod_mam_utils.delay/2

delete_arcid_elem/3

Specs

mod_mam_utils.delete_arcid_elem/3

delete_x_user_element/1

Specs

mod_mam_utils.delete_x_user_element/1

encode_compact_uuid/2

Specs

mod_mam_utils.encode_compact_uuid/2

Create a message ID (UID).
It removes a leading 0 from 64-bit binary representation. It puts node id as a last byte. The maximum date, that can be encoded is {{4253, 5, 31}, {22, 20, 37}}.

encode_jids/1

Specs

mod_mam_utils.encode_jids/1

error_on_sql_error/3

expand_minified_jid/2

Specs

mod_mam_utils.expand_minified_jid/2

expand_minified_jid/3

Specs

mod_mam_utils.expand_minified_jid/3

external_binary_to_mess_id/1

Specs

mod_mam_utils.external_binary_to_mess_id/1

Decode a message ID received from the user.

features/2

field_to_value/1

Specs

mod_mam_utils.field_to_value/1

find_field/2

Specs

mod_mam_utils.find_field/2

Return first matched field

form_borders_decode/1

Specs

mod_mam_utils.form_borders_decode/1

form_decode_optimizations/1

Specs

mod_mam_utils.form_decode_optimizations/1

form_field/2

form_field_mess_id/2

Specs

mod_mam_utils.form_field_mess_id/2

form_field_value/2

Specs

mod_mam_utils.form_field_value/2

form_field_value_s/2

form_to_text/1

Specs

mod_mam_utils.form_to_text/1

form_type_field/1

forwarded/3

Specs

mod_mam_utils.forwarded/3

full_jid/1

Specs

mod_mam_utils.full_jid/1

generate_message_id/0

Specs

mod_mam_utils.generate_message_id/0

get_one_of_path/2

Specs

mod_mam_utils.get_one_of_path/2

get_one_of_path/3

Specs

mod_mam_utils.get_one_of_path/3

get_origin_id/1

get_retract_id/1

get_retract_id/3

has_any/1

has_chat_marker/1

has_full_text_search/2

Specs

mod_mam_utils.has_full_text_search/2

has_message_retraction/2

Specs

mod_mam_utils.has_message_retraction/2

is_archivable_message/4

Specs

mod_mam_utils.is_archivable_message/4

In order to be archived, the message must be of type "normal", "chat" or "groupchat". It also must include a body or chat marker, as long as it doesn't include "result", "delay" or "no-store" elements.

is_arcid_elem_for/3

Specs

mod_mam_utils.is_arcid_elem_for/3

Return true, if the first element points on By.

is_complete_result_page/4

Specs

mod_mam_utils.is_complete_result_page/4

Returns true, if the current page is the final one in the result set. If there are more pages with messages, than this function returns false.

PageSize - maximum number of messages extracted in one lookup. TotalCount - total number of messages in the Result Set. Result Set - is subset of all messages in user's archive, in a specified time period. MessageRows - stuff we are about to send to the user. Params - lookup parameters, coming from mam_iq module.

TotalCount and Offset can be undefined, in case we use IsSimple=true. IsSimple=true tells the server not to do heavy SELECT COUNT(*) queries.
TODO It is possible to set complete flag WITH IsSimple=true, if we select one extra message from archive, but don't send it to the client. It's the most efficient way to query archive, if the client side does not care about the total number of messages and if it's stateless (i.e. web interface).

is_jid_element/1

is_jid_in_user_roster/2

Specs

mod_mam_utils.is_jid_in_user_roster/2

is_mam_namespace/1

is_mam_result_message/1

is_most_recent_page/4

Specs

mod_mam_utils.is_most_recent_page/4

Returns true, if the current page contains the most recent messages. If there are some more recent messages in archive, this function returns false.

is_policy_violation/4

is_valid_message/4

is_valid_message_type/3

is_x_user_element/1

jid_to_opt_binary/2

Specs

mod_mam_utils.jid_to_opt_binary/2

make_fin_element/4

Specs

mod_mam_utils.make_fin_element/4

mam_features/0

maybe_add_arcid_elems/4

Specs

mod_mam_utils.maybe_add_arcid_elems/4

maybe_encode_compact_uuid/2

Specs

mod_mam_utils.maybe_encode_compact_uuid/2

maybe_external_binary_to_mess_id/1

Specs

mod_mam_utils.maybe_external_binary_to_mess_id/1

maybe_get_result_namespace/1

maybe_integer/2

Specs

mod_mam_utils.maybe_integer/2

maybe_last/1

Specs

mod_mam_utils.maybe_last/1

maybe_log_deprecation/1

maybe_max/2

Specs

mod_mam_utils.maybe_max/2

maybe_microseconds/1

Specs

mod_mam_utils.maybe_microseconds/1

mod_mam_utils.maybe_microseconds/1

Return a unix timestamp in microseconds.
"maybe" means, that the function may return undefined.

maybe_min/2

Specs

mod_mam_utils.maybe_min/2

maybe_next_id/1

Specs

mod_mam_utils.maybe_next_id/1

maybe_previous_id/1

Specs

mod_mam_utils.maybe_previous_id/1

maybe_set_client_xmlns/2

Specs

mod_mam_utils.maybe_set_client_xmlns/2

mess_id_to_external_binary/1

Specs

mod_mam_utils.mess_id_to_external_binary/1

Encode a message ID to pass it to the user.

message_form/3

Specs

mod_mam_utils.message_form/3

message_form_fields/3

normalize_search_text/1

Specs

mod_mam_utils.normalize_search_text/1

Normalize given text to improve text search in some MAM backends. This normalization involves making text all lowercase, replacing some word separators ([, .:;-?!]) with given one (by default "%") and removing all unicode characters that are considered non-alphanumerical. For example, text: "My cat, was eaten by: my dog?!? Why...?!?" will be normalized as: "my%cat%was%eaten%by%my%dog%why"

normalize_search_text/2

Specs

mod_mam_utils.normalize_search_text/2

packet_to_search_body/3

Specs

mod_mam_utils.packet_to_search_body/3

packet_to_x_user_jid/1

Specs

mod_mam_utils.packet_to_x_user_jid/1

parse_jid_list/2

Specs

mod_mam_utils.parse_jid_list/2

parse_prefs/1

Specs

mod_mam_utils.parse_prefs/1

replace_arcid_elem/4

Specs

mod_mam_utils.replace_arcid_elem/4

replace_from_attribute/2

replace_x_user_element/4

Specs

mod_mam_utils.replace_x_user_element/4

result/4

Specs

mod_mam_utils.result/4

Generates tag <result />. This element will be added in each forwarded message.

result_prefs/4

Specs

mod_mam_utils.result_prefs/4

result_query/2

Specs

mod_mam_utils.result_query/2

result_set/4

Specs

mod_mam_utils.result_set/4

Generates <set /> tag.
This element will be added into "iq/query".

retracted_element/1

retraction_features/2

rsm_ns_binary/0

send_message/3

Specs

mod_mam_utils.send_message/3

skip_bad_jids/1

success_sql_execute/3

Specs

mod_mam_utils.success_sql_execute/3

success_sql_query/2

Specs

mod_mam_utils.success_sql_query/2

tag_id/2

Specs

mod_mam_utils.tag_id/2

tombstone/2

undefined_to_empty/1

valid_behavior/1

Specs

mod_mam_utils.valid_behavior/1

wait_shaper/3

Specs

mod_mam_utils.wait_shaper/3

wrap_message/6

Specs

mod_mam_utils.wrap_message/6

Forms <forwarded/> element, according to the XEP.

wrap_message/7

Specs

mod_mam_utils.wrap_message/7

wrapper_id/0

Specs

mod_mam_utils.wrapper_id/0

Returns a UUIDv4 canonical form binary.

x_user_item/3