mod_muc (MongooseIM v4.0.1)

Types

access()

Specs

t:mod_muc.access/0

affiliation()

Specs

t:mod_muc.affiliation/0

from_to_packet()

Specs

t:mod_muc.from_to_packet/0

muc_online_room()

Specs

t:mod_muc.muc_online_room/0

muc_registered()

Specs

t:mod_muc.muc_registered/0

muc_room()

Specs

t:mod_muc.muc_room/0

nick()

Specs

t:mod_muc.nick/0

packet()

Specs

t:mod_muc.packet/0

role()

Specs

t:mod_muc.role/0

room()

Specs

t:mod_muc.room/0

room_host()

Specs

t:mod_muc.room_host/0

state()

Specs

t:mod_muc.state/0

Functions

all_room_pids/1

broadcast_service_message/2

Specs

mod_muc.broadcast_service_message/2

can_access_identity/3

Specs

mod_muc.can_access_identity/3

can_access_room/3

Specs

mod_muc.can_access_room/3

can_use_nick/4

Specs

mod_muc.can_use_nick/4

check_user_can_create_room/4

Specs

mod_muc.check_user_can_create_room/4

clean_table_from_bad_node/1

Specs

mod_muc.clean_table_from_bad_node/1

clean_table_from_bad_node/2

Specs

mod_muc.clean_table_from_bad_node/2

code_change/3

Callback implementation for :gen_server.code_change/3.

config_metrics/1

Callback implementation for c::mongoose_module_metrics.config_metrics/1.

count_hibernated_rooms/1

count_hibernated_rooms/2

create_instant_room/5

Specs

mod_muc.create_instant_room/5

Create a room. If Opts = default, the default room options are used. Else use the passed options as defined in mod_muc_room.

default_host/0

Specs

mod_muc.default_host/0

ensure_metrics/1

flush/0

Specs

mod_muc.flush/0

forget_room/3

Specs

mod_muc.forget_room/3

get_nick/3

get_persistent_vh_rooms/1

Specs

mod_muc.get_persistent_vh_rooms/1

get_registered_room_or_route_error/3

Specs

mod_muc.get_registered_room_or_route_error/3

get_registered_room_or_route_error_from_packet/6

get_registered_room_or_route_error_from_presence/6

get_room_pos/2

Specs

mod_muc.get_room_pos/2

Return the position of desired room in the list of rooms. The room must exist in the list. The count starts in 0.

get_room_pos/3

get_vh_rooms/1

Specs

mod_muc.get_vh_rooms/1

get_vh_rooms/2

Specs

mod_muc.get_vh_rooms/2

handle_call/3

Callback implementation for :gen_server.handle_call/3.

handle_cast/2

Callback implementation for :gen_server.handle_cast/2.

handle_info/2

Callback implementation for :gen_server.handle_info/2.

hibernated_rooms_number/0

hibernated_rooms_number/1

init/1

Specs

mod_muc.init/1

Callback implementation for :gen_server.init/1.

iq_disco_info/3

Specs

mod_muc.iq_disco_info/3

iq_disco_items/4

Specs

mod_muc.iq_disco_items/4

iq_get_register_info/4

Specs

mod_muc.iq_get_register_info/4

iq_get_unique/1

Specs

mod_muc.iq_get_unique/1

Get a pseudo unique Room Name. The Room Name is generated as a hash of the requester JID, the local time and a random salt.
<<"pseudo">> because we don't verify that there is not a room with the returned Name already created, nor mark the generated Name as <<"already used">>. But in practice, it is unique enough. See http://xmpp.org/extensions/xep-0045.html#createroom-unique

iq_get_vcard/1

Specs

mod_muc.iq_get_vcard/1

iq_set_register_info/5

Specs

mod_muc.iq_set_register_info/5

iq_set_unregister_info/4

Specs

mod_muc.iq_set_unregister_info/4

is_room_owner/3

Specs

mod_muc.is_room_owner/3

load_permanent_rooms/6

Specs

mod_muc.load_permanent_rooms/6

muc_room_pid/2

Specs

mod_muc.muc_room_pid/2

online_rooms_number/0

online_rooms_number/1

process_iq_disco_items/4

Specs

mod_muc.process_iq_disco_items/4

process_iq_register_set/5

Specs

mod_muc.process_iq_register_set/5

process_packet/5

Specs

mod_muc.process_packet/5

Callback implementation for c::mongoose_packet_handler.process_packet/5.

process_register/7

Specs

mod_muc.process_register/7

register_room/3

Specs

mod_muc.register_room/3

register_room_or_stop_if_duplicate/3

restore_room/3

Specs

mod_muc.restore_room/3

room_destroyed/3

Specs

mod_muc.room_destroyed/3

This function is called by a room in three situations: A) The owner of the room destroyed it B) The only participant of a temporary room leaves it C) mod_muc:stop was called, and each room is being terminated In this case, the mod_muc process died before the room processes So the message sending must be catched

room_jid_to_pid/1

Specs

mod_muc.room_jid_to_pid/1

room_to_item/4

route_by_nick/3

Specs

mod_muc.route_by_nick/3

route_by_type/3

Specs

mod_muc.route_by_type/3

route_to_online_room/2

route_to_room/3

Specs

mod_muc.route_to_room/3

set_nick/4

set_persistent_rooms_timer/1

start/2

Specs

mod_muc.start/2

Callback implementation for c::gen_mod.start/2.

start_link/2

Specs

mod_muc.start_link/2

start_new_room/11

Specs

mod_muc.start_new_room/11

start_supervisor/1

Specs

mod_muc.start_supervisor/1

stop/1

Specs

mod_muc.stop/1

Callback implementation for c::gen_mod.stop/1.

stop_gen_server/1

stop_if_hibernated/3

stop_if_hibernated/4

stop_if_hibernated_for_specified_time/4

stop_supervisor/1

Specs

mod_muc.stop_supervisor/1

store_room/4

Specs

mod_muc.store_room/4

terminate/2

Callback implementation for :gen_server.terminate/2.

unset_nick/3

xfield/5

Specs

mod_muc.xfield/5