mod_muc_log (MongooseIM v4.0.1)

Types

command()

Specs

t:mod_muc_log.command/0

dir_name()

Specs

t:mod_muc_log.dir_name/0

dir_type()

Specs

t:mod_muc_log.dir_type/0

file_format()

Specs

t:mod_muc_log.file_format/0

jid_nick()

Specs

t:mod_muc_log.jid_nick/0

jid_nick_role()

Specs

t:mod_muc_log.jid_nick_role/0

logstate()

Specs

t:mod_muc_log.logstate/0

room()

Specs

t:mod_muc_log.room/0

Functions

add_message_to_log/5

Specs

mod_muc_log.add_message_to_log/5

add_to_log/5

Specs

mod_muc_log.add_to_log/5

add_to_log2/5

Specs

mod_muc_log.add_to_log2/5

build_filename_string/6

Specs

mod_muc_log.build_filename_string/6

calc_hour_offset/1

Specs

mod_muc_log.calc_hour_offset/1

check_access_log/2

Specs

mod_muc_log.check_access_log/2

close_previous_log/3

Specs

mod_muc_log.close_previous_log/3

Try to close the previous day log, if it exists

code_change/3

Callback implementation for :gen_server.code_change/3.

create_image_files/1

Specs

mod_muc_log.create_image_files/1

fw/2

Specs

mod_muc_log.fw/2

fw/3

Specs

mod_muc_log.fw/3

get_dateweek/2

Specs

mod_muc_log.get_dateweek/2

get_proc_name/1

get_room_existence_string/2

Specs

mod_muc_log.get_room_existence_string/2

get_room_info/2

get_room_name/1

Specs

mod_muc_log.get_room_name/1

get_roomconfig_text/1

Specs

mod_muc_log.get_roomconfig_text/1

get_timestamp_daydiff/2

Specs

mod_muc_log.get_timestamp_daydiff/2

calculate day before

group_by_role/1

Specs

mod_muc_log.group_by_role/1

Users = [{JID, Nick, Role}]

handle_call/3

Specs

mod_muc_log.handle_call/3

Callback implementation for :gen_server.handle_call/3.

handle_cast/2

Specs

mod_muc_log.handle_cast/2

mod_muc_log.handle_cast/2

Callback implementation for :gen_server.handle_cast/2.

handle_info/2

Callback implementation for :gen_server.handle_info/2.

htmlize/1

htmlize The default behaviour is to ignore the nofollow spam prevention on links (NoFollow=false)

htmlize/2

htmlize/3

The NoFollow parameter tell if the spam prevention should be applied to the link found. true means 'apply nofollow on links'.

htmlize2/2

image_base64/1

init/1

Specs

mod_muc_log.init/1

Callback implementation for :gen_server.init/1.

link_regexp/1

Regexp link. Add the nofollow rel attribute when required

make_dir_rec/1

Specs

mod_muc_log.make_dir_rec/1

put_header/11

Specs

mod_muc_log.put_header/11

put_header_css/2

Specs

mod_muc_log.put_header_css/2

put_header_script/1

put_room_config/4

Specs

mod_muc_log.put_room_config/4

put_room_occupants/4

Specs

mod_muc_log.put_room_occupants/4

render_config_value/4

Specs

mod_muc_log.render_config_value/4

role_users_to_string/2

Specs

mod_muc_log.role_users_to_string/2

roomconfig_to_binary/3

Specs

mod_muc_log.roomconfig_to_binary/3

roomoccupants_to_binary/2

Specs

mod_muc_log.roomoccupants_to_binary/2

Users = [{JID, Nick, Role}]

set_room_occupants/4

Specs

mod_muc_log.set_room_occupants/4

start/2

Specs

mod_muc_log.start/2

Callback implementation for c::gen_mod.start/2.

start_link/2

Specs

mod_muc_log.start_link/2

Starts the server

stop/1

Specs

mod_muc_log.stop/1

Callback implementation for c::gen_mod.stop/1.

terminate/2

Callback implementation for :gen_server.terminate/2.

write_last_lines/3

Specs

mod_muc_log.write_last_lines/3