mod_muc_room (MongooseIM v4.0.1)

Types

activity()

Specs

t:mod_muc_room.activity/0

affiliations_map()

Specs

t:mod_muc_room.affiliations_map/0

config()

Specs

t:mod_muc_room.config/0

find_changed_items_res()

Specs

t:mod_muc_room.find_changed_items_res/0

fsm_return()

Specs

t:mod_muc_room.fsm_return/0

info_msg()

Specs

t:mod_muc_room.info_msg/0

lqueue()

Specs

t:mod_muc_room.lqueue/0

mod_muc_room_user()

Specs

t:mod_muc_room.mod_muc_room_user/0

new_user_strategy()

Specs

t:mod_muc_room.new_user_strategy/0

res_row()

Specs

t:mod_muc_room.res_row/0

routed_iq()

Specs

t:mod_muc_room.routed_iq/0

routed_message()

Specs

t:mod_muc_room.routed_message/0

routed_nick_iq()

Specs

t:mod_muc_room.routed_nick_iq/0

routed_nick_message()

Specs

t:mod_muc_room.routed_nick_message/0

sessions_map()

Specs

t:mod_muc_room.sessions_map/0

stanzaid()

Specs

t:mod_muc_room.stanzaid/0

state()

Specs

t:mod_muc_room.state/0

statename()

Specs

t:mod_muc_room.statename/0

update_inbox_for_muc_payload()

Specs

t:mod_muc_room.update_inbox_for_muc_payload/0

user()

Specs

t:mod_muc_room.user/0

users_map()

Specs

t:mod_muc_room.users_map/0

users_pairs()

Specs

t:mod_muc_room.users_pairs/0

Functions

access_admin/1

Specs

mod_muc_room.access_admin/1

access_persistent/1

Specs

mod_muc_room.access_persistent/1

add_message_to_history/4

Specs

mod_muc_room.add_message_to_history/4

add_new_user/4

Specs

mod_muc_room.add_new_user/4

add_online_user/4

Specs

mod_muc_room.add_online_user/4

add_to_log/3

Specs

mod_muc_room.add_to_log/3

add_user_presence/3

Specs

mod_muc_room.add_user_presence/3

add_user_presence_un/3

Specs

mod_muc_room.add_user_presence_un/3

affiliation_to_binary/1

Specs

mod_muc_room.affiliation_to_binary/1

binary_to_affiliation/1

Specs

mod_muc_room.binary_to_affiliation/1

binary_to_role/1

Specs

mod_muc_room.binary_to_role/1

boolxfield/4

Specs

mod_muc_room.boolxfield/4

broadcast_room_packet/5

calc_shift/2

Specs

mod_muc_room.calc_shift/2

calc_shift/4

Specs

mod_muc_room.calc_shift/4

can_access_identity/2

Specs

mod_muc_room.can_access_identity/2

Return true if UserJID can read real user JIDs

can_access_room/2

Specs

mod_muc_room.can_access_room/2

Return true if UserJID can read room messages

can_change_ra/7

Specs

mod_muc_room.can_change_ra/7

can_change_subject/2

Specs

mod_muc_room.can_change_subject/2

can_read_conference/2

can_send_broadcasts/2

can_send_to_conference/2

can_user_access_identity/2

change_config/2

Specs

mod_muc_room.change_config/2

change_nick/3

Specs

mod_muc_room.change_nick/3

change_stanzaid/2

Specs

mod_muc_room.change_stanzaid/2

change_stanzaid/3

change_subject_error/5

change_subject_if_allowed/4

check_and_strip_visitor_status/3

Specs

mod_muc_room.check_and_strip_visitor_status/3

check_changed_item/9

Specs

mod_muc_room.check_changed_item/9

check_decline_invitation/1

Specs

mod_muc_room.check_decline_invitation/1

Check in the packet is a decline. If so, also returns the splitted packet. This function must be catched, because it crashes when the packet is not a decline message.

check_error_kick/1

Specs

mod_muc_room.check_error_kick/1

Check if the elements of this error stanza indicate that the sender is a dead participant. If so, return true to kick the participant.

check_invitation/4

Specs

mod_muc_room.check_invitation/4

check_password/2

check_subject/1

Specs

mod_muc_room.check_subject/1

check_voice_approval/4

Specs

mod_muc_room.check_voice_approval/4

choose_new_user_password_strategy/3

Specs

mod_muc_room.choose_new_user_password_strategy/3

choose_new_user_strategy/6

Specs

mod_muc_room.choose_new_user_strategy/6

choose_nick_change_strategy/3

Specs

mod_muc_room.choose_nick_change_strategy/3

clean_treap/2

Specs

mod_muc_room.clean_treap/2

code_change/4

config_opt_to_feature/3

Specs

mod_muc_room.config_opt_to_feature/3

count_maxchars_shift/3

Specs

mod_muc_room.count_maxchars_shift/3

count_maxstanzas_shift/2

Specs

mod_muc_room.count_maxstanzas_shift/2

count_seconds_shift/2

Specs

mod_muc_room.count_seconds_shift/2

count_stanza_shift/3

Specs

mod_muc_room.count_stanza_shift/3

count_users/1

Specs

mod_muc_room.count_users/1

create_invite/4

Specs

mod_muc_room.create_invite/4

create_invite_message_elem/4

Specs

mod_muc_room.create_invite_message_elem/4

create_password_elem/1

Specs

mod_muc_room.create_password_elem/1

decide_fate_message/4

Specs

mod_muc_room.decide_fate_message/4

Decide the fate of the message and its sender Returns: continue_delivery | forget_message | {expulse_sender, Reason}

decode_destination_jid/1

Specs

mod_muc_room.decode_destination_jid/1

decode_http_auth_response/1

decode_json_auth_response/1

decode_reason/1

Specs

mod_muc_room.decode_reason/1

destroy_room/2

Specs

mod_muc_room.destroy_room/2

destroy_temporary_room_if_empty/2

Specs

mod_muc_room.destroy_temporary_room_if_empty/2

disco_item/2

Specs

mod_muc_room.disco_item/2

do_add_new_user/5

do_route_iq/4

Specs

mod_muc_room.do_route_iq/4

do_stop_persistent_room/2

element_size/1

erase_matched_users/2

Specs

mod_muc_room.erase_matched_users/2

erase_matched_users_map/3

Specs

mod_muc_room.erase_matched_users_map/3

expulse_participant/4

Specs

mod_muc_room.expulse_participant/4

extract_history/2

Specs

mod_muc_room.extract_history/2

extract_password/1

Specs

mod_muc_room.extract_password/1

extract_role_or_affiliation/1

Specs

mod_muc_room.extract_role_or_affiliation/1

feature/1

Specs

mod_muc_room.feature/1

filter_presence/1

Specs

mod_muc_room.filter_presence/1

find_changed_items/7

Specs

mod_muc_room.find_changed_items/7

find_invite_elems/1

Specs

mod_muc_room.find_invite_elems/1

find_jids_by_nick/2

Specs

mod_muc_room.find_jids_by_nick/2

foreach_matched_jid/3

Specs

mod_muc_room.foreach_matched_jid/3

foreach_matched_user/3

Specs

mod_muc_room.foreach_matched_user/3

foreach_user/2

Specs

mod_muc_room.foreach_user/2

get_affected_jid/3

Specs

mod_muc_room.get_affected_jid/3

get_affiliation/2

Specs

mod_muc_room.get_affiliation/2

get_config/3

Specs

mod_muc_room.get_config/3

get_default_role/2

Specs

mod_muc_room.get_default_role/2

get_default_room_maxusers/1

Specs

mod_muc_room.get_default_room_maxusers/1

get_error_condition/1

Specs

mod_muc_room.get_error_condition/1

get_error_condition2/1

Specs

mod_muc_room.get_error_condition2/1

get_field/2

Specs

mod_muc_room.get_field/2

get_max_users/1

Specs

mod_muc_room.get_max_users/1

get_max_users_admin_threshold/1

Specs

mod_muc_room.get_max_users_admin_threshold/1

get_mucroom_disco_items/1

Specs

mod_muc_room.get_mucroom_disco_items/1

get_participant_data/2

Specs

mod_muc_room.get_participant_data/2

Get information of this participant, or default values. If the JID is not a participant, return values for a service message.

get_role/2

Specs

mod_muc_room.get_role/2

get_room_users/1

Specs

mod_muc_room.get_room_users/1

get_roomdesc_reply/3

Specs

mod_muc_room.get_roomdesc_reply/3

get_roomdesc_tail/2

Specs

mod_muc_room.get_roomdesc_tail/2

get_service_affiliation/2

Specs

mod_muc_room.get_service_affiliation/2

get_service_max_users/1

Specs

mod_muc_room.get_service_max_users/1

get_title/1

Specs

mod_muc_room.get_title/1

get_user_activity/2

Specs

mod_muc_room.get_user_activity/2

getmemberlist_field/1

Specs

mod_muc_room.getmemberlist_field/1

handle_event/3

handle_http_auth_result/6

handle_info/3

Specs

mod_muc_room.handle_info/3

handle_iq_vcard/5

Specs

mod_muc_room.handle_iq_vcard/5

handle_iq_vcard2/7

Specs

mod_muc_room.handle_iq_vcard2/7

handle_new_user/5

Specs

mod_muc_room.handle_new_user/5

handle_roommessage_from_nonparticipant/4

Specs

mod_muc_room.handle_roommessage_from_nonparticipant/4

Handle a message sent to the room by a non-participant. If it is a decline, send to the inviter. Otherwise, an error message is sent to the sender.

handle_sync_event/4

init/1

Specs

mod_muc_room.init/1

A room is created. Depending on request type (MUC/groupchat 1.0) the next state is determined accordingly (a locked room for MUC or an instant one for groupchat).

initial_state/2

Specs

mod_muc_room.initial_state/2

Receive the room-creating Stanza. Will crash if any other stanza is received in this state.

invite_body_elem/4

Specs

mod_muc_room.invite_body_elem/4

invite_body_text/4

Specs

mod_muc_room.invite_body_text/4

iq_admin_allowed/5

Specs

mod_muc_room.iq_admin_allowed/5

iq_disco_info_extras/2

Specs

mod_muc_room.iq_disco_info_extras/2

is_allowed_log_change/3

Specs

mod_muc_room.is_allowed_log_change/3

is_allowed_nonparticipant/2

Specs

mod_muc_room.is_allowed_nonparticipant/2

Check if this non participant can send message to room.
XEP-0045 v1.23: 7.9 Sending a Message to All Occupants an implementation MAY allow users with certain privileges (e.g., a room owner, room admin, or service-level admin) to send messages to the room even if those users are not occupants.

is_allowed_persistent_change/3

Specs

mod_muc_room.is_allowed_persistent_change/3

is_allowed_room_name_desc_limits/2

Specs

mod_muc_room.is_allowed_room_name_desc_limits/2

Check if the Room Name and Room Description defined in the Data Form are conformant to the configured limits

is_empty_map/1

Specs

mod_muc_room.is_empty_map/1

is_empty_room/1

Specs

mod_muc_room.is_empty_room/1

is_first_session/2

Specs

mod_muc_room.is_first_session/2

is_last_session/2

Specs

mod_muc_room.is_last_session/2

is_new_nick_of_online_user/3

Specs

mod_muc_room.is_new_nick_of_online_user/3

is_next_session_of_occupant/3

is_nick_change_public/2

Specs

mod_muc_room.is_nick_change_public/2

is_nick_exists/2

Specs

mod_muc_room.is_nick_exists/2

is_occupant_or_admin/2

Specs

mod_muc_room.is_occupant_or_admin/2

Check if the user is occupant of the room, or at least is an admin or owner.

is_owner/2

Specs

mod_muc_room.is_owner/2

is_password_required/2

is_password_settings_correct/2

Specs

mod_muc_room.is_password_settings_correct/2

Return false if: <<"the password for a password-protected room is blank">>

is_query_allowed/1

Specs

mod_muc_room.is_query_allowed/1

is_room_anonymous/1

is_room_non_anonymous/1

is_room_owner/2

Specs

mod_muc_room.is_room_owner/2

is_user_limit_reached/3

Specs

mod_muc_room.is_user_limit_reached/3

is_user_moderator/2

is_user_online/2

Specs

mod_muc_room.is_user_online/2

is_user_online_iq/3

Specs

mod_muc_room.is_user_online_iq/3

is_visitor/2

Specs

mod_muc_room.is_visitor/2

items_with_affiliation/2

Specs

mod_muc_room.items_with_affiliation/2

items_with_role/2

Specs

mod_muc_room.items_with_role/2

locked_error/3

Specs

mod_muc_room.locked_error/3

In the locked state StateData contains the same settings it previously held for the normal_state. The fsm awaits either a confirmation or a configuration form from the creator. Responds with error to any other queries.

locked_state/2

Specs

mod_muc_room.locked_state/2

Destroy room / confirm instant room / configure room

locked_state_process_owner_iq/5

Specs

mod_muc_room.locked_state_process_owner_iq/5

lookup_affiliation/2

Specs

mod_muc_room.lookup_affiliation/2

lqueue_cut/2

Specs

mod_muc_room.lqueue_cut/2

lqueue_in/2

Specs

mod_muc_room.lqueue_in/2

If the message queue limit is set to 0, do not store messages. Otherwise, rotate messages in the queue store.

lqueue_new/1

Specs

mod_muc_room.lqueue_new/1

lqueue_to_list/1

Specs

mod_muc_room.lqueue_to_list/1

ls/2

make_http_auth_request/4

make_opts/1

Specs

mod_muc_room.make_opts/1

map_foreach_value/2

Specs

mod_muc_room.map_foreach_value/2

maxusers_field/2

maybe_prepare_room_queue/2

maybe_stop_persistent_room/3

mk_send_nick_change/8

muc_user_item/4

Specs

mod_muc_room.muc_user_item/4

muc_user_x/1

Specs

mod_muc_room.muc_user_x/1

next_normal_state/1

nick_available_presence/5

Specs

mod_muc_room.nick_available_presence/5

nick_unavailable_presence/5

Specs

mod_muc_room.nick_unavailable_presence/5

normal_state/2

Specs

mod_muc_room.normal_state/2

notify_config_change_and_get_type/5

Specs

mod_muc_room.notify_config_change_and_get_type/5

notify_users_modified/1

occupant_jid/2

Specs

mod_muc_room.occupant_jid/2

parse_history_val/2

Specs

mod_muc_room.parse_history_val/2

perform_http_auth/6

prepare_room_queue/1

Specs

mod_muc_room.prepare_room_queue/1

presence/2

Specs

mod_muc_room.presence/2

presence_stanza_of_type_unavailable/1

Specs

mod_muc_room.presence_stanza_of_type_unavailable/1

privatexfield/4

process_admin_item_set/3

process_admin_item_set_unsafe/3

process_admin_items_set/4

Specs

mod_muc_room.process_admin_items_set/4

process_authorized_iq_owner/6

Specs

mod_muc_room.process_authorized_iq_owner/6

process_authorized_submit_owner/3

Specs

mod_muc_room.process_authorized_submit_owner/3

process_groupchat_message/3

Specs

mod_muc_room.process_groupchat_message/3

process_iq_admin/5

Specs

mod_muc_room.process_iq_admin/5

process_iq_disco_info/4

Specs

mod_muc_room.process_iq_disco_info/4

process_iq_disco_items/4

Specs

mod_muc_room.process_iq_disco_items/4

process_iq_owner/6

Specs

mod_muc_room.process_iq_owner/6

process_message_from_allowed_user/3

process_presence/4

Specs

mod_muc_room.process_presence/4

Process presence stanza and destroy the room, if it is empty.

process_presence/5

Specs

mod_muc_room.process_presence/5

process_presence1/4

Specs

mod_muc_room.process_presence1/4

process_presence_error/4

Specs

mod_muc_room.process_presence_error/4

process_presence_nick_change/5

Specs

mod_muc_room.process_presence_nick_change/5

process_presence_unavailable/3

Specs

mod_muc_room.process_presence_unavailable/3

process_simple_presence/3

Specs

mod_muc_room.process_simple_presence/3

read_hibernate_timeout/1

remove_each_occupant_from_room/2

Specs

mod_muc_room.remove_each_occupant_from_room/2

Service Removes Each Occupant

Send only one presence stanza of type "unavailable" to each occupant so that the user knows he or she has been removed from the room.
If extended presence information specifying the JID of an alternate location and the reason for the room destruction was provided by the room owner, the presence stanza MUST include that information.

remove_nonmembers/1

Specs

mod_muc_room.remove_nonmembers/1

remove_online_user/2

Specs

mod_muc_room.remove_online_user/2

remove_online_user/3

Specs

mod_muc_room.remove_online_user/3

replace_subelement/2

Specs

mod_muc_room.replace_subelement/2

Given an element and a new subelement, replace the instance of the subelement in element with the new subelement.

reply_not_authorized/5

reply_service_unavailable/5

reply_with_timeout/3

rewrite_next_state/2

Specs

mod_muc_room.rewrite_next_state/2

rfield/4

Specs

mod_muc_room.rfield/4

rfieldt/3

Specs

mod_muc_room.rfieldt/3

role_to_binary/1

Specs

mod_muc_room.role_to_binary/1

route/5

Specs

mod_muc_room.route/5

route_error/4

Specs

mod_muc_room.route_error/4

route_invitation/5

Specs

mod_muc_room.route_invitation/5

route_iq/3

Specs

mod_muc_room.route_iq/3

route_message/2

Specs

mod_muc_room.route_message/2

route_nick_iq/2

Specs

mod_muc_room.route_nick_iq/2

route_nick_message/2

Specs

mod_muc_room.route_nick_message/2

route_voice_approval/5

Specs

mod_muc_room.route_voice_approval/5

run_join_room_hook/2

Specs

mod_muc_room.run_join_room_hook/2

run_leave_room_hook/2

Specs

mod_muc_room.run_leave_room_hook/2

run_update_inbox_for_muc_hook/2

Specs

mod_muc_room.run_update_inbox_for_muc_hook/2

safe_send_kickban_presence/4

safe_send_kickban_presence/5

save_persistent_room_state/1

schedule_queue_processing_when_empty/4

Specs

mod_muc_room.schedule_queue_processing_when_empty/4

search_affiliation/2

Specs

mod_muc_room.search_affiliation/2

search_role/2

Specs

mod_muc_room.search_role/2

send_config_update/2

Specs

mod_muc_room.send_config_update/2

send_decline_invitation/3

Specs

mod_muc_room.send_decline_invitation/3

Send the decline to the inviter user. The original stanza must be slightly modified.

send_error_only_occupants/5

Specs

mod_muc_room.send_error_only_occupants/5

send_existing_presence/4

Specs

mod_muc_room.send_existing_presence/4

send_existing_presences/2

Specs

mod_muc_room.send_existing_presences/2

send_history/3

Specs

mod_muc_room.send_history/3

send_invitation/4

Specs

mod_muc_room.send_invitation/4

send_kickban_presence/4

Specs

mod_muc_room.send_kickban_presence/4

send_kickban_presence/5

Specs

mod_muc_room.send_kickban_presence/5

send_kickban_presence1/5

Specs

mod_muc_room.send_kickban_presence1/5

send_new_presence/2

Specs

mod_muc_room.send_new_presence/2

send_new_presence/3

Specs

mod_muc_room.send_new_presence/3

send_new_presence_to/4

Specs

mod_muc_room.send_new_presence_to/4

send_new_presence_to_single/7

send_new_presence_un/2

Specs

mod_muc_room.send_new_presence_un/2

send_new_presence_un/3

Specs

mod_muc_room.send_new_presence_un/3

send_nick_change/10

send_nick_changing/3

Specs

mod_muc_room.send_nick_changing/3

send_subject/3

Specs

mod_muc_room.send_subject/3

send_to_all_users/2

Specs

mod_muc_room.send_to_all_users/2

send_to_occupants/2

Specs

mod_muc_room.send_to_occupants/2

send_update_presence/3

Specs

mod_muc_room.send_update_presence/3

set_affiliation/3

Specs

mod_muc_room.set_affiliation/3

set_affiliation_and_reason/4

Specs

mod_muc_room.set_affiliation_and_reason/4

set_config/2

Specs

mod_muc_room.set_config/2

set_opts/2

Specs

mod_muc_room.set_opts/2

set_role/3

Specs

mod_muc_room.set_role/3

set_xoption/2

Specs

mod_muc_room.set_xoption/2

stanzaid_pack/2

Specs

mod_muc_room.stanzaid_pack/2

stanzaid_unpack/1

Specs

mod_muc_room.stanzaid_unpack/1

start/8

Specs

mod_muc_room.start/8

start/10

Specs

mod_muc_room.start/10

start_link/8

start_link/10

status_code/1

Specs

mod_muc_room.status_code/1

stop/1

stop_if_only_owner_is_online/3

store_room_if_persistent/1

Specs

mod_muc_room.store_room_if_persistent/1

store_user_activity/3

Specs

mod_muc_room.store_user_activity/3

stringxfield/4

strip_status/1

Specs

mod_muc_room.strip_status/1

tab_add_online_user/2

Specs

mod_muc_room.tab_add_online_user/2

tab_count_user/1

Specs

mod_muc_room.tab_count_user/1

tab_remove_online_user/2

Specs

mod_muc_room.tab_remove_online_user/2

terminate/3

Specs

mod_muc_room.terminate/3

Purpose: Shutdown the fsm

unavailable_presence/2

unsafe_check_invitation/4

Specs

mod_muc_room.unsafe_check_invitation/4

update_matched_users/3

Specs

mod_muc_room.update_matched_users/3

update_matched_users_map/3

Specs

mod_muc_room.update_matched_users_map/3

uri_encode/1

user_to_item/2

Specs

mod_muc_room.user_to_item/2

which_property_changed/2

Specs

mod_muc_room.which_property_changed/2

whois_field/2

Specs

mod_muc_room.whois_field/2

xfield/5

Specs

mod_muc_room.xfield/5