mod_offline_chatmarkers behaviour (MongooseIM v4.0.1)

This module optimizes offline storage for chat markers in the next way:

1) It filters out chat marker packets processed by mod_smart_markers:

* These packets can be identified by the extra permanent Acc timestamp field added by mod_smart_markers.

* These packets are not going to mod_offline (notice the difference in priorities for the offline_message_hook handlers)

* The information about these chat markers is stored in DB, timestamp added by mod_smart_markers is important here!

2) After all the offline messages are inserted by mod_offline (notice the difference in priorities for the resend_offline_messages_hook handlers), this module adds the latest chat markers as the last offline messages:

* It extracts chat markers data stored for the user in the DB (with timestamps)

* Requests cached chat markers from mod_smart_markers that has timestamp older or equal to the stored one.
* Generates and inserts chat markers as the last offline messages

Functions

add_default_backend/1

build_one2one_chatmarker_msg/2

build_room_chatmarker_msg/3

deps/2

Specs

mod_offline_chatmarkers.deps/2

get_room/2

get_thread/1

hooks/1

inspect_packet/4

make_route_item/5

marker/2

maybe_store_chat_marker/4

offline_chatmarkers/3

pop_offline_messages/3

process_row/3

remove_user/3

start/2

Callback implementation for c::gen_mod.start/2.

stop/1

Callback implementation for c::gen_mod.stop/1.

thread/1

Callbacks

get(jid)

Specs

c:mod_offline_chatmarkers.get/1

init(host, opts)

Specs

c:mod_offline_chatmarkers.init/2

maybe_store(jid, thread, room, timestamp)

Specs

c:mod_offline_chatmarkers.maybe_store/4

remove_user(jid)

Specs

c:mod_offline_chatmarkers.remove_user/1