mod_offline_riak (MongooseIM v4.0.1)

Riak backend for last activity XEP

The backend uses the existing riak connection pool, which is "globally" defined in the ejebberd.cfg file. Therefore, we don't need to start anything in the init function.

The module follows the approach taken by the other riak backends - it creates the following bucket {<<"offline">>, <<"example.com">>} for each xmpp domain.

   Data Layout:
   KV: {Username_and_timestamp:binary, Status:binary}
   2i: [Username:binary, Timestamp:integer]

Functions

bucket_type/1

Specs

mod_offline_riak.bucket_type/1

count_offline_messages/3

Specs

mod_offline_riak.count_offline_messages/3

Callback implementation for c::mod_offline.count_offline_messages/3.

extract_permanent_fields/1

fetch_messages/2

Callback implementation for c::mod_offline.fetch_messages/2.

fetch_msg/4

init/2

Specs

mod_offline_riak.init/2

Callback implementation for c::mod_offline.init/2.

key/2

Specs

mod_offline_riak.key/2

maybe_decode_timestamp/1

maybe_encode_timestamp/1

pop_messages/2

Specs

mod_offline_riak.pop_messages/2

Callback implementation for c::mod_offline.pop_messages/2.

pop_msg/4

read_user_idx/2

remove_expired_messages/1

Specs

mod_offline_riak.remove_expired_messages/1

Callback implementation for c::mod_offline.remove_expired_messages/1.

remove_old_messages/2

Specs

mod_offline_riak.remove_old_messages/2

Callback implementation for c::mod_offline.remove_old_messages/2.

remove_user/2

Specs

mod_offline_riak.remove_user/2

Callback implementation for c::mod_offline.remove_user/2.

set_obj_user_metadata/2

write_messages/3

Specs

mod_offline_riak.write_messages/3

Callback implementation for c::mod_offline.write_messages/3.

write_msg/3