mod_pubsub (MongooseIM v4.0.1)

The module mod_pubsub is the core of the PubSub extension. It relies on PubSub plugins for a large part of its functions.

Types

accessModel()

Specs

t:mod_pubsub.accessModel/0

affiliation()

Specs

t:mod_pubsub.affiliation/0

host()

Specs

t:mod_pubsub.host/0

hostPEP()

Specs

t:mod_pubsub.hostPEP/0

For example, it can be :
     {"bob", "example.org", []}
.

hostPubsub()

Specs

t:mod_pubsub.hostPubsub/0

hostPubsub()

Specs

t:mod_pubsub.hostPubsub/0

itemId()

Specs

t:mod_pubsub.itemId/0

nodeId()

Specs

t:mod_pubsub.nodeId/0

A node is defined by a list of its ancestors. The last element is the name of the current node. For example:
     <<"/home/localhost/user">>

nodeIdx()

Specs

t:mod_pubsub.nodeIdx/0

nodeOption()

Specs

t:mod_pubsub.nodeOption/0

Example:
     {deliver_payloads, true}

nodeOptions()

Specs

t:mod_pubsub.nodeOptions/0

nodeOptions()

Specs

t:mod_pubsub.nodeOptions/0

payload()

Specs

t:mod_pubsub.payload/0

publishModel()

Specs

t:mod_pubsub.publishModel/0

publishOptions()

Specs

t:mod_pubsub.publishOptions/0

pubsubItem()

Specs

t:mod_pubsub.pubsubItem/0

pubsubLastItem()

Specs

t:mod_pubsub.pubsubLastItem/0

pubsubNode()

Specs

t:mod_pubsub.pubsubNode/0

pubsubState()

Specs

t:mod_pubsub.pubsubState/0

state()

Specs

t:mod_pubsub.state/0

subId()

Specs

t:mod_pubsub.subId/0

subOption()

Specs

t:mod_pubsub.subOption/0

subOption()

Specs

t:mod_pubsub.subOption/0

subOptions()

Specs

t:mod_pubsub.subOptions/0

subOptions()

Specs

t:mod_pubsub.subOptions/0

subscription()

Specs

t:mod_pubsub.subscription/0

Functions

add_extended_headers/2

add_headers/4

add_hooks/2

add_message_type/2

Specs

mod_pubsub.add_message_type/2

add_opt/3

add_pep_iq_handlers/2

adhoc_get_pending_parse_options/2

adhoc_request/6

Processes an Ad Hoc Command.

affiliation_to_string/1

all_metrics/0

autocreate_if_supported_and_publish/8

broadcast_auto_retract_notification/7

broadcast_config_notification/6

broadcast_created_node/6

broadcast_publish_item/10

broadcast_purge_node/5

broadcast_removed_node/6

broadcast_retract_items/6

broadcast_retract_items/7

broadcast_stanza/9

broadcast_stanza/10

broadcast_step/2

cache_backend/1

caps_recognised/4

check_plugin_features_and_acc_affs/4

check_subs_limit/3

collection_shim/1

command_disco_info/3

config/2

config/3

config_metrics/1

Callback implementation for c::mongoose_module_metrics.config_metrics/1.

count_subscribed/1

create_node/5

Create new pubsub nodes

In addition to method-specific error conditions, there are several general reasons why the node creation request might fail:

The service does not support node creation.

Only entities that are registered with the service are allowed to create nodes but the requesting entity is not registered.

The requesting entity does not have sufficient privileges to create nodes.

The requested Node already exists.

The request did not include a Node and "instant nodes" are not supported.

ote: node creation is a particular case, error return code is evaluated at many places:

iq_pubsub checks if service supports node creation (type exists)

create_node checks if instant nodes are supported

create_node asks node plugin if entity have sufficient privilege

nodetree create_node checks if nodeid already exists

node plugin create_node just sets default affiliation/subscription

create_node/7

Specs

mod_pubsub.create_node/7

create_node_authorized_transaction/6

create_node_make_reply/1

create_node_reply/2

create_node_transaction/7

db_call_fun/3

default_host/0

Specs

mod_pubsub.default_host/0

delete_hooks/2

delete_item/4

Specs

mod_pubsub.delete_item/4

Delete item from a PubSub node.

The permission to delete an item must be verified by the plugin implementation.

There are several reasons why the item retraction request might fail:

The publisher does not have sufficient privileges to delete the requested item.

The node or item does not exist.

The request does not specify a node.

The request does not include an element or the element does not specify an ItemId.

The node does not support persistent items.

The service does not support the deletion of items.

delete_item/5

delete_item_transaction/4

delete_node/3

Specs

mod_pubsub.delete_node/3

Delete specified node and all childs.

There are several reasons why the node deletion request might fail:

The requesting entity does not have sufficient privileges to delete the node.

The node is the root collection node, which cannot be deleted.

The specified node does not exist.

delete_node_transaction/4

delete_pep_iq_handlers/1

deps/2

dirty/4

disco_features/3

disco_identity/3

disco_items/3

Specs

mod_pubsub.disco_items/3

disco_local_features/5

Specs

mod_pubsub.disco_local_features/5

disco_local_identity/4

disco_local_identity/5

Specs

mod_pubsub.disco_local_identity/5

disco_sm_features/5

Specs

mod_pubsub.disco_sm_features/5

disco_sm_identity/5

Specs

mod_pubsub.disco_sm_identity/5

disco_sm_items/5

Specs

mod_pubsub.disco_sm_items/5

dispatch_items/5

do_route/7

Specs

mod_pubsub.do_route/7

ensure_metrics/1

event_stanza/1

event_stanza/2

event_stanza_with_els/2

extended_error/2

extended_error/3

extended_headers/1

extract_item_id/2

extract_item_ids/1

feature/1

features/0

features/2

filter_node_options/1

find_authorization_response/1

get_affiliations/3

Specs

mod_pubsub.get_affiliations/3

get_affiliations/4

Specs

mod_pubsub.get_affiliations/4

Return the list of affiliations as an XMPP response.

get_affiliations_transaction/3

get_allowed_items_call/6

get_allowed_items_call/7

get_cached_item/2

Specs

mod_pubsub.get_cached_item/2

get_collection_subscriptions/2

get_configure/4

get_configure_transaction/6

get_configure_xfields/4

get_contacts_for_sending_last_item/2

get_default/4

get_item/3

get_items/2

get_items/7

Specs

mod_pubsub.get_items/7

Return the items of a given node.

The number of items to return is limited by MaxItems.

The permission are not checked in this function.

get_items_transaction/7

get_items_with_limit/7

get_last_item/4

get_last_items/5

get_max_items_node/1

get_max_subscriptions_node/1

get_node_subs/2

get_node_subs_by_depth/3

get_node_subscriptions_transaction/3

get_nodes_for_sending_last_item/2

get_nodes_owned_by/2

get_option/2

get_option/3

get_options/5

get_options_transaction/6

get_parent/3

get_pending_nodes/3

get_personal_data/2

Specs

mod_pubsub.get_personal_data/2

get_presence_and_roster_permissions/5

Specs

mod_pubsub.get_presence_and_roster_permissions/5

get_resource_state/3

Specs

mod_pubsub.get_resource_state/3

get_roster_info/4

get_sub_options_xml/7

get_subscriptions/3

get_subscriptions/4

Return the list of subscriptions as an XMPP response.

get_subscriptions_for_send_last/3

get_subscriptions_transaction/3

handle_authorization_response/5

handle_authorization_response_transaction/6

handle_call/3

Callback implementation for :gen_server.handle_call/3.

handle_pep_authorization_response/1

handle_pep_authorization_response/6

handle_remote_hook/4

hooks/0

host/1

in_subscription/6

Specs

mod_pubsub.in_subscription/6

init/1

Specs

mod_pubsub.init/1

Callback implementation for :gen_server.init/1.

init_backend/2

init_plugin/5

init_plugins/3

Call the init/1 function for each plugin declared in the config file. The default plugin module is implicit.
The Erlang code for the plugin is located in a module called node_plugin. The 'node_' prefix is mandatory.

The modules are initialized in alphetical order and the list is checked and sorted to ensure that each module is initialized only once.

See node_hometree:init/1 for an example implementation.

init_send_loop/1

iq_action_to_metric_name/1

iq_command/6

iq_disco_info/4

iq_disco_items/4

Specs

mod_pubsub.iq_disco_items/4

iq_disco_items_transaction/5

iq_get_vcard/1

iq_pubsub/6

Specs

mod_pubsub.iq_pubsub/6

iq_pubsub/8

Specs

mod_pubsub.iq_pubsub/8

iq_pubsub_action/6

iq_pubsub_get_items/4

iq_pubsub_get_options/4

iq_pubsub_owner/6

Specs

mod_pubsub.iq_pubsub_owner/6

iq_pubsub_owner_action/6

iq_pubsub_set_create/4

iq_pubsub_set_options/3

iq_pubsub_set_publish/4

iq_pubsub_set_retract/4

iq_pubsub_set_subscribe/4

iq_pubsub_set_unsubscribe/4

iq_sm/4

Specs

mod_pubsub.iq_sm/4

is_last_item_cache_enabled/1

is_subscribed/3

is_to_deliver/5

Check if a notification must be delivered or not based on node and subscription options.

item_attr/1

item_attr/2

items_els/1

items_event_stanza/2

make_and_wrap_sub_xform/4

make_error_reply/2

match_option/3

max_items/2

Return the maximum number of items for a given node.

Unlimited means that there is no limit in the number of items that can be stored.

maybe_add_shim_headers/5

maybe_default_node/3

maybe_start_cache_module/2

metric_name/3

node_action/4

node_attr/1

node_call/4

node plugin call.

node_disco_info/3

node_disco_info/5

node_options/2

node_owners_action/4

node_owners_call/4

node_plugin_options/2

node_to_deliver/2

nodes_to_deliver/5

notify_send_loop/2

notify_subscription_change/4

on_user_offline/5

out_subscription/5

Specs

mod_pubsub.out_subscription/5

parse_create_node_options_if_possible/3

payload_by_option/3

payload_xmlelements/1

Specs

mod_pubsub.payload_xmlelements/1

payload_xmlelements/2

pep_hooks/0

plugin/1

plugin/2

plugin_call/3

plugin_features/2

plugins/1

presence_can_deliver/2

Specs

mod_pubsub.presence_can_deliver/2

presence_probe/4

process_jid_to_deliver/5

process_jids_to_deliver/4

process_packet/5

Specs

mod_pubsub.process_packet/5

Callback implementation for c::mongoose_packet_handler.process_packet/5.

publish_item/6

Specs

mod_pubsub.publish_item/6

Publish item to a PubSub node.

The permission to publish an item must be verified by the plugin implementation.

There are several reasons why the publish request might fail:

The requesting entity does not have sufficient privileges to publish.

The node does not support item publication.

The node does not exist.

The payload size exceeds a service-defined limit.

The item contains more than one payload element or the namespace of the root payload element does not match the configured namespace for the node.

The request does not match the node configuration.

publish_item/7

publish_item/8

purge_item_of_offline_user/5

purge_node/3

Specs

mod_pubsub.purge_node/3

Delete all items of specified node owned by JID.

There are several reasons why the node purge request might fail:

The node or service does not support node purging.

The requesting entity does not have sufficient privileges to purge the node.

The node is not configured to persist items.

The specified node does not exist.

purge_node_transaction/3

purge_offline/1

purge_offline/3

remove_user/3

remove_user_per_plugin_safe/3

report_iq_action_metrics_after_return/5

report_iq_action_metrics_before_result/3

select_type/2

select_type/3

select_type/4

send_authorization_approval/4

Send a message to JID with the supplied Subscription

send_authorization_request/2

send_items/7

Resend the items of a node to the user.

send_last_item_to_jid/3

send_last_items_from_owner/3

send_last_pep_items/3

send_last_pubsub_items/2

send_last_pubsub_items_for_plugin/3

send_loop/1

send_pending_auth_events/4

Send a subscription approval form to Owner for all pending subscriptions on Host and Node.

send_pending_node_form/4

Sends the process pending subscriptions XForm for Host to Owner.

serverhost/1

service_jid/1

Specs

mod_pubsub.service_jid/1

set_affiliations/4

Specs

mod_pubsub.set_affiliations/4

set_affiliations_transaction/4

set_cached_item/5

set_configure/4

set_configure_submit/5

set_configure_transaction/4

set_configure_valid_transaction/3

set_options/5

set_options_helper/6

set_options_transaction/5

set_subscription_transaction/6

set_subscriptions/4

set_subscriptions_transaction/5

set_validated_affiliations_transaction/4

set_xoption/3

start/2

Callback implementation for c::gen_mod.start/2.

start_cache_module/2

start_link/2

state_can_deliver/2

Specs

mod_pubsub.state_can_deliver/2

stop/1

Callback implementation for c::gen_mod.stop/1.

store_config_in_ets/6

string_allow_to_boolean/1

string_to_affiliation/1

string_to_ljid/1

string_to_subscription/1

sub_option_can_deliver/3

sub_to_deliver/4

subid_shim/1

subscribe_node/5

Specs

mod_pubsub.subscribe_node/5

Accepts or rejects subcription requests on a PubSub node.

There are several reasons why the subscription request might fail:

The bare JID portions of the JIDs do not match.

The node has an access model of "presence" and the requesting entity is not subscribed to the owner's presence.

The node has an access model of "roster" and the requesting entity is not in one of the authorized roster groups.

The node has an access model of "whitelist" and the requesting entity is not on the whitelist.

The service requires payment for subscriptions to the node.

The requesting entity is anonymous and the service does not allow anonymous entities to subscribe.

The requesting entity has a pending subscription.

The requesting entity is blocked from subscribing (e.g., because having an affiliation of outcast).

The node does not support subscriptions.

The node does not exist.

See also node_hometree:subscribe_node/5.

subscribe_node_reply/2

subscribe_node_reply/3

subscribe_node_transaction/5

subscribe_node_transaction_step1/6

subscribe_node_transaction_step2/6

subscribe_node_transaction_step3/6

subscribe_node_transaction_step4/5

subscribed_nodes_by_jid/2

subscription_to_string/1

subscription_to_xmlel/2

terminate_plugins/4

timestamp/0

transaction/4

tree/1

tree/2

tree_action/3

tree_call/3

node tree plugin call.

uniqid/0

Specs

mod_pubsub.uniqid/0

unset_cached_item/2

unsubscribe_node/5

Specs

mod_pubsub.unsubscribe_node/5

Unsubscribe JID from the Node.

There are several reasons why the unsubscribe request might fail:

The requesting entity has multiple subscriptions to the node but does not specify a subscription ID.

The request does not specify an existing subscriber.

The requesting entity does not have sufficient privileges to unsubscribe the specified JID.

The node does not exist.

The request specifies a subscription ID that is not valid or current.

unsubscribe_user/2

unsubscribe_user/3

unsubscribe_user_per_plugin/4

update_auth/7

user_resource/3

user_resources/2

validate_and_set_options_helper/6