mod_roster behaviour (MongooseIM v4.0.1)

Roster management (Mnesia storage).
Includes support for XEP-0237: Roster Versioning. The roster versioning follows an all-or-nothing strategy: - If the version supplied by the client is the latest, return an empty response. - If not, return the entire new roster (with updated version string). Roster version is a hash digest of the entire roster. No additional data is stored in DB.

Types

get_user_roster_strategy()

Specs

t:mod_roster.get_user_roster_strategy/0

roster()

Specs

t:mod_roster.roster/0

sub_presence()

Specs

t:mod_roster.sub_presence/0

subscription_state()

Specs

t:mod_roster.subscription_state/0

Functions

ask_to_pending/1

autoreply_to_type/1

build_pending/3

choose_get_user_roster_strategy/3

Specs

mod_roster.choose_get_user_roster_strategy/3

config_metrics/1

Callback implementation for c::mongoose_module_metrics.config_metrics/1.

create_sub_el/2

del_roster_t/3

do_process_iq_get/3

Specs

mod_roster.do_process_iq_get/3

do_process_item_set/4

Specs

mod_roster.do_process_item_set/4

fill_subscription_lists/6

get_jid_info/4

Specs

mod_roster.get_jid_info/4

get_personal_data/2

Specs

mod_roster.get_personal_data/2

get_roster/2

get_roster_entry/2

get_roster_entry/3

get_roster_entry_t/3

get_roster_old/2

get_roster_old/3

get_subscription_lists/3

Specs

mod_roster.get_subscription_lists/3

get_user_roster/2

Specs

mod_roster.get_user_roster/2

get_user_roster_based_on_version/4

get_user_roster_db_versioning/3

get_user_roster_hash_versioning/3

get_user_roster_no_versioning/2

get_user_rosters_length/2

get_versioning_feature/2

hooks/1

in_auto_reply/3

in_state_change/3

in_subscription/6

Specs

mod_roster.in_subscription/6

item_to_map/1

Specs

mod_roster.item_to_map/1

item_to_xml/1

jid_arg_to_lower/1

Specs

mod_roster.jid_arg_to_lower/1

modify_roster_item/4

out_state_change/3

out_subscription/5

Specs

mod_roster.out_subscription/5

process_iq/4

Specs

mod_roster.process_iq/4

process_iq_get/3

Specs

mod_roster.process_iq_get/3

process_iq_set/3

Specs

mod_roster.process_iq_set/3

process_item_attrs/2

process_item_attrs_ws/2

process_item_els/2

process_item_set/3

Specs

mod_roster.process_item_set/3

process_item_set_t/3

process_local_iq/4

Specs

mod_roster.process_local_iq/4

process_subscription/6

process_subscription_transaction/5

push_item/3

push_item_final/4

push_item_version/4

push_item_without_version/4

read_roster_version/2

remove_from_roster/2

Specs

mod_roster.remove_from_roster/2

remove from roster - in practice it means changing subscription state to 'remove'

remove_test_user/2

remove_user/2

remove_user/3

roster_hash/1

roster_record_to_gdpr_entry/1

roster_subscribe_t/4

roster_version/2

roster_version_on_db/1

Specs

mod_roster.roster_version_on_db/1

roster_versioning_enabled/1

Specs

mod_roster.roster_versioning_enabled/1

send_presence_type/3

send_unsubscribing_presence/2

send_unsubscription_to_rosteritems/2

set_items/2

set_roster_entry/4

Specs

mod_roster.set_roster_entry/4

add a contact to roster, or update

set_roster_entry/5

Specs

mod_roster.set_roster_entry/5

set_roster_item/4

Specs

mod_roster.set_roster_item/4

this is run when a roster item is to be added, updated or removed the interface of this func could probably be a bit simpler

start/2

Callback implementation for c::gen_mod.start/2.

stop/1

Callback implementation for c::gen_mod.stop/1.

transaction/2

try_send_unsubscription_to_rosteritems/2

update_roster_t/4

write_roster_version/2

write_roster_version/3

write_roster_version_t/2

Callbacks

del_roster_t(lUser, lServer, lJid)

Specs

c:mod_roster.del_roster_t/3

get_roster(lUser, lServer)

Specs

c:mod_roster.get_roster/2

get_roster_entry(lUser, lServer, jid)

Specs

c:mod_roster.get_roster_entry/3

get_roster_entry(lUser, lServer, jid, atom)

Specs

c:mod_roster.get_roster_entry/4

get_roster_entry_t(lUser, lServer, jid)

Specs

c:mod_roster.get_roster_entry_t/3

get_roster_entry_t(lUser, lServer, jid, atom)

Specs

c:mod_roster.get_roster_entry_t/4

get_subscription_lists(acc, lUser, lServer)

Specs

c:mod_roster.get_subscription_lists/3

init(host, opts)

Specs

c:mod_roster.init/2

raw_to_record(lServer, item)

Specs

c:mod_roster.raw_to_record/2

read_roster_version(lUser, lServer)

Specs

c:mod_roster.read_roster_version/2

remove_user(lUser, lServer)

Specs

c:mod_roster.remove_user/2

roster_subscribe_t(lUser, lServer, lJid, sJid)

Specs

c:mod_roster.roster_subscribe_t/4

transaction(lServer, f)

Specs

c:mod_roster.transaction/2

update_roster_t(lUser, lServer, lJid, item)

Specs

c:mod_roster.update_roster_t/4

write_roster_version(lUser, lServer, inTransaction, ver)

Specs

c:mod_roster.write_roster_version/4