mod_smart_markers behaviour (MongooseIM v4.0.1)

This module implements storage of the latest chat markers sent by the users. This can be used to optimize mod_offline functionality, or to implement custom fetching protocol and avoid storage of chat markers in MAM.

Please be aware of the next implementation details:

1) Current implementation is based on user_send_packet hook. It doesn't work for s2s connections, but usage of another hook (e.g. filter_local_packet) makes implementation harder and results in multiple processing of one and the same chat marker notification (sent to different users by MUC). However that is the only possible way to deal with group chat messages sent from the room to the user over s2s.

                                              S2S
                                               +
                                               |
                  +--------------------+       |
                  |                    |       |   filter
                  |                    +--------------->
         send     |                    |       |   filter
         +------->+       ROOM         +--------------->
                  |                    |       |   filter
                  |                    +--------------->
                  |                    |       |
                  +--------------------+       |
                                               |
                                               +

2) DB backend requires us to provide host information, and the host is always the recipient's server in case one2one messages, and a master domain of the MUC service in case of groupchat.

3) It is the client application's responsibility to ensure that chat markers move only forward. There is no verification of chat markers in this module, it just stores the latest chat marker information sent by the user.
4) MUC light doesn't have message serialization! So it doesn't guaranty one and the same message order for different users. This can result in a race condition situation when different users track (and mark) different messages as the last in a chat history. However, this is a rare situation, and it self recovers on the next message in the room. Anyway storing chat markers in MAM doesn't fix this problem.

Types

chat_marker()

Specs

t:mod_smart_markers.chat_marker/0

chat_marker_type()

Specs

t:mod_smart_markers.chat_marker_type/0

chat_type()

Specs

t:mod_smart_markers.chat_type/0

maybe_thread()

Specs

t:mod_smart_markers.maybe_thread/0

Functions

add_default_backend/1

can_access_room/2

Specs

mod_smart_markers.can_access_room/2

check_chat_marker_attributes/2

Specs

mod_smart_markers.check_chat_marker_attributes/2

extract_chat_markers/4

Specs

mod_smart_markers.extract_chat_markers/4

get_chat_markers/1

Specs

mod_smart_markers.get_chat_markers/1

get_chat_markers/4

Specs

mod_smart_markers.get_chat_markers/4

get_host/2

Specs

mod_smart_markers.get_host/2

get_thread/1

Specs

mod_smart_markers.get_thread/1

hooks/1

Specs

mod_smart_markers.hooks/1

is_chat_marker_element/1

Specs

mod_smart_markers.is_chat_marker_element/1

is_valid_message/3

Specs

mod_smart_markers.is_valid_message/3

maybe_update_chat_markers/5

Specs

mod_smart_markers.maybe_update_chat_markers/5

start/2

Specs

mod_smart_markers.start/2

Callback implementation for c::gen_mod.start/2.

stop/1

Specs

mod_smart_markers.stop/1

Callback implementation for c::gen_mod.stop/1.

user_send_packet/4

Specs

mod_smart_markers.user_send_packet/4

Callbacks

get_chat_markers(host, to, thread, timestamp)

Specs

c:mod_smart_markers.get_chat_markers/4

init(host, opts)

Specs

c:mod_smart_markers.init/2

update_chat_marker(host, chatMarker)

Specs

c:mod_smart_markers.update_chat_marker/2