mongoose_api_common (MongooseIM v4.0.1)

MongooseIM REST API backend

This module handles the client HTTP REST requests, then respectively convert them to Commands from mongoose_commands and execute with admin` privileges. It supports responses with appropriate HTTP Status codes returned to the client. This module implements behaviour introduced in ejabberd_cowboy which is %% built on top of the cowboy library. The method supported: GET, POST, PUT, DELETE. Only JSON format. The library "jiffy" used to serialize and deserialized JSON data. REQUESTS The module is based on mongoose_commands registry. The root http path for a command is build based on the "category" field. %% Its always used as path a prefix. The commands are translated to HTTP API in the following manner:

command of action "read" will be called by GET request command of action "create" will be called by POST request command of action "update" will be called by PUT request command of action "delete" will be called by DELETE request

The args of the command will be filled with the values provided in path %% bindings or body parameters, depending of the method type: - for command of action "read" or "delete" all the args are pulled from the path bindings. The path should be constructed of pairs "/arg_name/arg_value" so that it could match the {arg_name, type} %% pattern in the command registry. E.g having the record of category "users" and args: [{username, binary}, {domain, binary}] we will have to make following GET request %% path: http://domain:port/api/users/username/Joe/domain/localhost and the command will be called with arguments "Joe" and "localhost"

- for command of action "create" or "update" args are pulled from the body JSON, except those that are on the "identifiers" list of the command. Those go to the path bindings. E.g having the record of category "animals", action "update" and args: [{species, binary}, {name, binary}, {age, integer}] and identifiers: [species, name] we can set the age for our elephant Ed in the PUT request: path: http://domain:port/api/species/elephant/name/Ed body: {"age" : "10"} and then the command will be called with arguments ["elephant", "Ed" and 10].

RESPONSES
The API supports some of the http status code like 200, 201, 400, 404 etc depending on the return value of the command execution and arguments checks. Additionally, when the command's action is "create" and it returns a value, it is concatenated to the path and return to the client in header "location" with response code 201 so that it could represent now a new created resource. If error occured while executing the command, the appropriate reason is returned in response body.

Types

arg_name()

Specs

t:mongoose_api_common.arg_name/0

arg_spec_list()

Specs

t:mongoose_api_common.arg_spec_list/0

arg_value()

Specs

t:mongoose_api_common.arg_value/0

args_applied()

Specs

t:mongoose_api_common.args_applied/0

correct_result()

Specs

t:mongoose_api_common.correct_result/0

error_result()

Specs

t:mongoose_api_common.error_result/0

http_api_state()

Specs

t:mongoose_api_common.http_api_state/0

method()

Specs

t:mongoose_api_common.method/0

optarg_spec_list()

Specs

t:mongoose_api_common.optarg_spec_list/0

Functions

action_to_method/1

add_bind/2

add_bindings/2

add_location_header/3

both_bind_and_body/1

Specs

mongoose_api_common.both_bind_and_body/1

check_and_extract_args/3

Specs

mongoose_api_common.check_and_extract_args/3

Checks if the arguments are correct. Return the arguments that can be applied to the execution of command.

convert_arg/2

Specs

mongoose_api_common.convert_arg/2

create_admin_url_path/1

Specs

mongoose_api_common.create_admin_url_path/1

create_params_proplist/1

Specs

mongoose_api_common.create_params_proplist/1

create_user_url_path/1

Specs

mongoose_api_common.create_user_url_path/1

error_code/1

error_response/4

Specs

mongoose_api_common.error_response/4

execute_command/3

Specs

mongoose_api_common.execute_command/3

get_allowed_methods/1

Specs

mongoose_api_common.get_allowed_methods/1

Returns list of allowed methods.

get_auth_details/1

Specs

mongoose_api_common.get_auth_details/1

handle_request/5

Specs

mongoose_api_common.handle_request/5

handle_result/4

Specs

mongoose_api_common.handle_result/4

is_known_auth_method/1

Specs

mongoose_api_common.is_known_auth_method/1

make_unauthorized_response/2

maybe_add_bindings/2

Specs

mongoose_api_common.maybe_add_bindings/2

maybe_add_caller/1

Specs

mongoose_api_common.maybe_add_caller/1

maybe_add_location_header/3

Specs

mongoose_api_common.maybe_add_location_header/3

maybe_add_subcategory/1

method_to_action/1

parse_request_body/1

Specs

mongoose_api_common.parse_request_body/1

process_request/4

Specs

mongoose_api_common.process_request/4

reload_dispatches/1

Reload all ejabberd_cowboy listeners. When a command is registered or unregistered, the routing paths that cowboy stores as a "dispatch" must be refreshed. Read more http://ninenines.eu/docs/en/cowboy/1.0/guide/routing/

to_atom/1

Specs

mongoose_api_common.to_atom/1