mongoose_cassandra behaviour (MongooseIM v4.0.1)

Types

fold_accumulator()

Specs

t:mongoose_cassandra.fold_accumulator/0

fold_fun()

Specs

t:mongoose_cassandra.fold_fun/0

parameters()

Specs

t:mongoose_cassandra.parameters/0

pool_name()

Specs

t:mongoose_cassandra.pool_name/0

query_name()

Specs

t:mongoose_cassandra.query_name/0

row()

Specs

t:mongoose_cassandra.row/0

rows()

Specs

t:mongoose_cassandra.rows/0

Functions

cql_foldl/7

Specs

mongoose_cassandra.cql_foldl/7

Execute read query to cassandra (select). This functions behaves much like the lists:foldl/3 but the input are pages from result of given query. Therefore each execution of given fun gets list of several result rows (by default 100 at most). --------------------------------------------------------

cql_read/5

Specs

mongoose_cassandra.cql_read/5

Execute read query to cassandra (select). Returns all rows at once even if there are several query pages. --------------------------------------------------------

cql_write/5

Specs

mongoose_cassandra.cql_write/5

Execute batch write query to cassandra (insert, update or delete). Note that Cassandra doesn't like big batches, therefore this function will try to split given rows into batches of 20 rows and will fall back to smaller batches if Cassandra rejects the query due to its size being to big. --------------------------------------------------------

cql_write_async/5

Specs

mongoose_cassandra.cql_write_async/5

Execute async batch write query to cassandra (insert, update or delete). Note that Cassandra doesn't like big batches and there's not retry login when query size will be exceeded like in cql_write/5. --------------------------------------------------------

now_timestamp/0

Return timestamp in microseconds

prepared_queries/0

tables/0

test_query/1

Specs

mongoose_cassandra.test_query/1

test_query/2

test_query_sql/0

total_count_queries/0

total_count_query/2

Specs

mongoose_cassandra.total_count_query/2

total_count_query_cql/1

Callbacks

prepared_queries()

Specs

c:mongoose_cassandra.prepared_queries/0