mongoose_cassandra_worker (MongooseIM v4.0.1)

Types

action()

Specs

t:mongoose_cassandra_worker.action/0

batch_mode()

Specs

t:mongoose_cassandra_worker.batch_mode/0

batch_mode_int()

Specs

t:mongoose_cassandra_worker.batch_mode_int/0

caller()

Specs

t:mongoose_cassandra_worker.caller/0

consistency_level()

Specs

t:mongoose_cassandra_worker.consistency_level/0

consistency_level_int()

Specs

t:mongoose_cassandra_worker.consistency_level_int/0

context_id()

Specs

t:mongoose_cassandra_worker.context_id/0

This value is used as a hash base in worker selection

cql_mref()

Specs

t:mongoose_cassandra_worker.cql_mref/0

cql_query()

Specs

t:mongoose_cassandra_worker.cql_query/0

cql_result()

Specs

t:mongoose_cassandra_worker.cql_result/0

cql_tag()

Specs

t:mongoose_cassandra_worker.cql_tag/0

Normally it's a reference(), but let's not depend on that

datatype()

Specs

t:mongoose_cassandra_worker.datatype/0

error_reason()

Specs

t:mongoose_cassandra_worker.error_reason/0

error_type()

Specs

t:mongoose_cassandra_worker.error_type/0

named_parameter()

Specs

t:mongoose_cassandra_worker.named_parameter/0

options()

Specs

t:mongoose_cassandra_worker.options/0

parameter()

Specs

t:mongoose_cassandra_worker.parameter/0

parameter_val()

Specs

t:mongoose_cassandra_worker.parameter_val/0

process_type()

Specs

t:mongoose_cassandra_worker.process_type/0

query_str()

Specs

t:mongoose_cassandra_worker.query_str/0

read_action()

Specs

t:mongoose_cassandra_worker.read_action/0

request()

Specs

t:mongoose_cassandra_worker.request/0

request_id()

Specs

t:mongoose_cassandra_worker.request_id/0

serial_consistency()

Specs

t:mongoose_cassandra_worker.serial_consistency/0

serial_consistency_int()

Specs

t:mongoose_cassandra_worker.serial_consistency_int/0

worker_state()

Specs

t:mongoose_cassandra_worker.worker_state/0

write_action()

Specs

t:mongoose_cassandra_worker.write_action/0

Functions

cleanup_request/2

Specs

mongoose_cassandra_worker.cleanup_request/2

code_change/3

Callback implementation for :gen_server.code_change/3.

cql_error_text/1

do_handle_error/4

Specs

mongoose_cassandra_worker.do_handle_error/4

do_handle_result/3

Specs

mongoose_cassandra_worker.do_handle_result/3

do_read/5

error_text/3

get_client/1

Specs

mongoose_cassandra_worker.get_client/1

get_client_loop/2

Specs

mongoose_cassandra_worker.get_client_loop/2

handle_call/3

Callback implementation for :gen_server.handle_call/3.

handle_cancel/3

Specs

mongoose_cassandra_worker.handle_cancel/3

handle_cast/2

Callback implementation for :gen_server.handle_cast/2.

handle_down/2

Specs

mongoose_cassandra_worker.handle_down/2

handle_error/2

Specs

mongoose_cassandra_worker.handle_error/2

handle_info/2

Callback implementation for :gen_server.handle_info/2.

handle_result/2

Specs

mongoose_cassandra_worker.handle_result/2

init/1

Callback implementation for :gen_server.init/1.

is_finished/2

Specs

mongoose_cassandra_worker.is_finished/2

maybe_reply/2

Specs

mongoose_cassandra_worker.maybe_reply/2

new_read_action/2

Specs

mongoose_cassandra_worker.new_read_action/2

new_request/3

Specs

mongoose_cassandra_worker.new_request/3

new_write_action/2

Specs

mongoose_cassandra_worker.new_write_action/2

prepare_options/1

Specs

mongoose_cassandra_worker.prepare_options/1

process_request/2

Specs

mongoose_cassandra_worker.process_request/2

query_type/1

read/7

Specs

mongoose_cassandra_worker.read/7

Execute read query to Cassandra (select). This functions behaves much like the lists:foldl/3 but the input are pages from result of given query. Therefore each execution of given fun gets list of several result rows (by default 100 at most).

response_code_to_binary/1

retry_info/4

Specs

mongoose_cassandra_worker.retry_info/4

schedule_retry/3

Specs

mongoose_cassandra_worker.schedule_retry/3

schedule_timeout/2

send_next_request/3

Specs

mongoose_cassandra_worker.send_next_request/3

terminate/2

Callback implementation for :gen_server.terminate/2.

update_req/2

write/5

Specs

mongoose_cassandra_worker.write/5

Execute batch write query to Cassandra (insert, update or delete). Note that Cassandra doesn't like big batches, therefore this function will try to split given rows into batches of 20 rows and will fall back to smaller batches if Cassandra rejects the query due to its size being to big.

write_async/5

Specs

mongoose_cassandra_worker.write_async/5

Same as @see write/5 but asynchronous (without response).