mongoose_cluster (MongooseIM v4.0.1)

Functions

change_schema_type/1

check_if_successful_copied/1

check_networking/1

delete_mnesia/0

detach_nodes/1

do_join/1

do_leave/0

do_remove_from_cluster/1

is_app_running/1

is_node_alive/1

join/1

Specs

mongoose_cluster.join/1

Join a cluster designated by ClusterMember. This drops all current connections and discards all persistent data from Mnesia. Use with caution! Next time the node starts, it will connect to other members automatically.

leave/0

Specs

mongoose_cluster.leave/0

Leave cluster. This drops all current connections and discards all persistent data from Mnesia. Use with caution! Next time the node starts, it will NOT connect to previous members. Remaining members will remove this node from the cluster Mnesia schema.

mnesia_nodes/0

node_trans/1

remove_dead_from_cluster/1

remove_from_cluster/1

Specs

mongoose_cluster.remove_from_cluster/1

Remove dead node from the cluster. The removing node must be down

rmrf/1

set_extra_db_nodes/1

table_type/2

unsafe_join/2

wait_for_pong/1

wait_for_pong/4

with_app_stopped/2