mongoose_config_flat (MongooseIM v4.0.1)

Pure flat config logic. No ets table manipulations, no Mnesia, no starting modules, no file reading here. Everything here is safe, side effect free. OK, logging is possible, but keep it to minimum.
Flat format is a list of pairs, corresponding to some part of config tree. Keys are: - [g, K] - global option; - [a, K] - acl list; - [l, listener, Address, Module] - enabled listener; - [l, listener_opt, Address, Module, OptName] - option with value; - [l, listener_simple_opt, Address, Module, Opt] - option without value; - [h, Host, modules] - we need this to build tree (expand options back); - [h, Host, module, Module] - for an enabled module; - [h, H, module_opt, Module, OptName] - option of module; - [h, Host, module_subopt, Module, OptName|Path] - if value of OptName is a proplist, we split is into module_subopt-s recursively.

Types

flat_option()

Specs

t:mongoose_config_flat.flat_option/0

flat_options()

Specs

t:mongoose_config_flat.flat_options/0

Functions

add_to_group/3

Specs

mongoose_config_flat.add_to_group/3

Append into a group for later use Returns updated Groups

can_be_flattened_value/1

do_split_node_specific_options_tree/1

does_pattern_match/2

Specs

mongoose_config_flat.does_pattern_match/2

Use erlang match spec to check if the Subject fits the Pattern. http://erlang.org/doc/apps/erts/match_spec.html

expand_all_opts/1

expand_global_config_group/2

expand_global_config_group_item/3

expand_host_config_group/2

expand_host_config_group_item/4

expand_host_config_group_item_value/4

expand_local_config_group/2

Rewrite LocalConfigGroup to {local_config, _, _} format using Groups for quick lookups

expand_local_config_group_item/4

expand_local_config_group_item_value/3

expand_module_opts/4

expand_module_subopts/4

expand_module_subopts_path/4

expand_opts/1

Convert flat list of options back to LocalConfig and HostLocalConfig Useful, but not used

find_matching_node_specific_pattern/2

flatten_acl_config_opt/2

flatten_all_opts/1

flatten_global_config_opt/2

flatten_global_config_opts/1

flatten_listen/3

flatten_listen_opts/3

flatten_local_config_host_opt/3

flatten_local_config_host_opts/1

flatten_local_config_opt/2

flatten_local_config_opts/1

flatten_module/3

flatten_module_opts/3

flatten_module_subopts/4

flatten_opts/2

Convert LocalConfig and HostLocalConfig to a flat list of options

flatten_subopt/3

flatten_subopts/2

group_flat_opt/2

group_flat_opts/1

group opts for faster lookups in expand_opts

last_init/1

make_listener_opts_from_groups/3

make_listeners_from_groups/1

make_module_opts_from_groups/3

make_modules_for_host_from_groups/2

node_specific_orphans/1

split_node_specific_options/2

split_node_specific_options/5

split_node_specific_options_tree/2

subtract_lists/2