mongoose_config_reload (MongooseIM v4.0.1)

Pure config reloading logic. No ets table manipulations, no Mnesia, no starting modules, no file reading here. Everything here is safe, side effect free. OK, logging is possible, but keep it to minimum.

Types

categorized_options()

Specs

t:mongoose_config_reload.categorized_options/0

compare_result()

Specs

t:mongoose_config_reload.compare_result/0

config_diff()

Specs

t:mongoose_config_reload.config_diff/0

config_version()

Specs

t:mongoose_config_reload.config_version/0

node_state()

Specs

t:mongoose_config_reload.node_state/0

reloading_change()

Specs

t:mongoose_config_reload.reloading_change/0

reloading_context()

Specs

t:mongoose_config_reload.reloading_context/0

state()

Specs

t:mongoose_config_reload.state/0

Functions

all_empty/2

all_equal_to/3

all_same/2

as_global/2

as_hosts/2

as_local/2

calculate_changes/1

calculate_changes_on_coordinator/1

Specs

mongoose_config_reload.calculate_changes_on_coordinator/1

calculate_changes_on_remote_node/1

Specs

mongoose_config_reload.calculate_changes_on_remote_node/1

calculate_inconsistent_opts/1

can_be_ignored/1

Specs

mongoose_config_reload.can_be_ignored/1

categorize_option/2

categorize_options/1

Specs

mongoose_config_reload.categorize_options/1

categorize_options_to_flat_global_config_opts/1

Specs

mongoose_config_reload.categorize_options_to_flat_global_config_opts/1

categorize_options_to_flat_local_config_opts/1

Specs

mongoose_config_reload.categorize_options_to_flat_local_config_opts/1

check/3

check_hosts/2

Specs

mongoose_config_reload.check_hosts/2

cluster_reload_version_check/1

cluster_reload_version_checks/1

compare_listeners/2

Specs

mongoose_config_reload.compare_listeners/2

compare_terms/4

Specs

mongoose_config_reload.compare_terms/4

compute_config_file_version/1

Specs

mongoose_config_reload.compute_config_file_version/1

compute_config_version/2

context_to_changes_to_apply/1

Specs

mongoose_config_reload.context_to_changes_to_apply/1

context_to_failed_checks/1

Specs

mongoose_config_reload.context_to_failed_checks/1

extend_node_state/1

extend_node_states/1

filter_out_empty_value_pairs/1

find_modules_to_change/5

find_modules_to_start/4

flat_global_opts_version/1

get_config_diff/2

Specs

mongoose_config_reload.get_config_diff/2

get_key_group/2

Specs

mongoose_config_reload.get_key_group/2

group_host_changes/1

Specs

mongoose_config_reload.group_host_changes/1

inconsistent_node_opts/1

is_not_host_specific/1

Specs

mongoose_config_reload.is_not_host_specific/1

is_special_config_opt/1

make_categorized_options/3

Specs

mongoose_config_reload.make_categorized_options/3

Make a categorize_options() map Does not work with acls

map_listeners/1

node_values/2

prepare_data_for_cluster_reloading_context/1

skip_special_config_opts/1

special_local_config_opts/0

state_to_categorized_options/1

Specs

mongoose_config_reload.state_to_categorized_options/1

state_to_flat_local_opts/1

It ignores global options

states_to_reloading_context/1

Specs

mongoose_config_reload.states_to_reloading_context/1

subtract_lists/2