mongoose_config_terms (MongooseIM v4.0.1)

Parsing of the 'cfg' file to Erlang terms

Types

config_problem()

Specs

t:mongoose_config_terms.config_problem/0

spec me better

Functions

config_filenames_to_include/1

delete_disallowed/2

Specs

mongoose_config_terms.delete_disallowed/2

Filter from the list of terms the disallowed. Returns a sublist of Terms without the ones which first element is included in Disallowed.

delete_disallowed2/2

describe_config_problem/2

Specs

mongoose_config_terms.describe_config_problem/2

describe_config_problem/3

Specs

mongoose_config_terms.describe_config_problem/3

find_plain_terms_for_file/2

get_plain_terms_file/1

Specs

mongoose_config_terms.get_plain_terms_file/1

Read an ejabberd configuration file and return the terms. Input is an absolute or relative path to an ejabberd config file. Returns a list of plain terms, in which the options 'include_config_file' were parsed and the terms in those files were included.

include_config_files/1

Specs

mongoose_config_terms.include_config_files/1

Include additional configuration files in the list of terms.

include_config_files/2

include_config_files/3

keep_only_allowed/2

Specs

mongoose_config_terms.keep_only_allowed/2

Keep from the list only the allowed terms. Returns a sublist of Terms with only the ones which first element is included in Allowed.