mongoose_config_utils (MongooseIM v4.0.1)

Short functions useful for config file manipulations. This stuff can be pure, but most likely not. It's for generic functions.

Types

config_line()

Specs

t:mongoose_config_utils.config_line/0

spec me better

Functions

exit_or_halt/1

Specs

mongoose_config_utils.exit_or_halt/1

If MongooseIM isn't yet running in this node, then halt the node

get_absolute_path/1

Specs

mongoose_config_utils.get_absolute_path/1

Convert configuration filename to absolute path. Input is an absolute or relative path to an ejabberd configuration file. And returns an absolute path to the configuration file.

get_config_lines/4

Specs

mongoose_config_utils.get_config_lines/4

get_config_lines2/5

is_file_readable/1

Specs

mongoose_config_utils.is_file_readable/1