mongoose_deprecations (MongooseIM v4.0.1)

This module is responsible for initialising and stopping stuff needed to handle different deprecation warinings as well as for exposing API for loging these deprecations. It checks whether a specific deprecation warning is not exceeding given frequency of logging.

Types

deprecation_tag()

Specs

t:mongoose_deprecations.deprecation_tag/0

Specifies the deprecation

log_level()

Specs

t:mongoose_deprecations.log_level/0

log_map()

Specs

t:mongoose_deprecations.log_map/0

unix_timestamp()

Specs

t:mongoose_deprecations.unix_timestamp/0

Functions

default_cooldown/0

Specs

mongoose_deprecations.default_cooldown/0

default_log_lvl/0

Specs

mongoose_deprecations.default_log_lvl/0

destroy_ets/0

Specs

mongoose_deprecations.destroy_ets/0

did_cooldown_elapse/2

Specs

mongoose_deprecations.did_cooldown_elapse/2

log/2

Specs

mongoose_deprecations.log/2

log/3

Specs

mongoose_deprecations.log/3

Should be used to log deprecation messages. It logs keeping proper frequency. Opts can be: * cooldown - the minimal interval (in milliseconds) to be held between logs. Default: 6 hours It is internally represented in microseconds but API requires milliseconds. * log_level - 'warning' or 'error'

log_with_lvl/2

Specs

mongoose_deprecations.log_with_lvl/2

maybe_log/4

Specs

mongoose_deprecations.maybe_log/4

milliseconds_to_microseconds/1

Specs

mongoose_deprecations.milliseconds_to_microseconds/1

prepare_ets/0

Specs

mongoose_deprecations.prepare_ets/0

Deprecation table will hold pairs in form: {deprecation_tag(), unix_timestamp()} and will indicate at what (unix) time last deprecation warining was logged concerning a deprecation connected to the deprecation tag specified in a key

start/0

Specs

mongoose_deprecations.start/0

Should be called before using the module. Sets everything needed up

stop/0

Specs

mongoose_deprecations.stop/0

Used after using the module, when we won't log deprecation messages again.