mongoose_elasticsearch (MongooseIM v4.0.1)

Provides functions to interact with ElasticSearch cluster via HTTP API. ------------------------------------------------------------------

Types

document()

Specs

t:mongoose_elasticsearch.document/0

id()

Specs

t:mongoose_elasticsearch.id/0

index()

Specs

t:mongoose_elasticsearch.index/0

query()

Specs

t:mongoose_elasticsearch.query/0

type()

Specs

t:mongoose_elasticsearch.type/0

Functions

count/3

Specs

mongoose_elasticsearch.count/3

Retrieves count of documents matching given search query in given ElasticSearch index.
See https://www.elastic.co/guide/en/elasticsearch/reference/6.2/search-count.html for more information.

delete_by_query/3

Specs

mongoose_elasticsearch.delete_by_query/3

Deletes documents matching a query in a given ElasticSearch index.
See https://www.elastic.co/guide/en/elasticsearch/reference/6.2/docs-delete-by-query.html for more information.

health/0

Specs

mongoose_elasticsearch.health/0

Returns the health status of the ElasticSearch cluster.
See https://www.elastic.co/guide/en/elasticsearch/reference/current/cluster-health.html for more information.

insert_document/4

Specs

mongoose_elasticsearch.insert_document/4

Tries to insert a document into given ElasticSearch index.
See https://www.elastic.co/guide/en/elasticsearch/reference/6.2/docs-index_.html for more information.

search/3

Specs

mongoose_elasticsearch.search/3

Runs a search query on a given ElasticSearch index.
See https://www.elastic.co/guide/en/elasticsearch/reference/6.2/search.html for more information.