mongoose_hooks (MongooseIM v4.0.1)

Hooks wrapper providing clear specifications for a hook caller.
Every hook has its own function in this module with specs as accurate as possible. This helps to have a static analysis of the hooks callers to make sure they pass the expected arguments.

Functions

adhoc_local_items/5

Specs

mongoose_hooks.adhoc_local_items/5

adhoc_sm_items/5

Specs

mongoose_hooks.adhoc_sm_items/5

amp_check_condition/4

Specs

mongoose_hooks.amp_check_condition/4

The amp_check_condition hook is called to determine whether the AMP strategy matches the given AMP rule.

amp_check_packet/4

Specs

mongoose_hooks.amp_check_packet/4

The amp_check_packet hook is called when one wants to check a message against amp rules. Calling it may result in mod_amp taking actions according to the rules and the event (i.e. notify the sender).

amp_determine_strategy/6

Specs

mongoose_hooks.amp_determine_strategy/6

The amp_determine_strategy hook is called when checking to determine which strategy will be chosen when executing AMP rules.

amp_error_action_triggered/2

Specs

mongoose_hooks.amp_error_action_triggered/2

The amp_error_action_triggered hook is called to inform that the error action has been triggered.

amp_notify_action_triggered/2

Specs

mongoose_hooks.amp_notify_action_triggered/2

The amp_notify_action_triggered hook is called to inform that the notify action has been triggered.

amp_verify_support/3

Specs

mongoose_hooks.amp_verify_support/3

The amp_verify_support hook is called when checking whether the host supports given AMP rules.

anonymous_purge_hook/3

Specs

mongoose_hooks.anonymous_purge_hook/3

The anonymous_purge_hook hook is called when anonymous user's data is removed.

auth_failed/2

Specs

mongoose_hooks.auth_failed/2

c2s_broadcast_recipients/6

Specs

mongoose_hooks.c2s_broadcast_recipients/6

c2s_filter_packet/6

Specs

mongoose_hooks.c2s_filter_packet/6

c2s_preprocessing_hook/3

Specs

mongoose_hooks.c2s_preprocessing_hook/3

c2s_presence_in/5

Specs

mongoose_hooks.c2s_presence_in/5

c2s_remote_hook/5

Specs

mongoose_hooks.c2s_remote_hook/5

c2s_stream_features/2

Specs

mongoose_hooks.c2s_stream_features/2

c2s_unauthenticated_iq/4

Specs

mongoose_hooks.c2s_unauthenticated_iq/4

c2s_update_presence/2

Specs

mongoose_hooks.c2s_update_presence/2

can_access_identity/4

Specs

mongoose_hooks.can_access_identity/4

The can_access_identity hook is called to determine if a given user can see the real identity of the people in a room.

can_access_room/4

Specs

mongoose_hooks.can_access_room/4

The can_access_room hook is called to determine if a given user can access a room.

caps_add/6

Specs

mongoose_hooks.caps_add/6

caps_recognised/5

Specs

mongoose_hooks.caps_recognised/5

caps_update/6

Specs

mongoose_hooks.caps_update/6

check_bl_c2s/2

Specs

mongoose_hooks.check_bl_c2s/2

disco_info/5

Specs

mongoose_hooks.disco_info/5

disco_info hook is called to extract information about the server.

disco_local_features/6

Specs

mongoose_hooks.disco_local_features/6

disco_local_features hook is called to extract features offered by the server.

disco_local_identity/6

Specs

mongoose_hooks.disco_local_identity/6

disco_local_identity hook is called to get the identity of the server.

disco_local_items/6

Specs

mongoose_hooks.disco_local_items/6

disco_local_items hook is called to extract items associated with the server.

disco_sm_features/6

Specs

mongoose_hooks.disco_sm_features/6

disco_sm_features hook is called to get the features of the client when a discovery IQ gets to session management.

disco_sm_identity/6

Specs

mongoose_hooks.disco_sm_identity/6

disco_sm_identity hook is called to get the identity of the client when a discovery IQ gets to session management.

disco_sm_items/6

Specs

mongoose_hooks.disco_sm_items/6

disco_sm_items hook is called to get the items associated with the client when a discovery IQ gets to session management.

ejabberd_ctl_process/2

Specs

mongoose_hooks.ejabberd_ctl_process/2

failed_to_store_message/2

Specs

mongoose_hooks.failed_to_store_message/2

filter_local_packet/2

Specs

mongoose_hooks.filter_local_packet/2

The filter_local_packet hook is called to filter out stanzas routed with mongoose_local_delivery.

filter_packet/1

Specs

mongoose_hooks.filter_packet/1

The filter_packet hook is called to filter out stanzas routed with mongoose_router_global.

filter_room_packet/3

Specs

mongoose_hooks.filter_room_packet/3

find_s2s_bridge/3

Specs

mongoose_hooks.find_s2s_bridge/3

find_s2s_bridge hook is called to find a s2s bridge to a foreign protocol when opening a socket to a different XMPP server fails.

forbidden_session_hook/3

Specs

mongoose_hooks.forbidden_session_hook/3

forget_room/4

Specs

mongoose_hooks.forget_room/4

The forget_room hook is called when a room is removed from the database.

get_key/3

Specs

mongoose_hooks.get_key/3

The get_key hook is called to extract a key from mod_keystore.

get_mam_muc_gdpr_data/3

Specs

mongoose_hooks.get_mam_muc_gdpr_data/3

get_mam_muc_gdpr_data hook is called to provide a user's archive for GDPR purposes.

get_mam_pm_gdpr_data/3

Specs

mongoose_hooks.get_mam_pm_gdpr_data/3

get_mam_pm_gdpr_data hook is called to provide a user's archive for GDPR purposes.

get_personal_data/3

Specs

mongoose_hooks.get_personal_data/3

get_personal_data hook is called to retrieve a user's personal data for GDPR purposes.

host_config_update/4

Specs

mongoose_hooks.host_config_update/4

inbox_unread_count/3

Specs

mongoose_hooks.inbox_unread_count/3

The inbox_unread_count hook is called to get the number of unread messages in the inbox for a user.

invitation_sent/7

Specs

mongoose_hooks.invitation_sent/7

is_muc_room_owner/4

Specs

mongoose_hooks.is_muc_room_owner/4

The is_muc_room_owner hooks is called to determine if a given user is a room's owner.
The hook's handler needs to expect the following arguments Acc, Room, User. The arguments and the return value types correspond to the following spec.

join_room/6

Specs

mongoose_hooks.join_room/6

The join_room hook is called when a user joins a MUC room.

leave_room/6

Specs

mongoose_hooks.leave_room/6

The leave_room hook is called when a user joins a MUC room.

local_send_to_resource_hook/5

Specs

mongoose_hooks.local_send_to_resource_hook/5

mam_archive_id/3

Specs

mongoose_hooks.mam_archive_id/3

The mam_archive_id hook is called to determine the id of an archive for a particular user or entity. The hook handler is expected to accept the following arguments: * Acc with an initial value of undefined, * Host as passed in the HooksServer variable, * OwnerJID,

and to return an integer value corresponding to the given owner's archive.
If a MAM backend doesn't support or doesn't require archive IDs, undefined may be returned.

mam_archive_message/3

Specs

mongoose_hooks.mam_archive_message/3

The mam_archive_message hook is called in order to store the message in the archive.

mam_archive_size/4

Specs

mongoose_hooks.mam_archive_size/4

The mam_archive_size hook is called to determine the size of the archive for a given JID

mam_get_behaviour/5

Specs

mongoose_hooks.mam_get_behaviour/5

The mam_get_behaviour hooks is called to determine if a message should be archived or not based on a given pair of JIDs.

mam_get_prefs/4

Specs

mongoose_hooks.mam_get_prefs/4

The mam_get_prefs hook is called to read the archive settings for a given user.

mam_lookup_messages/3

Specs

mongoose_hooks.mam_lookup_messages/3

The mam_lookup_messages hook is to retrieve archived messages for given search parameters.

mam_muc_archive_id/3

Specs

mongoose_hooks.mam_muc_archive_id/3

The mam_muc_archive_id hook is called to determine the archive ID for a particular room. The hook handler is expected to accept the following arguments: * Acc with initial value undefined, * Host as passed in HooksServer variable, * OwnerJID,

and return an integer value corresponding to the given owner's archive.
If a MAM backend doesn't support or doesn't require archive IDs, undefined may be returned.

mam_muc_archive_message/3

Specs

mongoose_hooks.mam_muc_archive_message/3

The mam_muc_archive_message hook is called in order to store the MUC message in the archive.

mam_muc_archive_size/4

Specs

mongoose_hooks.mam_muc_archive_size/4

The mam_muc_archive_size hook is called to determine the archive's size for a given room.

mam_muc_flush_messages/3

Specs

mongoose_hooks.mam_muc_flush_messages/3

The mam_muc_flush_messages hook is run after the async bulk write happens for MUC messages despite the result of the write.

mam_muc_get_behaviour/5

Specs

mongoose_hooks.mam_muc_get_behaviour/5

The mam_muc_get_behaviour hooks is called to determine if a message should be archived or not based on the given room and user JIDs.

mam_muc_get_prefs/4

Specs

mongoose_hooks.mam_muc_get_prefs/4

The mam_muc_get_prefs hook is called to read the archive settings for a given room.

mam_muc_lookup_messages/3

Specs

mongoose_hooks.mam_muc_lookup_messages/3

The mam_muc_lookup_messages hook is to retrieve archived MUC messages for any given search parameters.

mam_muc_remove_archive/4

Specs

mongoose_hooks.mam_muc_remove_archive/4

The mam_muc_remove_archive hook is called in order to remove the entire archive for a particular user.

mam_muc_set_prefs/7

Specs

mongoose_hooks.mam_muc_set_prefs/7

The mam_muc_set_prefs hook is called to set a room's archive preferences.
It's possible to set which JIDs are always or never allowed in the archive

mam_remove_archive/4

Specs

mongoose_hooks.mam_remove_archive/4

The mam_remove_archive hook is called in order to remove the entire archive for a particular user.

mam_set_prefs/7

Specs

mongoose_hooks.mam_set_prefs/7

The mam_set_prefs hook is called to set a user's archive preferences.
It's possible to set which JIDs are always or never allowed in the archive

mod_global_distrib_known_recipient/5

Specs

mongoose_hooks.mod_global_distrib_known_recipient/5

The mod_global_distrib_known_recipient hook is called when the recipient is known to global_distrib.

mod_global_distrib_unknown_recipient/2

Specs

mongoose_hooks.mod_global_distrib_unknown_recipient/2

The mod_global_distrib_unknown_recipient hook is called when the recipient is unknown to global_distrib.

muc_room_pid/3

Specs

mongoose_hooks.muc_room_pid/3

The muc_room_pid hooks is called to get the pid for a given room's JID

offline_groupchat_message_hook/5

Specs

mongoose_hooks.offline_groupchat_message_hook/5

offline_message_hook/5

Specs

mongoose_hooks.offline_message_hook/5

packet_to_component/3

Specs

mongoose_hooks.packet_to_component/3

presence_probe_hook/5

Specs

mongoose_hooks.presence_probe_hook/5

privacy_check_packet/6

Specs

mongoose_hooks.privacy_check_packet/6

privacy_get_user_list/3

Specs

mongoose_hooks.privacy_get_user_list/3

privacy_iq_get/6

Specs

mongoose_hooks.privacy_iq_get/6

privacy_iq_set/5

Specs

mongoose_hooks.privacy_iq_set/5

privacy_updated_list/4

Specs

mongoose_hooks.privacy_updated_list/4

pubsub_create_node/6

Specs

mongoose_hooks.pubsub_create_node/6

The pubsub_create_node hook is called to inform that a pubsub node is created.

pubsub_delete_node/5

Specs

mongoose_hooks.pubsub_delete_node/5

The pubsub_delete_node hook is called to inform that a pubsub node is deleted.

pubsub_publish_item/7

Specs

mongoose_hooks.pubsub_publish_item/7

The pubsub_publish_item hook is called to inform that a pubsub item is published.

push_notifications/4

Specs

mongoose_hooks.push_notifications/4

The push_notifications hook is called to push notifications.

register_command/2

Specs

mongoose_hooks.register_command/2

The register_command hook is called when a command is registered in mongoose_commands.

register_subhost/3

Specs

mongoose_hooks.register_subhost/3

The register_subhost hook is called when a component is registered for ejabberd_router.

register_user/3

Specs

mongoose_hooks.register_user/3

The register_user hook is called when a user is successfully registered in an authentication backend.

remove_user/3

Specs

mongoose_hooks.remove_user/3

The remove_user hook is called when a user is removed.

resend_offline_messages_hook/3

Specs

mongoose_hooks.resend_offline_messages_hook/3

rest_user_send_packet/5

Specs

mongoose_hooks.rest_user_send_packet/5

The rest_user_send_packet hook is called when a user sends a message using the REST API.

room_packet/6

Specs

mongoose_hooks.room_packet/6

The room_packet hook is called when a message is added to room's history.

room_send_packet/3

Specs

mongoose_hooks.room_send_packet/3

The room_send_packet hook is called when a message is sent to a room.

roster_get/4

Specs

mongoose_hooks.roster_get/4

The roster_get hook is called to extract a user's roster.

roster_get_jid_info/4

Specs

mongoose_hooks.roster_get_jid_info/4

The roster_get_jid_info hook is called to determine the subscription state between a given pair of users. The hook handlers need to expect following arguments: * Acc with an initial value of {none, []}, * LUser, a stringprepd username part of the roster's owner, * LServer, a stringprepd server part of the roster's owner (same value as HookServer), * RemoteBareJID, a bare JID of the other user.
The arguments and the return value types correspond to the following spec.

roster_get_subscription_lists/3

Specs

mongoose_hooks.roster_get_subscription_lists/3

The roster_get_subscription_lists hook is called to extract user's subscription list.

roster_get_versioning_feature/2

Specs

mongoose_hooks.roster_get_versioning_feature/2

The roster_get_versioning_feature hook is called to determine if roster versioning is enabled.

roster_groups/2

Specs

mongoose_hooks.roster_groups/2

The roster_groups hook is called to extract roster groups.

roster_in_subscription/7

Specs

mongoose_hooks.roster_in_subscription/7

The roster_in_subscription hook is called to determine if a subscription presence is routed to a user.

roster_out_subscription/5

Specs

mongoose_hooks.roster_out_subscription/5

The roster_out_subscription hook is called when a user sends out subscription.

roster_process_item/2

Specs

mongoose_hooks.roster_process_item/2

The roster_process_item hook is called when a user's roster is set.

roster_push/4

Specs

mongoose_hooks.roster_push/4

The roster_push hook is called when a roster item is being pushed and roster versioning is not enabled.

roster_set/5

Specs

mongoose_hooks.roster_set/5

The roster_set hook is called when a user's roster is set through an IQ.

s2s_allow_host/3

Specs

mongoose_hooks.s2s_allow_host/3

s2s_allow_host hook is called to check whether a server should be allowed to be connected to.
A handler can decide that a server should not be allowed and pass this information to the caller.

s2s_connect_hook/3

Specs

mongoose_hooks.s2s_connect_hook/3

s2s_connect_hook hook is called when a s2s connection is established.

s2s_receive_packet/2

Specs

mongoose_hooks.s2s_receive_packet/2

s2s_receive_packet hook is called when an incoming stanza is routed by the server.

s2s_send_packet/5

Specs

mongoose_hooks.s2s_send_packet/5

s2s_send_packet hook is called when a message is routed.

s2s_stream_features/2

Specs

mongoose_hooks.s2s_stream_features/2

s2s_stream_features hook is used to extract the stream management features supported by the server.

session_cleanup/5

Specs

mongoose_hooks.session_cleanup/5

The session_cleanup hook is called when sm backend cleans up a user's session.

session_opening_allowed_for_user/3

Specs

mongoose_hooks.session_opening_allowed_for_user/3

set_presence_hook/5

Specs

mongoose_hooks.set_presence_hook/5

set_vcard/4

Specs

mongoose_hooks.set_vcard/4

The set_vcard hook is called when the caller wants to set the VCard.

sm_broadcast/6

Specs

mongoose_hooks.sm_broadcast/6

sm_filter_offline_message/5

Specs

mongoose_hooks.sm_filter_offline_message/5

sm_register_connection_hook/4

Specs

mongoose_hooks.sm_register_connection_hook/4

sm_remove_connection_hook/6

Specs

mongoose_hooks.sm_remove_connection_hook/6

unacknowledged_message/3

Specs

mongoose_hooks.unacknowledged_message/3

unregister_command/2

Specs

mongoose_hooks.unregister_command/2

The unregister_command hook is called when a command is unregistered from mongoose_commands.

unregister_subhost/2

Specs

mongoose_hooks.unregister_subhost/2

The unregister_subhost hook is called when a component is unregistered from ejabberd_router.

unset_presence_hook/5

Specs

mongoose_hooks.unset_presence_hook/5

update_inbox_for_muc/2

Specs

mongoose_hooks.update_inbox_for_muc/2

user_available_hook/3

Specs

mongoose_hooks.user_available_hook/3

user_ping_response/5

Specs

mongoose_hooks.user_ping_response/5

The user_ping_response hook is called when a user responds to a ping.

user_ping_timeout/3

Specs

mongoose_hooks.user_ping_timeout/3

The user_ping_timeout hook is called when there is a timeout when waiting for a ping response from a user.

user_receive_packet/6

Specs

mongoose_hooks.user_receive_packet/6

user_send_packet/5

Specs

mongoose_hooks.user_send_packet/5

A hook called when a user sends an XMPP stanza. The hook's handler is expected to accept four parameters: Acc, From, To and Packet The arguments and the return value types correspond to the following spec.

user_sent_keep_alive/2

Specs

mongoose_hooks.user_sent_keep_alive/2

vcard_set/4

Specs

mongoose_hooks.vcard_set/4

The vcard_set hook is called to inform that the vcard has been set in mod_vcard backend.

xmpp_bounce_message/2

Specs

mongoose_hooks.xmpp_bounce_message/2

xmpp_send_element/3

Specs

mongoose_hooks.xmpp_send_element/3

xmpp_stanza_dropped/5

Specs

mongoose_hooks.xmpp_stanza_dropped/5

The xmpp_stanza_dropped hook is called to inform that an xmpp stanza has been dropped.