mongoose_iq (MongooseIM v4.0.1)

Functions to work with iq record.

============================================================================== Copyright 2015 Erlang Solutions Ltd.

Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at

http://www.apache.org/licenses/LICENSE-2.0
Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License. ==============================================================================

Types

iq_acc_field()

Specs

t:mongoose_iq.iq_acc_field/0

Functions

command/1

Specs

mongoose_iq.command/1

empty_result_iq/1

info/1

Specs

mongoose_iq.info/1

iq_to_sub_el/1

sub_el_to_command/1

Specs

mongoose_iq.sub_el_to_command/1

try_to_handle_iq/4

Specs

mongoose_iq.try_to_handle_iq/4

Generic error handling code for IQ. Use this function, instead of try ... catch _:_ -> ok end.

update_acc_info/1

Specs

mongoose_iq.update_acc_info/1

update_and_get/2

Specs

mongoose_iq.update_and_get/2

xmlns/1

Specs

mongoose_iq.xmlns/1