mongoose_service behaviour (MongooseIM v4.0.1)

Types

options()

Specs

t:mongoose_service.options/0

service()

Specs

t:mongoose_service.service/0

Functions

assert_loaded/1

Specs

mongoose_service.assert_loaded/1

check_deps/1

check_deps/2

ensure_loaded/1

Specs

mongoose_service.ensure_loaded/1

ensure_loaded/2

Specs

mongoose_service.ensure_loaded/2

get_deps/1

Specs

mongoose_service.get_deps/1

get_service_opts/1

Specs

mongoose_service.get_service_opts/1

is_app_running/1

Specs

mongoose_service.is_app_running/1

is_loaded/1

Specs

mongoose_service.is_loaded/1

loaded_services_with_opts/0

Specs

mongoose_service.loaded_services_with_opts/0

purge_service/1

Specs

mongoose_service.purge_service/1

to be used as an emergency feature if serviced crashed while stopping and is not running but still lingers in the services tab

run_start_service/2

run_stop_service/1

start/0

Specs

mongoose_service.start/0

start_deps/1

Specs

mongoose_service.start_deps/1

start_service/2

Specs

mongoose_service.start_service/2

stop/0

Specs

mongoose_service.stop/0

stop_service/1

Specs

mongoose_service.stop_service/1

Callbacks

start(opts)

Specs

c:mongoose_service.start/1

stop()

Specs

c:mongoose_service.stop/0

optional: -callback deps() -> [service()].