mongoose_wpool behaviour (MongooseIM v4.0.1)

This is here because there are pool options which have to be given when calling the pool (selection strategy, timeout), while we want to set it once for the pool and not worry about them later, hence additional storage.

Types

host()

Specs

t:mongoose_wpool.host/0

name()

Specs

t:mongoose_wpool.name/0

tag()

Specs

t:mongoose_wpool.tag/0

type()

Specs

t:mongoose_wpool.type/0

Functions

call/2

call/3

call/4

call/5

call_callback/3

call_start_callback/2

cast/2

cast/3

cast/4

cast/5

default_opts/1

ensure_started/0

expand_pools/2

get_pool/3

get_pool_settings/3

get_pools/0

get_unique_types/1

get_worker/1

get_worker/2

get_worker/3

is_configured/1

Specs

mongoose_wpool.is_configured/1

make_callback_module_name/1

Specs

mongoose_wpool.make_callback_module_name/1

make_pool_name/1

make_pool_name/3

Specs

mongoose_wpool.make_pool_name/3

start/1

start/2

start/3

start/4

start/5

start/7

start_configured_pools/0

start_configured_pools/1

start_configured_pools/2

start_sup_pool/3

Specs

mongoose_wpool.start_sup_pool/3

this function starts the worker_pool's pool under a specific supervisor in MongooseIM application. It's needed for 2 reasons: 1. We want to have a full control of all the pools and its restarts 2. When a pool is started via wpool:start_pool it's supposed be called by a supervisor, if not, there is no way to stop the pool.

stats/3

stop/0

stop/1

stop/2

stop/3

Callbacks

default_opts()
(optional)

Specs

c:mongoose_wpool.default_opts/0

init()

Specs

c:mongoose_wpool.init/0

is_supported_strategy(strategy)
(optional)

Specs

c:mongoose_wpool.is_supported_strategy/1

start(host, tag, wPoolOpts, connOpts)

Specs

c:mongoose_wpool.start/4

stop(host, tag)

Specs

c:mongoose_wpool.stop/2