node_flat (MongooseIM v4.0.1)

The module node_flat is the default PubSub plugin.
It is used as a default for all unknown PubSub node type. It can serve as a developer basis and reference to build its own custom pubsub node types.

PubSub plugin nodes are using the gen_node behaviour.

Types

accessModel()

Specs

t:node_flat.accessModel/0

affiliation()

Specs

t:node_flat.affiliation/0

Functions

access_model_to_subscription/1

Specs

node_flat.access_model_to_subscription/1

authenticate_unsubscribe/4

authorize_get_item/5

authorize_subscription/7

Specs

node_flat.authorize_subscription/7

based_on/0

Callback implementation for c::gen_pubsub_node.based_on/0.

can_fetch_item/2

create_node/2

Callback implementation for c::gen_pubsub_node.create_node/2.

create_node_permission/6

Checks if the current user has the permission to create the requested node
In flat node, any unused node name is allowed. The access parameter is also checked. This parameter depends on the value of the access_createnode ACL value in ejabberd config file.

del_item/2

Delete an item from database.

del_items/2

delete_foreign_item/3

delete_item/4

Triggers item deletion.

Default plugin: The user performing the deletion must be the node owner or a publisher, or PublishModel being open.

delete_node/1

Callback implementation for c::gen_pubsub_node.delete_node/1.

features/0

Callback implementation for c::gen_pubsub_node.features/0.

get_affiliation/2

Callback implementation for c::gen_pubsub_node.get_affiliation/2.

get_entity_affiliations/2

Callback implementation for c::gen_pubsub_node.get_entity_affiliations/2.

get_entity_subscriptions/2

Callback implementation for c::gen_pubsub_node.get_entity_subscriptions/2.

get_entity_subscriptions_loop/3

get_item/2

Returns an item (one item list), given its reference.

get_item/7

Callback implementation for c::gen_pubsub_node.get_item/7.

get_item_name/3

Callback implementation for c::gen_pubsub_node.get_item_name/3.

get_items/3

Returns the list of stored items for a given node.
For the default PubSub module, items are stored in Mnesia database.

We can consider that the pubsub_item table have been created by the main mod_pubsub module.

PubSub plugins can store the items where they wants (for example in a relational database), or they can even decide not to persist any items.

get_items_if_authorised/3

Callback implementation for c::gen_pubsub_node.get_items_if_authorised/3.

get_node_affiliations/1

Callback implementation for c::gen_pubsub_node.get_node_affiliations/1.

get_node_subscriptions/1

Callback implementation for c::gen_pubsub_node.get_node_subscriptions/1.

get_pending_nodes/2

Returns a list of Owner's nodes on Host with pending subscriptions.

get_permition/5

get_subscriptions/2

Callback implementation for c::gen_pubsub_node.get_subscriptions/2.

init/3

Callback implementation for c::gen_pubsub_node.init/3.

is_subscribed/1

make_pubsub_item/8

make_subid/0

node_to_path/1

Return the path of the node. In flat it's just node id.

options/0

Callback implementation for c::gen_pubsub_node.options/0.

path_to_node/1

Callback implementation for c::gen_pubsub_node.path_to_node/1.

publish_item/9

Publishes the item passed as parameter.

The mechanism works as follow:

The main PubSub module prepares the item to publish and passes the result of the preparation as a mod_pubsub:pubsubItem() record.

This function gets the prepared record and several other parameters and can decide to:

reject the publication;

allow the publication as is, letting the main module perform the database persistance;

allow the publication, modifying the record. The main module will store the modified record;

allow it, but perform the needed persistance operations.

The selected behaviour depends on the return parameter:

{error, Reason}: an iq error result will be return. No publication is actually performed.

true: Publication operation is allowed, based on the unmodified record passed in parameter Item. If the Item parameter contains an error, no subscription will actually be performed.

{true, Item}: Publication operation is allowed, but the mod_pubsub:pubsubItem() record returned replaces the value passed in parameter Item. The persistance will be performed by the main module.

{true, done}: Publication operation is allowed, but the mod_pubsub:pubsubItem() will be considered as already stored and no further persistance operation will be performed. This case is used, when the plugin module is doing the persistance by itself or when it want to completly disable persistance.

In the default plugin module, the record is unchanged.

purge_node/2

Callback implementation for c::gen_pubsub_node.purge_node/2.

remove_extra_items/3

This function is used to remove extra items, most notably when the maximum number of items has been reached.

This function is used internally by the core PubSub module, as no permission check is performed.

In the default plugin module, the oldest items are removed, but other rules can be used.

If another PubSub plugin wants to delegate the item removal (and if the plugin is using the default pubsub storage), it can implements this function like this:
     remove_extra_items(Nidx, MaxItems, ItemIds) ->
            node_default:remove_extra_items(Nidx, MaxItems, ItemIds).

remove_user/2

Callback implementation for c::gen_pubsub_node.remove_user/2.

remove_user_by_affiliation_and_node/2

Specs

node_flat.remove_user_by_affiliation_and_node/2

set_affiliation/3

Callback implementation for c::gen_pubsub_node.set_affiliation/3.

set_item/1

Write an item into database.

set_subscriptions/4

should_delete_when_owner_removed/0

Callback implementation for c::gen_pubsub_node.should_delete_when_owner_removed/0.

subscribe_node/8

Accepts or rejects subcription requests on a PubSub node.

The mechanism works as follow:

The main PubSub module prepares the subscription and passes the result of the preparation as a record.

This function gets the prepared record and several other parameters and can decide to:

reject the subscription;

allow it as is, letting the main module perform the database persistance;

allow it, modifying the record. The main module will store the modified record;

allow it, but perform the needed persistance operations.

The selected behaviour depends on the return parameter:

{error, Reason}: an IQ error result will be returned. No subscription will actually be performed.

true: Subscribe operation is allowed, based on the unmodified record passed in parameter SubscribeResult. If this parameter contains an error, no subscription will be performed.

{true, PubsubState}: Subscribe operation is allowed, but the mod_pubsub:pubsubState() record returned replaces the value passed in parameter SubscribeResult.

{true, done}: Subscribe operation is allowed, but the mod_pubsub:pubsubState() will be considered as already stored and no further persistance operation will be performed. This case is used, when the plugin module is doing the persistance by itself or when it want to completly disable persistance.

In the default plugin module, the record is unchanged.

terminate/2

Callback implementation for c::gen_pubsub_node.terminate/2.

unsubscribe_node/4

Unsubscribe the Subscriber from the Node.