nodetree_tree (MongooseIM v4.0.1)

The module nodetree_tree is the default PubSub node tree plugin.
It is used as a default for all unknown PubSub node type. It can serve as a developer basis and reference to build its own custom pubsub node tree types.

PubSub node tree plugins are using the gen_nodetree behaviour.

The API isn't stabilized yet. The pubsub plugin development is still a work in progress. However, the system is already useable and useful as is. Please, send us comments, feedback and improvements.

Functions

check_parent_and_its_owner_list/3

create_node/6

Callback implementation for c::gen_pubsub_nodetree.create_node/6.

delete_node/2

Callback implementation for c::gen_pubsub_nodetree.delete_node/2.

get_node/1

Callback implementation for c::gen_pubsub_nodetree.get_node/1.

get_node/2

Callback implementation for c::gen_pubsub_nodetree.get_node/2.

get_nodes/2

Callback implementation for c::gen_pubsub_nodetree.get_nodes/2.

get_parentnodes_tree/3

Default node tree does not handle parents, return a list containing just this node.

get_subnodes/3

Callback implementation for c::gen_pubsub_nodetree.get_subnodes/3.

init/3

Callback implementation for c::gen_pubsub_nodetree.init/3.

set_node/1

Callback implementation for c::gen_pubsub_nodetree.set_node/1.

terminate/2

Callback implementation for c::gen_pubsub_nodetree.terminate/2.