service_admin_extra_roster (MongooseIM v4.0.1)

Types

delete_action()

Specs

t:service_admin_extra_roster.delete_action/0

jids_nick_subs_ask_grp()

Specs

t:service_admin_extra_roster.jids_nick_subs_ask_grp/0

list_action()

Specs

t:service_admin_extra_roster.list_action/0

push_action()

Specs

t:service_admin_extra_roster.push_action/0

simple_roster()

Specs

t:service_admin_extra_roster.simple_roster/0

subs()

Specs

t:service_admin_extra_roster.subs/0

Functions

add_rosteritem/7

Specs

service_admin_extra_roster.add_rosteritem/7

apply_action/2

build_broadcast/3

Specs

service_admin_extra_roster.build_broadcast/3

build_iq_roster_push/1

Specs

service_admin_extra_roster.build_iq_roster_push/1

build_list_users/3

Specs

service_admin_extra_roster.build_list_users/3

build_roster_item/2

Specs

service_admin_extra_roster.build_roster_item/2

commands/0

Specs

service_admin_extra_roster.commands/0

decide_rip/2

decide_rip_jid/2

delete_rosteritem/4

Specs

service_admin_extra_roster.delete_rosteritem/4

do_add_rosteritem/5

get_roster/2

Specs

service_admin_extra_roster.get_roster/2

is_regexp_match/2

make_roster/1

Specs

service_admin_extra_roster.make_roster/1

Note: if a contact is in several groups, the contact is returned several times, each one in a different group.

possible_subs_binary/0

process_rosteritems/5

Specs

service_admin_extra_roster.process_rosteritems/5

push_alltoall/2

Specs

service_admin_extra_roster.push_alltoall/2

push_roster/3

Specs

service_admin_extra_roster.push_roster/3

push_roster_all/1

Specs

service_admin_extra_roster.push_roster_all/1

push_roster_item/3

Specs

service_admin_extra_roster.push_roster_item/3

Push to the roster of account LU@LS the contact U@S. The specific action to perform is defined in Action.

rip/3

Specs

service_admin_extra_roster.rip/3

roster_list_to_binary/1

Specs

service_admin_extra_roster.roster_list_to_binary/1

subscribe/6

Specs

service_admin_extra_roster.subscribe/6

returns result of mnesia or rdbms transaction

subscribe_all/1

Specs

service_admin_extra_roster.subscribe_all/1

subscribe_all/2

subscribe_roster/2

Specs

service_admin_extra_roster.subscribe_roster/2

unsubscribe/2

Specs

service_admin_extra_roster.unsubscribe/2

returns result of mnesia or rdbms transaction

validate_regexps/1