service_admin_extra_stanza (MongooseIM v4.0.1)

Functions

build_packet/2

Specs

service_admin_extra_stanza.build_packet/2

commands/0

Specs

service_admin_extra_stanza.commands/0

route_packet/3

Specs

service_admin_extra_stanza.route_packet/3

send_message_chat/3

Specs

service_admin_extra_stanza.send_message_chat/3

Send a chat message to a Jabber account.

send_message_headline/4

Specs

service_admin_extra_stanza.send_message_headline/4

Send a headline message to a Jabber account.

send_packet_all_resources/3

Specs

service_admin_extra_stanza.send_packet_all_resources/3

Send a packet to a Jabber account. If a resource was specified in the JID, the packet is sent only to that specific resource. If no resource was specified in the JID, and the user is remote or local but offline, the packet is sent to the bare JID. If the user is local and is online in several resources, the packet is sent to all its resources.

send_packet_all_resources_2/3

Specs

service_admin_extra_stanza.send_packet_all_resources_2/3

send_stanza_c2s/4

Specs

service_admin_extra_stanza.send_stanza_c2s/4