shaper_srv (MongooseIM v4.0.1)

Shared shapers.

Types

key()

Specs

t:shaper_srv.key/0

state()

Specs

t:shaper_srv.state/0

Functions

child_spec/1

Specs

shaper_srv.child_spec/1

child_specs/0

Specs

shaper_srv.child_specs/0

code_change/3

Callback implementation for :gen_server.code_change/3.

create_shaper/1

Specs

shaper_srv.create_shaper/1

default_shaper/0

delete_old_shapers/1

Specs

shaper_srv.delete_old_shapers/1

find_or_create_shaper/2

Specs

shaper_srv.find_or_create_shaper/2

get_shaper_name/4

Specs

shaper_srv.get_shaper_name/4

handle_call/3

Callback implementation for :gen_server.handle_call/3.

handle_cast/2

Callback implementation for :gen_server.handle_cast/2.

handle_info/2

Callback implementation for :gen_server.handle_info/2.

init/1

Callback implementation for :gen_server.init/1.

init_dicts/1

Specs

shaper_srv.init_dicts/1

new_key/3

Specs

shaper_srv.new_key/3

reply_after/3

Specs

shaper_srv.reply_after/3

It is a small hack This function calls this in more efficient way: timer:apply_after(DelayMs, gen_server, reply, [From, Reply]).

request_shaper_name/1

Specs

shaper_srv.request_shaper_name/1

reset_all_shapers/1

Ask all shaper servers to forget current shapers and read settings again

reset_shapers/1

Ask server to forget its shapers

save_shaper/3

Specs

shaper_srv.save_shaper/3

select_worker/2

Specs

shaper_srv.select_worker/2

start_link/1

Specs

shaper_srv.start_link/1

subtract_seconds/2

Specs

shaper_srv.subtract_seconds/2

terminate/2

Callback implementation for :gen_server.terminate/2.

update_access_time/3

Specs

shaper_srv.update_access_time/3

wait/4

Specs

shaper_srv.wait/4

Shapes the caller from executing the action.

worker_count/1

worker_name/2

Specs

shaper_srv.worker_name/2

worker_names/1

Specs

shaper_srv.worker_names/1

worker_number/2

Specs

shaper_srv.worker_number/2

worker_prefix/0

Specs

shaper_srv.worker_prefix/0